### 2022-09-01 #4726 @editorialbot
[PRE REVIEW]: Hexatomic: An extensible, OS-independent platform for deep multi-layer linguistic annotation of corpora
**Submitting author:** @sdruskat (<a href="http://orcid.org/0000-0003-4925-7248">Stephan Druskat</a>)
**Repository:** https://github.com/hexatomic/hexatomic
**Branch with paper.md** (empty if default branch): JOSS-paper
**Version:** v1.0.1
