### 2022-09-02 #4726 @sdruskat
@Kevin-Mattheus-Moerman, thanks for handling this. I'm looking forward for this to be assigned an editor. I'm fully aware that JOSS operates on a voluntary basis, and that we're all very busy, so thanks very much in advance to everyone who'll handle this submission.

We'll look into potential reviewers and will make suggestions here.
