### 2022-09-07 #4726 @sdruskat
@thomaskrause and me have come up with the following list of potential reviewers, all of who have developed open source software in linguistics:

- @berndmoos
- @reckart
- @raphywink
