### 2022-09-07 ca942a8
Add empty media player project
