+++ bundles/org.corpus_tools.hexatomic.mediaplayer/src/main/java/org/corpus_tools/hexatomic/mediaplayer/MediaPlayer.java
@@ -1,0 +1,15 @@
+package org.corpus_tools.hexatomic.mediaplayer;
+
+import javax.annotation.PostConstruct;
+import org.corpus_tools.hexatomic.core.ProjectManager;
+import org.eclipse.e4.ui.model.application.ui.basic.MPart;
+import org.eclipse.swt.widgets.Composite;
+
+public class MediaPlayer {
+
+  @PostConstruct
+  public void postConstruct(Composite parent, MPart part, ProjectManager projectManager) {
+
+  }
+
+}
