### 2022-09-07 fc43b84
Load a template for the video/audio player HTML when the media player editor is loaded
