### 2022-09-07 93e65c0
Add selection manager to media player
