+++ bundles/org.corpus_tools.hexatomic.graph/src/main/java/org/corpus_tools/hexatomic/graph/internal/GraphLayoutParameterWidget.java
@@ -46,18 +46,18 @@ public class GraphLayoutParameterWidget extends Composite {
     }
 
     private void updateConfigFromInterface() {
-      config.setVerticalNodeMargin(((double) scalePercentMargin.getSelection()) / 10.0);
-      config.setHorizontalTokenMargin(((double) scaleTokenMargin.getSelection()) / 10.0);
-      config.setTokenRankOffset(scaleTokenRank.getSelection());
+      config.setVerticalNodeMargin(((double) scaleVerticalMargin.getSelection()) / 10.0);
+      config.setHorizontalTokenMargin(((double) scaleHorizontalTokenMargin.getSelection()) / 10.0);
+      config.setTokenRankOffset(scaleVerticalTokenMargin.getSelection());
     }
 
     @Override
     public void widgetSelected(SelectionEvent e) {
       updateConfigFromInterface();
       // Update all the labels
-      lblPercentMargin.setText("" + config.getVerticalNodeMargin());
-      lblTokenMargin.setText("" + config.getHorizontalTokenMargin());
-      lblTokenRank.setText("" + config.getTokenRankOffset());
+      lblVerticalMargin.setText("" + config.getVerticalNodeMargin());
+      lblHorizontalTokenMargin.setText("" + config.getHorizontalTokenMargin());
+      lblVerticalTokenMargin.setText("" + config.getTokenRankOffset());
 
       // Send the event that the config has changed
       if (eventBroker != null) {
@@ -67,16 +67,14 @@ public class GraphLayoutParameterWidget extends Composite {
   }
 
   public static final String PARAM_CHANGED_TOPIC = "GRAPH_EDITOR/GRAPH_LAYOUT_PARAMETER/CHANGED";
-  private Label lblPercentMargin;
-  private Scale scalePercentMargin;
+  private Label lblVerticalMargin;
+  private Scale scaleVerticalMargin;
 
   private GraphDisplayConfiguration config;
-  private Label lblTokenMarginCaption;
-  private Scale scaleTokenMargin;
-  private Label lblTokenMargin;
-  private Label lblTokenRankCaption;
-  private Scale scaleTokenRank;
-  private Label lblTokenRank;
+  private Scale scaleHorizontalTokenMargin;
+  private Label lblHorizontalTokenMargin;
+  private Scale scaleVerticalTokenMargin;
+  private Label lblVerticalTokenMargin;
 
   /**
    * Create a graph layout manipulation widget.
@@ -90,72 +88,73 @@ public class GraphLayoutParameterWidget extends Composite {
 
     config = new GraphDisplayConfiguration();
 
-    Label lblPercentMarginCaption = new Label(this, SWT.WRAP);
-    GridData lblPercentMarginCaptionGridData =
+    GridData captionGridData =
         new GridData(SWT.LEFT, SWT.CENTER, false, false, 1, 1);
-    lblPercentMarginCaptionGridData.widthHint = 80;
-    lblPercentMarginCaption.setLayoutData(lblPercentMarginCaptionGridData);
-    lblPercentMarginCaption.setToolTipText("Vertical margin between nodes.\n"
+    captionGridData.widthHint = 80;
+    
+    GridData scaleGridData = new GridData(SWT.FILL, SWT.CENTER, true, false, 1, 1);
+    
+    
+    Label captionVerticalMargin = new Label(this, SWT.WRAP);
+    captionVerticalMargin.setLayoutData(captionGridData);
+    captionVerticalMargin.setToolTipText("Vertical margin between nodes.\n"
         + "This is measure in \"times of the node height\".  "
         + "So for \"0\" there is no margin, for \"1\" the margin has the same height as the node, "
[…]
