### 2022-09-08 1459be3
Apply SWT workaround for scales in the expand item everytime the control is resized
