### 2022-09-08 81b5bed
Don't reuse GridData object
