### 2022-09-08 41cd48a
Use icons instead of verbose text
