### 2022-09-08 9596246
Restructure graph editor documentation so that filtering the view is its own section
