+++ docs/user/src/SUMMARY.md
@@ -9,7 +9,8 @@
 	- [Editing the corpus structure](usage/corpus-structure.md)
 	- [Opening an editor](usage/open-editor.md)
 	- [Graph Editor](usage/graph-editor/README.md)
-    	- [Editing the graph](usage/graph-editor/console.md)
+    	- [Filter visible annotations](usage/graph-editor/filter.md)
+    	- [Editing the graph](usage/graph-editor/console.md) 
   	- [Grid Editor](usage/grid-editor/README.md)
     	- [Editing the grid](usage/grid-editor/editing.md)
 		    - [Editing annotation columns](usage/grid-editor/editing-columns.md)
