### 2022-09-08 e8f887e
Document the graph layout parameter
