+++ docs/user/src/SUMMARY.md
@@ -10,7 +10,8 @@
 	- [Opening an editor](usage/open-editor.md)
 	- [Graph Editor](usage/graph-editor/README.md)
     	- [Filter visible annotations](usage/graph-editor/filter.md)
-    	- [Editing the graph](usage/graph-editor/console.md) 
+    	- [Editing the graph](usage/graph-editor/console.md)
+    	- [Change layout parameters](usage/graph-editor/layout-params.md) 
   	- [Grid Editor](usage/grid-editor/README.md)
     	- [Editing the grid](usage/grid-editor/editing.md)
 		    - [Editing annotation columns](usage/grid-editor/editing-columns.md)
