+++ docs/user/src/usage/graph-editor/layout-params.md
@@ -1,0 +1,45 @@
+# Change layout parameters
+
+Depending on the shown annotations and the specific graph, the default visualization of the graph might get cluttered.
+The graph editor will try to arrange the nodes using a default layout algorithm, which parameters you can adjust by expanding the "Graph Layout" section.
+
+![Graph editor view with expanded layout parameters](./layout-params.png)
+
+## Horizontal margin between token
+
+![Horizontal margin parameter slider](./layout-param-horizontal-margin.png)
+
+
+There is some horizontal space between the token at the bottom of the graph.
+If the space is increased, this indirectly affects the space between other nodes, too.
+
+This is measured in "times of the average token width".
+So for \"0\" there is no margin, for "1" the margin has  the same width as the average token node, and for "2" the margin is twice as high as the average token node width.
+
+![Effect of the settings 0, 1 and 2 for the horizontal margin parameter](horizontal-margin-examples.png)
+
+## Vertical margin between nodes
+
+![Vertical margin parameter slider](./layout-param-vertical-margin.png)
+
+Configures the vertical margin between *all* nodes.
+
+This is measure in "times of the node height".
+So for "0" there is no margin, for "1" the margin has the same height as the node.
+
+![Effect of the settings 0, 1 and 2 for the vertical margin parameter](vertical-margin-examples.png)
+
+
+## Vertical margin between token and non-token
+
+![Vertical margin between token and non-token parameter slider](./layout-param-token-margin.png)
+
+
+Token are grouped in the lowest rank (at the bottom)
+To allow space for pointing relations, you can add a margin between the token row and the lowest annotation nodes.
+This margin is in addition to the [vertical margin between nodes](#vertical-margin-between-nodes).
+
+A margin of "1" means there is one level left empty, an offset of "0"
+means there is no additional space except for the regular vertical margin.
+
+![Effect of the settings 0 and 3 for the vertical token margin parameter](./vertical-token-margin-examples.png)
