### 2022-09-08 9e9060e
Format automatically created resource manager class
