+++ bundles/org.corpus_tools.hexatomic.graph/src/main/java/org/eclipse/wb/swt/SwtResourceManager.java
@@ -1,0 +1,458 @@
+/*******************************************************************************
+ * Copyright (c) 2011 Google, Inc. 
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0 which accompanies this
+ * distribution, and is available at http://www.eclipse.org/legal/epl-v10.html
+ * 
+ * <p>Contributors: 
+ * Google, Inc. - initial API and implementation
+ * </p>
+ *******************************************************************************/
+
+package org.eclipse.wb.swt;
+
+import java.io.FileInputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.util.HashMap;
+import java.util.Map;
+import org.eclipse.swt.SWT;
+import org.eclipse.swt.graphics.Color;
+import org.eclipse.swt.graphics.Cursor;
+import org.eclipse.swt.graphics.Font;
+import org.eclipse.swt.graphics.FontData;
+import org.eclipse.swt.graphics.GC;
+import org.eclipse.swt.graphics.Image;
+import org.eclipse.swt.graphics.ImageData;
+import org.eclipse.swt.graphics.RGB;
+import org.eclipse.swt.graphics.Rectangle;
+import org.eclipse.swt.widgets.Display;
+
+/**
+ * Utility class for managing OS resources associated with SWT controls such as colors, fonts,
+ * images, etc.
+ * 
+ * <p>
+ * !!! IMPORTANT !!! Application code must explicitly invoke the <code>dispose()</code> method to
+ * release the operating system resources managed by cached objects when those objects and OS
+ * resources are no longer needed (e.g. on application shutdown)
+ * </p>
+ * 
+ * <p>
+ * This class may be freely distributed as part of any application or plugin.
+ * </p>
+ * 
+ * @author scheglov_ke
+ * @author Dan Rubel
+ */
+public class SwtResourceManager {
+  ////////////////////////////////////////////////////////////////////////////
+  //
+  // Color
+  //
+  ////////////////////////////////////////////////////////////////////////////
+  private static Map<RGB, Color> m_colorMap = new HashMap<RGB, Color>();
+
+  /**
+   * Returns the system {@link Color} matching the specific ID.
+   * 
+   * @param systemColorID the ID value for the color
+   * @return the system {@link Color} matching the specific ID
+   */
+  public static Color getColor(int systemColorID) {
+    Display display = Display.getCurrent();
+    return display.getSystemColor(systemColorID);
+  }
+
+  /**
+   * Returns a {@link Color} given its red, green and blue component values.
+   * 
+   * @param r the red component of the color
+   * @param g the green component of the color
+   * @param b the blue component of the color
+   * @return the {@link Color} matching the given red, green and blue component values
+   */
+  public static Color getColor(int r, int g, int b) {
+    return getColor(new RGB(r, g, b));
+  }
+
+  /**
+   * Returns a {@link Color} given its RGB value.
+   * 
+   * @param rgb the {@link RGB} value of the color
+   * @return the {@link Color} matching the RGB value
+   */
+  public static Color getColor(RGB rgb) {
+    Color color = m_colorMap.get(rgb);
+    if (color == null) {
+      Display display = Display.getCurrent();
+      color = new Color(display, rgb);
+      m_colorMap.put(rgb, color);
+    }
+    return color;
+  }
+
+  /**
+   * Dispose of all the cached {@link Color}'s.
+   */
+  public static void disposeColors() {
+    for (Color color : m_colorMap.values()) {
+      color.dispose();
+    }
+    m_colorMap.clear();
+  }
+
[…]
