### 2022-09-08 48d8a1d
Updated CITATION.cff (only new snapshot version and date)
