+++ CITATION.cff
@@ -34,8 +34,8 @@ identifiers:
   type: doi
   value: 10.5281/zenodo.6900689
 doi: 10.5281/zenodo.6900689
-version: 1.0.1
-date-released: 2022-08-30
+version: 1.1.0-SNAPSHOT
+date-released: 2022-09-08
 references:
 - type: software
   title: SWT Resource Manager
