### 2022-09-14 dbe5c05
Fix some of the code smell in the resource management code
