+++ bundles/org.corpus_tools.hexatomic.core/src/main/java/org/eclipse/wb/swt/ResourceManager.java
@@ -1,12 +1,11 @@
 /*******************************************************************************
- * Copyright (c) 2011 Google, Inc. and others 
- * All rights reserved. This program and the accompanying
+ * Copyright (c) 2011 Google, Inc. and others All rights reserved. This program and the accompanying
  * materials are made available under the terms of the Eclipse Public License v1.0 which accompanies
- * this distribution, and is available at http://www.eclipse.org/legal/epl-v10.html 
+ * this distribution, and is available at http://www.eclipse.org/legal/epl-v10.html
  *
- * <p>Contributors:
- * Google, Inc. - initial API and implementation 
- * Wim Jongman - 1.8 and higher compliance
+ * <p>
+ * Contributors: Google, Inc. - initial API and implementation Wim Jongman - 1.8 and higher
+ * compliance
  * </p>
  *******************************************************************************/
 
@@ -14,8 +13,6 @@ package org.eclipse.wb.swt;
 
 import java.io.File;
 import java.io.InputStream;
-import java.lang.reflect.Constructor;
-import java.lang.reflect.Method;
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.HashMap;
@@ -65,8 +62,7 @@ public class ResourceManager extends SwtResourceManager {
   /**
    * The map where we store our images.
    */
-  private static Map<ImageDescriptor, Image> m_descriptorImageMap =
-      new HashMap<ImageDescriptor, Image>();
+  private static Map<ImageDescriptor, Image> descriptorImageMap = new HashMap<>();
 
   /**
    * Returns an {@link ImageDescriptor} stored in the file at the specified path relative to the
@@ -104,19 +100,14 @@ public class ResourceManager extends SwtResourceManager {
     if (descriptor == null) {
       return null;
     }
-    Image image = m_descriptorImageMap.get(descriptor);
-    if (image == null) {
-      image = descriptor.createImage();
-      m_descriptorImageMap.put(descriptor, image);
-    }
-    return image;
+    return descriptorImageMap.computeIfAbsent(descriptor, d -> d.createImage());
   }
 
   /**
    * Maps images to decorated images.
    */
   @SuppressWarnings("unchecked")
-  private static Map<Image, Map<Image, Image>>[] m_decoratedImageMap = new Map[LAST_CORNER_KEY];
+  private static Map<Image, Map<Image, Image>>[] decoratedImageMap = new Map[LAST_CORNER_KEY];
 
   /**
    * Returns an {@link Image} composed of a base image decorated by another image.
@@ -142,17 +133,13 @@ public class ResourceManager extends SwtResourceManager {
     if (corner <= 0 || corner >= LAST_CORNER_KEY) {
       throw new IllegalArgumentException("Wrong decorate corner");
     }
-    Map<Image, Map<Image, Image>> cornerDecoratedImageMap = m_decoratedImageMap[corner];
+    Map<Image, Map<Image, Image>> cornerDecoratedImageMap = decoratedImageMap[corner];
     if (cornerDecoratedImageMap == null) {
-      cornerDecoratedImageMap = new HashMap<Image, Map<Image, Image>>();
-      m_decoratedImageMap[corner] = cornerDecoratedImageMap;
+      cornerDecoratedImageMap = new HashMap<>();
+      decoratedImageMap[corner] = cornerDecoratedImageMap;
     }
-    Map<Image, Image> decoratedMap = cornerDecoratedImageMap.get(baseImage);
-    if (decoratedMap == null) {
-      decoratedMap = new HashMap<Image, Image>();
-      cornerDecoratedImageMap.put(baseImage, decoratedMap);
-    }
-    //
+    Map<Image, Image> decoratedMap =
+        cornerDecoratedImageMap.computeIfAbsent(baseImage, i -> new HashMap<>());
[…]
