+++ bundles/org.corpus_tools.hexatomic.core/src/main/java/org/eclipse/wb/swt/SwtResourceManager.java
@@ -1,11 +1,10 @@
 /*******************************************************************************
- * Copyright (c) 2011 Google, Inc. 
- * All rights reserved. This program and the accompanying materials
+ * Copyright (c) 2011 Google, Inc. All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0 which accompanies this
  * distribution, and is available at http://www.eclipse.org/legal/epl-v10.html
  * 
- * <p>Contributors: 
- * Google, Inc. - initial API and implementation
+ * <p>
+ * Contributors: Google, Inc. - initial API and implementation
  * </p>
  *******************************************************************************/
 
@@ -51,7 +50,7 @@ public class SwtResourceManager {
   // Color
   //
   ////////////////////////////////////////////////////////////////////////////
-  private static Map<RGB, Color> m_colorMap = new HashMap<RGB, Color>();
+  private static Map<RGB, Color> colorMap = new HashMap<>();
 
   /**
    * Returns the system {@link Color} matching the specific ID.
@@ -83,11 +82,11 @@ public class SwtResourceManager {
    * @return the {@link Color} matching the RGB value
    */
   public static Color getColor(RGB rgb) {
-    Color color = m_colorMap.get(rgb);
+    Color color = colorMap.get(rgb);
     if (color == null) {
       Display display = Display.getCurrent();
       color = new Color(display, rgb);
-      m_colorMap.put(rgb, color);
+      colorMap.put(rgb, color);
     }
     return color;
   }
@@ -96,10 +95,10 @@ public class SwtResourceManager {
    * Dispose of all the cached {@link Color}'s.
    */
   public static void disposeColors() {
-    for (Color color : m_colorMap.values()) {
+    for (Color color : colorMap.values()) {
       color.dispose();
     }
-    m_colorMap.clear();
+    colorMap.clear();
   }
 
   ////////////////////////////////////////////////////////////////////////////
@@ -110,7 +109,7 @@ public class SwtResourceManager {
   /**
    * Maps image paths to images.
    */
-  private static Map<String, Image> m_imageMap = new HashMap<String, Image>();
+  private static Map<String, Image> imageMap = new HashMap<>();
 
   /**
    * Returns an {@link Image} encoded by the specified {@link InputStream}.
@@ -138,14 +137,14 @@ public class SwtResourceManager {
    * @return the {@link Image} stored in the file at the specified path
    */
   public static Image getImage(String path) {
-    Image image = m_imageMap.get(path);
+    Image image = imageMap.get(path);
     if (image == null) {
       try {
         image = getImage(new FileInputStream(path));
-        m_imageMap.put(path, image);
+        imageMap.put(path, image);
       } catch (Exception e) {
         image = getMissingImage();
-        m_imageMap.put(path, image);
+        imageMap.put(path, image);
       }
     }
     return image;
@@ -161,14 +160,14 @@ public class SwtResourceManager {
    */
   public static Image getImage(Class<?> clazz, String path) {
     String key = clazz.getName() + '|' + path;
-    Image image = m_imageMap.get(key);
+    Image image = imageMap.get(key);
     if (image == null) {
       try {
         image = getImage(clazz.getResourceAsStream(path));
-        m_imageMap.put(key, image);
+        imageMap.put(key, image);
       } catch (Exception e) {
         image = getMissingImage();
-        m_imageMap.put(key, image);
[…]
