### 2022-09-14 a087cd3
Fix checkstyle violation
