### 2022-09-14 6eaec2f
Fix more code smells and remove some duplicated functionality from the base class
