+++ bundles/org.corpus_tools.hexatomic.core/src/main/java/org/eclipse/wb/swt/SwtResourceManager.java
@@ -24,7 +24,6 @@ import org.eclipse.swt.graphics.GC;
 import org.eclipse.swt.graphics.Image;
 import org.eclipse.swt.graphics.ImageData;
 import org.eclipse.swt.graphics.RGB;
-import org.eclipse.swt.graphics.Rectangle;
 import org.eclipse.swt.widgets.Display;
 
 /**
@@ -44,7 +43,7 @@ import org.eclipse.swt.widgets.Display;
  * @author scheglov_ke
  * @author Dan Rubel
  */
-public class SwtResourceManager {
+public abstract class SwtResourceManager {
   ////////////////////////////////////////////////////////////////////////////
   //
   // Color
@@ -82,12 +81,10 @@ public class SwtResourceManager {
    * @return the {@link Color} matching the RGB value
    */
   public static Color getColor(RGB rgb) {
-    Color color = colorMap.get(rgb);
-    if (color == null) {
+    Color color = colorMap.computeIfAbsent(rgb, key -> {
       Display display = Display.getCurrent();
-      color = new Color(display, rgb);
-      colorMap.put(rgb, color);
-    }
+      return new Color(display, key);
+    });
     return color;
   }
 
@@ -211,96 +208,20 @@ public class SwtResourceManager {
    * Internal value.
    */
   protected static final int LAST_CORNER_KEY = 5;
-  /**
-   * Maps images to decorated images.
-   */
-  @SuppressWarnings("unchecked")
-  private static Map<Image, Map<Image, Image>>[] decoratedImageMap = new Map[LAST_CORNER_KEY];
 
-  /**
-   * Returns an {@link Image} composed of a base image decorated by another image.
-   * 
-   * @param baseImage the base {@link Image} that should be decorated
-   * @param decorator the {@link Image} to decorate the base image
-   * @return {@link Image} The resulting decorated image
-   */
-  public static Image decorateImage(Image baseImage, Image decorator) {
-    return decorateImage(baseImage, decorator, BOTTOM_RIGHT);
-  }
 
-  /**
-   * Returns an {@link Image} composed of a base image decorated by another image.
-   * 
-   * @param baseImage the base {@link Image} that should be decorated
-   * @param decorator the {@link Image} to decorate the base image
-   * @param corner the corner to place decorator image
-   * @return the resulting decorated {@link Image}
-   */
-  public static Image decorateImage(final Image baseImage, final Image decorator,
-      final int corner) {
-    if (corner <= 0 || corner >= LAST_CORNER_KEY) {
-      throw new IllegalArgumentException("Wrong decorate corner");
+  private static void disposeLoadedImages() {
+    for (Image image : imageMap.values()) {
+      image.dispose();
     }
-    Map<Image, Map<Image, Image>> cornerDecoratedImageMap = decoratedImageMap[corner];
-    if (cornerDecoratedImageMap == null) {
-      cornerDecoratedImageMap = new HashMap<>();
-      decoratedImageMap[corner] = cornerDecoratedImageMap;
-    }
-    Map<Image, Image> decoratedMap = cornerDecoratedImageMap.get(baseImage);
-    if (decoratedMap == null) {
-      decoratedMap = new HashMap<>();
-      cornerDecoratedImageMap.put(baseImage, decoratedMap);
-    }
-    //
-    Image result = decoratedMap.get(decorator);
-    if (result == null) {
-      Rectangle bib = baseImage.getBounds();
-      Rectangle dib = decorator.getBounds();
-      //
-      result = new Image(Display.getCurrent(), bib.width, bib.height);
-      //
-      GC gc = new GC(result);
-      gc.drawImage(baseImage, 0, 0);
-      if (corner == TOP_LEFT) {
-        gc.drawImage(decorator, 0, 0);
-      } else if (corner == TOP_RIGHT) {
[…]

[…]
