### 2022-09-14 3f9bc35
Use ResourceManager to create console font in graph editor
