+++ bundles/org.corpus_tools.hexatomic.graph/src/main/java/org/corpus_tools/hexatomic/console/ConsoleView.java
@@ -42,7 +42,7 @@ import org.eclipse.swt.events.VerifyEvent;
 import org.eclipse.swt.events.VerifyListener;
 import org.eclipse.swt.graphics.Font;
 import org.eclipse.swt.graphics.FontData;
-import org.eclipse.swt.widgets.Display;
+import org.eclipse.wb.swt.ResourceManager;
 
 public class ConsoleView implements Runnable, IDocumentListener, VerifyListener {
 
@@ -66,8 +66,7 @@ public class ConsoleView implements Runnable, IDocumentListener, VerifyListener 
    * @param projectManager An project manager object.
    * @param graph The Salt graph to edit.
    */
-  public ConsoleView(SourceViewer view, UISynchronize sync,
-      ProjectManager projectManager,
+  public ConsoleView(SourceViewer view, UISynchronize sync, ProjectManager projectManager,
       SDocumentGraph graph) {
     this.document = view.getDocument();
     this.sync = sync;
@@ -208,21 +207,27 @@ public class ConsoleView implements Runnable, IDocumentListener, VerifyListener 
     private void verifyCtrlPlusKey(VerifyEvent e) {
       e.doit = false;
       FontData[] fd = styledText.getFont().getFontData();
-
-      fd[0].setHeight(fd[0].getHeight() + 1);
-      styledText.setFont(new Font(Display.getCurrent(), fd[0]));
+      if (fd.length > 0) {
+        Font newFont =
+            ResourceManager.getFont(fd[0].getName(), fd[0].getHeight() + 1, fd[0].getStyle());
+        styledText.setFont(newFont);
+      }
     }
 
     private void verifyCtrlMinusKey(VerifyEvent e) {
       e.doit = false;
       FontData[] fd = styledText.getFont().getFontData();
 
-      if (fd[0].getHeight() > 6) {
-        fd[0].setHeight(fd[0].getHeight() - 1);
+      if (fd.length > 0) {
+        // Minimum font size is 6
+        if (fd[0].getHeight() > 6) {
+          Font newFont =
+              ResourceManager.getFont(fd[0].getName(), fd[0].getHeight() - 1, fd[0].getStyle());
+          styledText.setFont(newFont);
+        }
       }
-      styledText.setFont(new Font(Display.getCurrent(), fd[0]));
     }
-    
+
     private void setCommand(String cmd) {
       if (cmd == null) {
         return;
