### 2022-09-14 62edd34
Use Resource manager to create the colors of the ColorPalette
