### 2022-09-14 246c3c3
Return temporary value directly
