### 2022-09-14 f2d6989
More code smell fixes
