### 2022-09-14 8dd1aa8
Add test for horizontal token margin
