+++ tests/org.corpus_tools.hexatomic.it.tests/src/main/java/org/corpus_tools/hexatomic/it/tests/TestGraphEditor.java
@@ -3,6 +3,7 @@ package org.corpus_tools.hexatomic.it.tests;
 import static org.eclipse.swtbot.swt.finder.matchers.WidgetMatcherFactory.widgetOfType;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
@@ -29,6 +30,8 @@ import org.corpus_tools.salt.common.SDocument;
 import org.corpus_tools.salt.common.SDocumentGraph;
 import org.corpus_tools.salt.common.SPointingRelation;
 import org.corpus_tools.salt.common.STextualDS;
+import org.corpus_tools.salt.common.SToken;
+import org.corpus_tools.salt.core.SNode;
 import org.eclipse.core.commands.ParameterizedCommand;
 import org.eclipse.draw2d.ScalableFigure;
 import org.eclipse.draw2d.Viewport;
@@ -56,6 +59,8 @@ import org.eclipse.swtbot.swt.finder.utils.WidgetTextDescription;
 import org.eclipse.swtbot.swt.finder.waits.DefaultCondition;
 import org.eclipse.swtbot.swt.finder.waits.ICondition;
 import org.eclipse.swtbot.swt.finder.widgets.SWTBotCheckBox;
+import org.eclipse.swtbot.swt.finder.widgets.SWTBotExpandBar;
+import org.eclipse.swtbot.swt.finder.widgets.SWTBotScale;
 import org.eclipse.swtbot.swt.finder.widgets.SWTBotStyledText;
 import org.eclipse.swtbot.swt.finder.widgets.SWTBotTable;
 import org.eclipse.swtbot.swt.finder.widgets.SWTBotText;
@@ -76,6 +81,7 @@ class TestGraphEditor {
   private static final String ANNOTATION_NAME = "Node Annotations";
   private static final String SPANS = "Spans";
   private static final String FILTER_VIEW = "Filter View";
+  private static final String GRAPH_LAYOUT = "Graph Layout";
   private static final String ANNOTATION_TYPES = "Annotation Types";
   private static final String CORPUS_EDITOR_PART_ID =
       "org.corpus_tools.hexatomic.corpusedit.part.corpusstructure";
@@ -96,6 +102,36 @@ class TestGraphEditor {
   private UiStatusReport uiStatus;
 
   private final Keyboard keyboard = KeyboardFactory.getAWTKeyboard();
+
+  private final class HorizontalTokenDistanceCondition extends DefaultCondition {
+    private final List<SToken> token;
+    private final double expected;
+    private final Graph g;
+
+    private HorizontalTokenDistanceCondition(List<SToken> token, double expected, Graph g) {
+      this.token = token;
+      this.expected = expected;
+      this.g = g;
+    }
+
+    private double getDistance() {
+      GraphNode t1 = getGraphNodeForSalt(bot, g, token.get(0));
+      GraphNode t2 = getGraphNodeForSalt(bot, g, token.get(1));
+      return t2.getNodeFigure().getBounds().getLeft()
+          .getDistance(t1.getNodeFigure().getBounds().getRight());
+    }
+
+    @Override
+    public boolean test() throws Exception {
+      return getDistance() == expected;
+    }
+
+    @Override
+    public String getFailureMessage() {
+      return "Horizontal distance between first token should have been " + expected + " but was "
+          + getDistance() + ".";
+    }
+  }
 
   private final class ConsoleFontSizeCondition implements ICondition {
[…]
