### 2022-09-14 3c22f33
Add test for vertical (non-token) margin
