### 2022-09-14 0a110f9
Use fixed values
