### 2022-09-14 663935b
Add more timeout when waiting for graph layout updates
