+++ tests/org.corpus_tools.hexatomic.it.tests/src/main/java/org/corpus_tools/hexatomic/it/tests/TestGraphEditor.java
@@ -75,6 +75,7 @@ import org.junit.jupiter.api.TestMethodOrder;
 @TestMethodOrder(OrderAnnotation.class)
 class TestGraphEditor {
 
+  private static final long GRAPH_LAYOUT_TIMEOUT = 5000;
   private static final String INF_STRUCT = "Inf-Struct";
   private static final String CONST = "const";
   private static final String SEARCH = "Search";
@@ -131,9 +132,8 @@ class TestGraphEditor {
     @Override
     public String getFailureMessage() {
       return "Horizontal distance between node " + leftNode.getName() + " and "
-          + rightNode.getName()
-          + " should have been " + expected + " but was "
-          + getDistance() + ".";
+          + rightNode.getName() + " should have been " + expected + " but was " + getDistance()
+          + ".";
     }
   }
 
@@ -447,7 +447,7 @@ class TestGraphEditor {
         public String getFailureMessage() {
           return "Second text segment was not checked";
         }
-      }, 5000);
+      }, GRAPH_LAYOUT_TIMEOUT);
     }
   }
 
@@ -721,27 +721,29 @@ class TestGraphEditor {
       // is updated
       horizontalScale.setValue(0);
       assertEquals("0.0", botLayout.label(1).getText());
-      bot.waitUntil(new HorizontalNodeDistanceCondition(token.get(0), token.get(1), 0.0, g));
+      bot.waitUntil(new HorizontalNodeDistanceCondition(token.get(0), token.get(1), 0.0, g),
+          GRAPH_LAYOUT_TIMEOUT);
 
       horizontalScale.setValue(5);
       assertEquals("0.5", botLayout.label(1).getText());
-      bot.waitUntil(new HorizontalNodeDistanceCondition(token.get(0), token.get(1),
-          65.0, g));
+      bot.waitUntil(new HorizontalNodeDistanceCondition(token.get(0), token.get(1), 65.0, g),
+          GRAPH_LAYOUT_TIMEOUT);
 
       horizontalScale.setValue(10);
       assertEquals("1.0", botLayout.label(1).getText());
-      bot.waitUntil(new HorizontalNodeDistanceCondition(token.get(0), token.get(1), 130.0, g));
+      bot.waitUntil(new HorizontalNodeDistanceCondition(token.get(0), token.get(1), 130.0, g),
+          GRAPH_LAYOUT_TIMEOUT);
 
       horizontalScale.setValue(14);
       assertEquals("1.4", botLayout.label(1).getText());
-      bot.waitUntil(new HorizontalNodeDistanceCondition(token.get(0), token.get(1),
-          182.0, g));
+      bot.waitUntil(new HorizontalNodeDistanceCondition(token.get(0), token.get(1), 182.0, g),
+          GRAPH_LAYOUT_TIMEOUT);
       horizontalScale.setValue(15);
 
       horizontalScale.setValue(20);
       assertEquals("2.0", botLayout.label(1).getText());
-      bot.waitUntil(new HorizontalNodeDistanceCondition(token.get(0), token.get(1),
-          260.0, g));
+      bot.waitUntil(new HorizontalNodeDistanceCondition(token.get(0), token.get(1), 260.0, g),
+          GRAPH_LAYOUT_TIMEOUT);
 
       // Change the vertical margin and compare the distance between the root node and a node below
       // it
@@ -751,21 +753,53 @@ class TestGraphEditor {
 
       verticalScale.setValue(0);
       assertEquals("0.0", botLayout.label(3).getText());
[…]
