### 2022-09-14 b3548a2
Allow for 20% tolerance, but at least 1 point
