### 2022-09-14 8038aca
Close graph editor after changing the different layout parameters
