### 2022-09-14 6e61e1e
Close editor window after each graph editor test
