### 2022-09-14 6153cd6
Fix another checkstyle violation
