### 2022-09-14 362e1ae
Use constants for repeatly used strings in test class
