### 2022-09-15 5825aeb
Removed unused (and untested) code and add a application shutdown hook to dispose the resources
