+++ bundles/org.corpus_tools.hexatomic.core/src/main/java/org/eclipse/wb/swt/ResourceManager.java
@@ -11,32 +11,20 @@
 
 package org.eclipse.wb.swt;
 
-import static org.eclipse.wb.swt.SwtResourceManager.BOTTOM_LEFT;
-import static org.eclipse.wb.swt.SwtResourceManager.BOTTOM_RIGHT;
-import static org.eclipse.wb.swt.SwtResourceManager.LAST_CORNER_KEY;
-import static org.eclipse.wb.swt.SwtResourceManager.TOP_LEFT;
-import static org.eclipse.wb.swt.SwtResourceManager.TOP_RIGHT;
-
-import java.io.File;
+import java.io.IOException;
 import java.io.InputStream;
-import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.Map;
 import org.eclipse.core.runtime.Platform;
-import org.eclipse.jface.resource.CompositeImageDescriptor;
-import org.eclipse.jface.resource.ImageDescriptor;
 import org.eclipse.swt.graphics.Image;
 import org.eclipse.swt.graphics.ImageData;
-import org.eclipse.swt.graphics.ImageDataProvider;
-import org.eclipse.swt.graphics.Point;
-import org.eclipse.swt.graphics.Rectangle;
+import org.eclipse.swt.widgets.Display;
 import org.osgi.framework.Bundle;
 
 /**
- * Utility class for managing OS resources associated with SWT/JFace controls such as colors, fonts,
- * images, etc.
+ * Utility class for managing OS resources associated with SWT/JFace controls such as images.
  * 
  * <p>
  * This class is created automatically when you fiddle around with images and colors in WB. You
@@ -69,128 +57,7 @@ public class ResourceManager {
     throw new IllegalStateException("Utility class that should not be instantiated");
   }
 
-  /**
-   * The map where we store our images.
-   */
-  private static Map<ImageDescriptor, Image> descriptorImageMap = new HashMap<>();
 
-  /**
-   * Returns an {@link ImageDescriptor} stored in the file at the specified path relative to the
-   * specified class.
-   * 
-   * @param clazz the {@link Class} relative to which to find the image descriptor.
-   * @param path the path to the image file.
-   * @return the {@link ImageDescriptor} stored in the file at the specified path.
-   */
-  public static ImageDescriptor getImageDescriptor(Class<?> clazz, String path) {
-    return ImageDescriptor.createFromFile(clazz, path);
-  }
-
-  /**
-   * Returns an {@link ImageDescriptor} stored in the file at the specified path.
-   * 
-   * @param path the path to the image file.
-   * @return the {@link ImageDescriptor} stored in the file at the specified path.
-   */
-  public static ImageDescriptor getImageDescriptor(String path) {
-    try {
-      return ImageDescriptor.createFromURL(new File(path).toURI().toURL());
-    } catch (MalformedURLException e) {
-      return null;
-    }
-  }
-
-  /**
-   * Returns an {@link Image} based on the specified {@link ImageDescriptor}.
-   * 
-   * @param descriptor the {@link ImageDescriptor} for the {@link Image}.
-   * @return the {@link Image} based on the specified {@link ImageDescriptor}.
-   */
-  public static Image getImage(ImageDescriptor descriptor) {
-    if (descriptor == null) {
-      return null;
-    }
-    return descriptorImageMap.computeIfAbsent(descriptor, ImageDescriptor::createImage);
-  }
-
-  /**
-   * Maps images to decorated images.
-   */
-  @SuppressWarnings("unchecked")
-  private static Map<Image, Map<Image, Image>>[] decoratedImageMap = new Map[LAST_CORNER_KEY];
-
-  /**
-   * Returns an {@link Image} composed of a base image decorated by another image.
-   * 
[…]
