+++ bundles/org.corpus_tools.hexatomic.core/src/main/java/org/eclipse/wb/swt/SwtResourceManager.java
@@ -10,25 +10,17 @@
 
 package org.eclipse.wb.swt;
 
-import java.io.FileInputStream;
-import java.io.IOException;
-import java.io.InputStream;
 import java.util.HashMap;
 import java.util.Map;
 import org.eclipse.swt.SWT;
 import org.eclipse.swt.graphics.Color;
-import org.eclipse.swt.graphics.Cursor;
 import org.eclipse.swt.graphics.Font;
 import org.eclipse.swt.graphics.FontData;
-import org.eclipse.swt.graphics.GC;
-import org.eclipse.swt.graphics.Image;
-import org.eclipse.swt.graphics.ImageData;
 import org.eclipse.swt.graphics.RGB;
 import org.eclipse.swt.widgets.Display;
 
 /**
- * Utility class for managing OS resources associated with SWT controls such as colors, fonts,
- * images, etc.
+ * Utility class for managing OS resources associated with SWT controls such as colors, fonts, etc.
  * 
  * <p>
  * !!! IMPORTANT !!! Application code must explicitly invoke the <code>dispose()</code> method to
@@ -55,17 +47,6 @@ public class SwtResourceManager {
   //
   ////////////////////////////////////////////////////////////////////////////
   private static Map<RGB, Color> colorMap = new HashMap<>();
-
-  /**
-   * Returns the system {@link Color} matching the specific ID.
-   * 
-   * @param systemColorID the ID value for the color
-   * @return the system {@link Color} matching the specific ID
-   */
-  public static Color getColor(int systemColorID) {
-    Display display = Display.getCurrent();
-    return display.getSystemColor(systemColorID);
-  }
 
   /**
    * Returns a {@link Color} given its red, green and blue component values.
@@ -97,132 +78,6 @@ public class SwtResourceManager {
       color.dispose();
     }
     colorMap.clear();
-  }
-
-  ////////////////////////////////////////////////////////////////////////////
-  //
-  // Image
-  //
-  ////////////////////////////////////////////////////////////////////////////
-  /**
-   * Maps image paths to images.
-   */
-  private static Map<String, Image> imageMap = new HashMap<>();
-
-  /**
-   * Returns an {@link Image} encoded by the specified {@link InputStream}.
-   * 
-   * @param stream the {@link InputStream} encoding the image data
-   * @return the {@link Image} encoded by the specified input stream
-   */
-  protected static Image getImage(InputStream stream) throws IOException {
-    try {
-      Display display = Display.getCurrent();
-      ImageData data = new ImageData(stream);
-      if (data.transparentPixel > 0) {
-        return new Image(display, data, data.getTransparencyMask());
-      }
-      return new Image(display, data);
-    } finally {
-      stream.close();
-    }
-  }
-
-  /**
-   * Returns an {@link Image} stored in the file at the specified path.
-   * 
-   * @param path the path to the image file
-   * @return the {@link Image} stored in the file at the specified path
-   */
-  public static Image getImage(String path) {
-    Image image = imageMap.get(path);
-    if (image == null) {
-      try {
-        image = getImage(new FileInputStream(path));
-        imageMap.put(path, image);
-      } catch (Exception e) {
-        image = getMissingImage();
-        imageMap.put(path, image);
-      }
-    }
-    return image;
-  }
-
-  /**
-   * Returns an {@link Image} stored in the file at the specified path relative to the specified
-   * class.
-   * 
-   * @param clazz the {@link Class} relative to which to find the image
-   * @param path the path to the image file, if starts with <code>'/'</code>
[…]
