### 2022-09-15 d12e3e4
Add tests for the resource managers
