### 2022-09-15 ff322c9
Remove empty test setup functions
