+++ tests/org.corpus_tools.hexatomic.core.tests/src/main/java/org/eclipse/wb/swt/TestSwtResourceManager.java
@@ -7,13 +7,9 @@ import org.eclipse.swt.SWT;
 import org.eclipse.swt.graphics.Color;
 import org.eclipse.swt.graphics.Font;
 import org.eclipse.swt.graphics.RGB;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 class TestSwtResourceManager {
-
-  @BeforeEach
-  void setUp() {}
 
   /**
    * Tests correct caching and disposal of fonts.
