### 2022-09-15 53d8f16
Fix code smell
