+++ bundles/org.corpus_tools.hexatomic.graph/src/main/java/org/corpus_tools/hexatomic/graph/internal/SaltGraphStyler.java
@@ -217,9 +217,4 @@ public class SaltGraphStyler extends LabelProvider implements ISelfStyleProvider
     return null;
   }
 
-  @Override
-  public void dispose() {
-    super.dispose();
-  }
-
 }
