### 2022-09-15 d1c8ee8
Execute some of the graph tests faster by polling more often when waiting until condition comes true
