+++ tests/org.corpus_tools.hexatomic.it.tests/src/main/java/org/corpus_tools/hexatomic/it/tests/TestGraphEditor.java
@@ -54,6 +54,7 @@ import org.eclipse.swtbot.swt.finder.finders.UIThreadRunnable;
 import org.eclipse.swtbot.swt.finder.keyboard.Keyboard;
 import org.eclipse.swtbot.swt.finder.keyboard.KeyboardFactory;
 import org.eclipse.swtbot.swt.finder.keyboard.Keystrokes;
+import org.eclipse.swtbot.swt.finder.utils.SWTBotPreferences;
 import org.eclipse.swtbot.swt.finder.utils.SWTUtils;
 import org.eclipse.swtbot.swt.finder.utils.WidgetTextDescription;
 import org.eclipse.swtbot.swt.finder.waits.DefaultCondition;
@@ -78,7 +79,6 @@ class TestGraphEditor {
 
   private static final String DOC1_SALT_ID = "salt:/rootCorpus/subCorpus1/doc1";
   private static final String DOC1_TITLE = "doc1 (Graph Editor)";
-  private static final long GRAPH_LAYOUT_TIMEOUT = 5000;
   private static final String INF_STRUCT = "Inf-Struct";
   private static final String CONST = "const";
   private static final String SEARCH = "Search";
@@ -411,7 +411,7 @@ class TestGraphEditor {
     docMenu.click();
     assertNotNull(docMenu.contextMenu("Open with Graph Editor").click());
 
-    bot.waitUntil(new GraphLoadedCondition(0));
+    bot.waitUntil(new GraphLoadedCondition(0), SWTBotPreferences.TIMEOUT, 100);
   }
 
   void openMinimalProjectStructure() {
@@ -425,7 +425,7 @@ class TestGraphEditor {
     docMenu.click();
     assertNotNull(docMenu.contextMenu("Open with Graph Editor").click());
 
-    bot.waitUntil(new GraphLoadedCondition("document_1"));
+    bot.waitUntil(new GraphLoadedCondition("document_1"), SWTBotPreferences.TIMEOUT, 100);
   }
 
   void enterCommand(String command) {
@@ -462,7 +462,7 @@ class TestGraphEditor {
         public String getFailureMessage() {
           return "Second text segment was not checked";
         }
-      }, GRAPH_LAYOUT_TIMEOUT);
+      });
     }
   }
 
@@ -492,10 +492,12 @@ class TestGraphEditor {
     // Initially, the zoom is adjusted to match the height, so moving up/down should not do anything
     keyboard.pressShortcut(Keystrokes.DOWN);
     bot.waitUntil(
-        new ViewLocationReachedCondition(viewPort, new Point(origLocation.x, origLocation.y)));
+        new ViewLocationReachedCondition(viewPort, new Point(origLocation.x, origLocation.y)),
+        SWTBotPreferences.TIMEOUT, 100);
     keyboard.pressShortcut(Keystrokes.UP);
     bot.waitUntil(
-        new ViewLocationReachedCondition(viewPort, new Point(origLocation.x, origLocation.y)));
+        new ViewLocationReachedCondition(viewPort, new Point(origLocation.x, origLocation.y)),
+        SWTBotPreferences.TIMEOUT, 100);
 
     // Zoom in so we can move the view with the arrow keys
     // Use the mock keyboard for this since AWT/SWT keyboards don't map the keypad keys yet.
@@ -508,41 +510,53 @@ class TestGraphEditor {
     // Scroll with arrow keys (left, right, up, down) and check that that view has been moved
     keyboard.pressShortcut(Keystrokes.RIGHT);
     bot.waitUntil(
[…]
