### 2022-09-20 1e72849
Add empty dialog, command, handler and menu entry for re-assign node names
