+++ bundles/org.corpus_tools.hexatomic.core/src/main/java/org/corpus_tools/hexatomic/core/ui/ReassignNodeNamesDialog.java
@@ -1,0 +1,11 @@
+package org.corpus_tools.hexatomic.core.ui;
+
+import org.eclipse.jface.dialogs.Dialog;
+import org.eclipse.swt.widgets.Shell;
+
+public class ReassignNodeNamesDialog extends Dialog {
+
+  public ReassignNodeNamesDialog(Shell shell) {
+    super(shell);
+  }
+}
