### 2022-09-20 8912fcb
Use a standard JFace question dialog
