### 2022-09-20 ccb2417
Implement re-assigning the node names
