### 2022-09-23 1bf5c67
Apply suggestions from code review
Co-authored-by: Stephan Druskat <sdruskat@users.noreply.github.com>
