### 2022-09-23 1c56304
Add a selection parameter to the command, so we can apply the refactoring to only a selection of the corpus.
