+++ bundles/org.corpus_tools.hexatomic.core/src/main/java/org/corpus_tools/hexatomic/core/handlers/ReassignNodeNamesHandler.java
@@ -1,20 +1,33 @@
 package org.corpus_tools.hexatomic.core.handlers;
 
 import java.lang.reflect.InvocationTargetException;
+import java.util.Arrays;
+import java.util.HashSet;
 import java.util.List;
-import java.util.Optional;
+import java.util.Set;
+import java.util.TreeSet;
 import java.util.stream.Collectors;
 import javax.inject.Inject;
+import javax.inject.Named;
+import org.corpus_tools.hexatomic.core.CommandParams;
 import org.corpus_tools.hexatomic.core.ProjectManager;
 import org.corpus_tools.hexatomic.core.errors.ErrorService;
+import org.corpus_tools.salt.common.SCorpus;
+import org.corpus_tools.salt.common.SCorpusDocumentRelation;
+import org.corpus_tools.salt.common.SCorpusRelation;
 import org.corpus_tools.salt.common.SDocument;
 import org.corpus_tools.salt.common.SDocumentGraph;
 import org.corpus_tools.salt.common.SSpan;
 import org.corpus_tools.salt.common.SStructure;
 import org.corpus_tools.salt.common.SToken;
+import org.corpus_tools.salt.core.GraphTraverseHandler;
+import org.corpus_tools.salt.core.SGraph.GRAPH_TRAVERSE_TYPE;
+import org.corpus_tools.salt.core.SNode;
+import org.corpus_tools.salt.core.SRelation;
 import org.eclipse.core.runtime.IProgressMonitor;
 import org.eclipse.e4.core.di.annotations.CanExecute;
 import org.eclipse.e4.core.di.annotations.Execute;
+import org.eclipse.e4.core.di.annotations.Optional;
 import org.eclipse.jface.dialogs.MessageDialog;
 import org.eclipse.jface.dialogs.ProgressMonitorDialog;
 import org.eclipse.jface.operation.IRunnableWithProgress;
@@ -28,11 +41,40 @@ public class ReassignNodeNamesHandler {
   @Inject
   ProjectManager projectManager;
 
+  private final class DocumentForCorpusTraverser implements GraphTraverseHandler {
+
+    private final Set<String> childDocumentIds = new TreeSet<String>();
+
+    @Override
+    public void nodeReached(GRAPH_TRAVERSE_TYPE traversalType, String traversalId, SNode currNode,
+        SRelation<SNode, SNode> relation, SNode fromNode, long order) {
+      // If the target node is a document, add it to our set
+      if (currNode instanceof SDocument) {
+        childDocumentIds.add(currNode.getId());
+      }
+    }
+
+    @Override
+    public void nodeLeft(GRAPH_TRAVERSE_TYPE traversalType, String traversalId, SNode currNode,
+        SRelation<SNode, SNode> relation, SNode fromNode, long order) {
+      // We already add the documents when we reach the nodes.
+    }
+
+    @Override
+    public boolean checkConstraint(GRAPH_TRAVERSE_TYPE traversalType, String traversalId,
+        @SuppressWarnings("rawtypes") SRelation relation, SNode currNode, long order) {
+      // Traverse to sub-corpora and documents
+      return relation == null || relation instanceof SCorpusDocumentRelation
+          || relation instanceof SCorpusRelation;
+    }
+  }
+
+
   protected class NameAssigner implements IRunnableWithProgress {
 
-    List<String> documentIds;
+    private Set<String> documentIds;
 
-    public NameAssigner(List<String> documentIds) {
+    public NameAssigner(Set<String> documentIds) {
       this.documentIds = documentIds;
     }
 
@@ -49,7 +91,7 @@ public class ReassignNodeNamesHandler {
         }
         monitor.subTask(docId);
 
[…]
