### 2022-09-23 34e8c91
Get the selected corpora/documents from the selection service
