### 2022-09-23 5ac8979
Add test for renaming nodes in all documents
