+++ tests/org.corpus_tools.hexatomic.it.tests/src/main/java/org/corpus_tools/hexatomic/it/tests/TestReassignNodeInNames.java
@@ -1,0 +1,155 @@
+package org.corpus_tools.hexatomic.it.tests;
+
+import static org.junit.Assert.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import java.io.File;
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Optional;
+import org.corpus_tools.hexatomic.core.CommandParams;
+import org.corpus_tools.hexatomic.core.ProjectManager;
+import org.corpus_tools.salt.common.SCorpusGraph;
+import org.corpus_tools.salt.common.SDocument;
+import org.corpus_tools.salt.common.SDocumentGraph;
+import org.corpus_tools.salt.common.SStructuredNode;
+import org.corpus_tools.salt.common.SToken;
+import org.corpus_tools.salt.core.SNode;
+import org.eclipse.core.commands.ParameterizedCommand;
+import org.eclipse.e4.core.commands.ECommandService;
+import org.eclipse.e4.core.commands.EHandlerService;
+import org.eclipse.e4.core.contexts.ContextInjectionFactory;
+import org.eclipse.e4.core.contexts.IEclipseContext;
+import org.eclipse.e4.core.services.events.IEventBroker;
+import org.eclipse.emf.common.util.URI;
+import org.eclipse.swtbot.e4.finder.widgets.SWTWorkbenchBot;
+import org.eclipse.swtbot.swt.finder.waits.DefaultCondition;
+import org.eclipse.swtbot.swt.finder.widgets.SWTBotShell;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.MethodOrderer.OrderAnnotation;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestMethodOrder;
+
+@TestMethodOrder(OrderAnnotation.class)
+class TestReassignNodeInNames {
+
+
+  private static final String REFACTOR = "Refactor";
+
+  private static final String EDIT = "Edit";
+
+  private static final String ALL_DOCUMENTS_LABEL = "Re-assign node names for all documents";
+
+  private final SWTWorkbenchBot bot = new SWTWorkbenchBot(TestHelper.getEclipseContext());
+
+  private URI exampleProjectUri;
+  private ECommandService commandService;
+  private EHandlerService handlerService;
+
+  private IEventBroker events;
+  private ProjectManager projectManager;
+
+  @BeforeEach
+  void setup() {
+    TestHelper.setKeyboardLayout();
+    IEclipseContext ctx = TestHelper.getEclipseContext();
+
+    projectManager = ContextInjectionFactory.make(ProjectManager.class, ctx);
+
+    events = ctx.get(IEventBroker.class);
+    assertNotNull(events);
+
+    commandService = ctx.get(ECommandService.class);
+    assertNotNull(commandService);
+
+    handlerService = ctx.get(EHandlerService.class);
+    assertNotNull(handlerService);
+
+    File exampleProjectDirectory = new File("../org.corpus_tools.hexatomic.core.tests/"
+        + "src/main/resources/org/corpus_tools/hexatomic/core/example-corpus/");
+    assertTrue(exampleProjectDirectory.isDirectory());
+    exampleProjectUri = URI.createFileURI(exampleProjectDirectory.getAbsolutePath());
+  }
+
+  void openDefaultExample() {
+    // Programmatically open the example corpus
+    Map<String, String> params = new HashMap<>();
+    params.put(CommandParams.LOCATION, exampleProjectUri.toFileString());
+    params.put(CommandParams.FORCE_CLOSE, "true");
+    ParameterizedCommand cmd = commandService
[…]
