+++ tests/org.corpus_tools.hexatomic.it.tests/src/main/java/org/corpus_tools/hexatomic/it/tests/TestReassignNodeInNames.java
@@ -1,6 +1,6 @@
 package org.corpus_tools.hexatomic.it.tests;
 
-import static org.junit.Assert.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
@@ -24,9 +24,12 @@ import org.eclipse.e4.core.contexts.ContextInjectionFactory;
 import org.eclipse.e4.core.contexts.IEclipseContext;
 import org.eclipse.e4.core.services.events.IEventBroker;
 import org.eclipse.emf.common.util.URI;
+import org.eclipse.swtbot.e4.finder.widgets.SWTBotView;
 import org.eclipse.swtbot.e4.finder.widgets.SWTWorkbenchBot;
 import org.eclipse.swtbot.swt.finder.waits.DefaultCondition;
 import org.eclipse.swtbot.swt.finder.widgets.SWTBotShell;
+import org.eclipse.swtbot.swt.finder.widgets.SWTBotTreeItem;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.MethodOrderer.OrderAnnotation;
 import org.junit.jupiter.api.Test;
@@ -36,11 +39,30 @@ import org.junit.jupiter.api.TestMethodOrder;
 class TestReassignNodeInNames {
 
 
+  private final class RefactoringFinishedCondition extends DefaultCondition {
+    @Override
+    public boolean test() throws Exception {
+      boolean hasMatchingShell = Arrays.asList(bot.shells()).stream()
+          .anyMatch(s -> s.isOpen() && SELECTED_DIALOG_CAPTION.equals(s.getText()));
+      return !hasMatchingShell && projectManager.canUndo();
+    }
+
+    @Override
+    public String getFailureMessage() {
+      return "Renaming operation did not finish.";
+    }
+  }
+
+  private static final String SELECTED_MENU_LABEL = "Re-assign node names for selection";
+  private static final String DOC1_ID = "salt:/rootCorpus/subCorpus1/doc1";
+  private static final String DOC2_ID = "salt:/rootCorpus/subCorpus1/doc2";
+  private static final String SELECTED_DIALOG_CAPTION =
+      "Re-assign node names for selected documents";
   private static final String REFACTOR = "Refactor";
-
   private static final String EDIT = "Edit";
-
   private static final String ALL_DOCUMENTS_LABEL = "Re-assign node names for all documents";
+  private static final String CORPUS_EDITOR_PART_ID =
+      "org.corpus_tools.hexatomic.corpusedit.part.corpusstructure";
 
   private final SWTWorkbenchBot bot = new SWTWorkbenchBot(TestHelper.getEclipseContext());
 
@@ -73,7 +95,17 @@ class TestReassignNodeInNames {
     exampleProjectUri = URI.createFileURI(exampleProjectDirectory.getAbsolutePath());
   }
 
-  void openDefaultExample() {
+  @AfterEach
+  void close() {
+    // Programmatically close the example corpus by opening a new fresh one
+    Map<String, String> params = new HashMap<>();
+    params.put(CommandParams.FORCE_CLOSE, "true");
+    ParameterizedCommand cmd = commandService
+        .createCommand("org.corpus_tools.hexatomic.core.command.new_salt_project", params);
+    handlerService.executeHandler(cmd);
+  }
+
+  SWTBotView openDefaultExample() {
     // Programmatically open the example corpus
     Map<String, String> params = new HashMap<>();
     params.put(CommandParams.LOCATION, exampleProjectUri.toFileString());
@@ -81,19 +113,40 @@ class TestReassignNodeInNames {
[…]
