### 2022-09-23 d8d27dd
Also test that the refactoring can be undone
