### 2022-09-23 77b4eb6
Close project in TestProjectManager
