+++ tests/org.corpus_tools.hexatomic.it.tests/src/main/java/org/corpus_tools/hexatomic/it/tests/TestProjectManager.java
@@ -46,6 +46,7 @@ import org.eclipse.swtbot.swt.finder.SWTBot;
 import org.eclipse.swtbot.swt.finder.finders.UIThreadRunnable;
 import org.eclipse.swtbot.swt.finder.waits.ICondition;
 import org.eclipse.swtbot.swt.finder.widgets.SWTBotShell;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.MethodOrderer.OrderAnnotation;
 import org.junit.jupiter.api.Order;
@@ -89,6 +90,11 @@ class TestProjectManager {
 
     errorService.clearLastException();
 
+    TestHelper.executeNewProjectCommand(commandService, handlerService);
+  }
+
+  @AfterEach
+  void close() {
     TestHelper.executeNewProjectCommand(commandService, handlerService);
   }
 
