### 2022-09-23 7f1d983
Rename refactoring operation to "Automatically generate node names ..."
