+++ docs/user/src/usage/refactoring/README.md
@@ -1,0 +1,10 @@
+# Refactoring
+
+Hexatomic offers some so-called refactoring operations, which perform changes on
+bulk. Refactoring operations can be applied to the whole project (via the menu
+**Edit -> Refactor**) or on selected documents or sub-corpora, by clicking right
+on selected item in the [“Corpus Structure” editor](../corpus-structure.md) and
+opening the **Refactor** context menu item.
+
+## Automatically generate node names
+
