### 2022-09-23 cc4a8bb
Document the "Automatically generate node names" refactoring operation
