+++ docs/user/src/usage/refactoring/README.md
@@ -8,3 +8,13 @@ opening the **Refactor** context menu item.
 
 ## Automatically generate node names
 
+Node names are e.g. used in the [Graph Editor
+console](../graph-editor/console.md). If these node names are not unique or too
+complex to easily reference them, you can replace all node names with
+automatically generated ones. Tokens will have the pattern `t1`, `t2`, `t3` and
+so on. They will be ordered by the position of the token in the text. Other
+nodes will have generic names like `n1`, `n2`, `n3` and so on.
+
+**Warning:** Some corpus preparation pipelines use the node names to merge files
+in different formats. If your corpus preparation pipeline requires specific node
+names, you should not use this refactoring operation.
