### 2022-09-23 0f36ee0
Reduce complexity of method
