### 2022-09-23 a481c88
Replace some strings with constants
