### 2022-09-23 1ecd615
Fix smaller code smells
