### 2022-09-23 a4f9990
Add test for reverting to last checkpoint
