+++ tests/org.corpus_tools.hexatomic.core.tests/src/main/java/org/corpus_tools/hexatomic/core/TestProjectManager.java
@@ -294,6 +294,39 @@ class TestProjectManager {
     }
   }
 
+  @Test
+  public void testRevertToLastCheckpoint() {
+
+    projectManager.open(exampleProjectUri);
+    assertFalse(projectManager.canUndo());
+    assertFalse(projectManager.canRedo());
+
+    Optional<SDocument> document = projectManager.getDocument(DOC3_ID, true);
+    assertTrue(document.isPresent());
+    if (document.isPresent()) {
+      SDocumentGraph docGraph = document.get().getDocumentGraph();
+      assertNotNull(docGraph);
+
+      // Change a token annotation without creating a checkpoint
+      List<SToken> token = docGraph.getSortedTokenByText();
+      token.get(0).createAnnotation("test", "anno", "0");
+      assertTrue(token.get(0).containsLabel("test::anno"));
+      token.get(1).createAnnotation("test", "anno", "1");
+      assertTrue(token.get(1).containsLabel("test::anno"));
+
+      assertFalse(projectManager.canRedo());
+      assertFalse(projectManager.canRedo());
+
+      // Revert changes and check that the created annotations are gone
+      projectManager.revertToLastCheckpoint();
+      assertFalse(token.get(0).containsLabel("test::anno"));
+      assertFalse(token.get(1).containsLabel("test::anno"));
+
+      assertFalse(projectManager.canRedo());
+      assertFalse(projectManager.canRedo());
+    }
+  }
+
   /**
    * If a checkpoint was added due to user interaction, all "redo" operations should become invalid.
    */
