+++ bundles/org.corpus_tools.hexatomic.core/src/main/java/org/corpus_tools/hexatomic/core/ProjectManager.java
@@ -619,7 +619,6 @@ public class ProjectManager {
    */
   public void revertToLastCheckpoint() {
 
-
     if (!uncommittedChanges.isEmpty()) {
       // Group uncommitted changes in a changeset
       ChangeSet uncommitedChangeSet = new ChangeSet(uncommittedChanges);
