### 2022-09-23 90ae4de
If nothing is selected, don't update anything
