### 2022-09-23 face253
Replace string with existing constant
