### 2022-09-26 063adbf
Get TLI index and overlapping token
