### 2022-09-26 ad6eb67
Enable scrolling of the table
