+++ bundles/org.corpus_tools.hexatomic.timeline/src/main/java/org/corpus_tools/hexatomic/timeline/TimelineAlignEditor.java
@@ -9,7 +9,11 @@ import org.corpus_tools.salt.common.SDocumentGraph;
 import org.eclipse.e4.ui.model.application.ui.basic.MPart;
 import org.eclipse.jface.layout.GridDataFactory;
 import org.eclipse.nebula.widgets.nattable.NatTable;
+import org.eclipse.nebula.widgets.nattable.grid.GridRegion;
 import org.eclipse.nebula.widgets.nattable.layer.DataLayer;
+import org.eclipse.nebula.widgets.nattable.selection.SelectionLayer;
+import org.eclipse.nebula.widgets.nattable.selection.config.DefaultRowSelectionLayerConfiguration;
+import org.eclipse.nebula.widgets.nattable.viewport.ViewportLayer;
 import org.eclipse.swt.SWT;
 import org.eclipse.swt.layout.GridLayout;
 import org.eclipse.swt.widgets.Composite;
@@ -35,11 +39,17 @@ public class TimelineAlignEditor {
     this.graph = getGraph();
     bodyDataProvider.setGraph(graph);
 
+    // Define scrollable body layer
     DataLayer bodyDataLayer = new DataLayer(bodyDataProvider);
+    SelectionLayer selectionLayer = new SelectionLayer(bodyDataLayer);
+    selectionLayer.addConfiguration(new DefaultRowSelectionLayerConfiguration());
+    ViewportLayer viewportLayer = new ViewportLayer(selectionLayer);
+    viewportLayer.setRegionName(GridRegion.BODY);
+
 
     NatTable natTable = new NatTable(parent,
-        SWT.NO_REDRAW_RESIZE | SWT.DOUBLE_BUFFERED | SWT.BORDER, bodyDataLayer);
-
+        SWT.NO_REDRAW_RESIZE | SWT.DOUBLE_BUFFERED | SWT.BORDER | SWT.H_SCROLL | SWT.V_SCROLL,
+        viewportLayer);
     GridDataFactory.fillDefaults().grab(true, true).applyTo(natTable);
   }
 
