+++ bundles/org.corpus_tools.hexatomic.timeline/src/main/java/org/corpus_tools/hexatomic/timeline/TimelineAlignEditor.java
@@ -3,14 +3,23 @@ package org.corpus_tools.hexatomic.timeline;
 import javax.annotation.PostConstruct;
 import javax.inject.Inject;
 import org.corpus_tools.hexatomic.core.ProjectManager;
+import org.corpus_tools.hexatomic.timeline.internal.data.TextualDsHeaderDataProvider;
 import org.corpus_tools.hexatomic.timeline.internal.data.TimelineTokenDataProvider;
+import org.corpus_tools.hexatomic.timeline.internal.data.TliCornerDataProvider;
+import org.corpus_tools.hexatomic.timeline.internal.data.TliRowHeaderDataProvider;
 import org.corpus_tools.salt.common.SDocument;
 import org.corpus_tools.salt.common.SDocumentGraph;
 import org.eclipse.e4.ui.model.application.ui.basic.MPart;
 import org.eclipse.jface.layout.GridDataFactory;
 import org.eclipse.nebula.widgets.nattable.NatTable;
 import org.eclipse.nebula.widgets.nattable.grid.GridRegion;
+import org.eclipse.nebula.widgets.nattable.grid.layer.ColumnHeaderLayer;
+import org.eclipse.nebula.widgets.nattable.grid.layer.CornerLayer;
+import org.eclipse.nebula.widgets.nattable.grid.layer.GridLayer;
+import org.eclipse.nebula.widgets.nattable.grid.layer.RowHeaderLayer;
+import org.eclipse.nebula.widgets.nattable.grid.layer.config.DefaultGridLayerConfiguration;
 import org.eclipse.nebula.widgets.nattable.layer.DataLayer;
+import org.eclipse.nebula.widgets.nattable.layer.ILayer;
 import org.eclipse.nebula.widgets.nattable.selection.SelectionLayer;
 import org.eclipse.nebula.widgets.nattable.selection.config.DefaultRowSelectionLayerConfiguration;
 import org.eclipse.nebula.widgets.nattable.viewport.ViewportLayer;
@@ -29,6 +38,12 @@ public class TimelineAlignEditor {
   @Inject
   private TimelineTokenDataProvider bodyDataProvider;
 
+  @Inject
+  private TextualDsHeaderDataProvider columnHeaderDataProvider;
+
+  @Inject
+  private TliRowHeaderDataProvider tliRowDataProvider;
+
   private SDocumentGraph graph;
 
 
@@ -38,6 +53,8 @@ public class TimelineAlignEditor {
 
     this.graph = getGraph();
     bodyDataProvider.setGraph(graph);
+    columnHeaderDataProvider.setGraph(graph);
+    tliRowDataProvider.setGraph(graph);
 
     // Define scrollable body layer
     DataLayer bodyDataLayer = new DataLayer(bodyDataProvider);
@@ -46,10 +63,31 @@ public class TimelineAlignEditor {
     ViewportLayer viewportLayer = new ViewportLayer(selectionLayer);
     viewportLayer.setRegionName(GridRegion.BODY);
 
+    // Header layer
+    DataLayer columnHeaderDataLayer = new DataLayer(columnHeaderDataProvider);
+    ILayer columnHeaderLayer =
+        new ColumnHeaderLayer(columnHeaderDataLayer, viewportLayer, selectionLayer);
+
+    // Row layer
+    DataLayer rowHeaderDataLayer = new DataLayer(tliRowDataProvider);
+    ILayer rowHeaderLayer =
+        new RowHeaderLayer(rowHeaderDataLayer, viewportLayer, selectionLayer);
+
+    // Corner
+    final TliCornerDataProvider cornerDataProvider =
[…]
