+++ bundles/org.corpus_tools.hexatomic.timeline/src/main/java/org/corpus_tools/hexatomic/timeline/internal/data/TextualDsHeaderDataProvider.java
@@ -1,0 +1,42 @@
+package org.corpus_tools.hexatomic.timeline.internal.data;
+
+import org.corpus_tools.salt.common.SDocumentGraph;
+import org.eclipse.e4.core.di.annotations.Creatable;
+import org.eclipse.nebula.widgets.nattable.data.IDataProvider;
+
+@Creatable
+public class TextualDsHeaderDataProvider implements IDataProvider {
+
+  private SDocumentGraph graph;
+
+  public void setGraph(SDocumentGraph graph) {
+    this.graph = graph;
+  }
+
+  @Override
+  public Object getDataValue(int columnIndex, int rowIndex) {
+    if (columnIndex == 0) {
+      return "Time";
+    } else {
+      return graph.getTextualDSs().get(columnIndex - 1).getName();
+    }
+  }
+
+  @Override
+  public void setDataValue(int columnIndex, int rowIndex, Object newValue) {
+    // TODO Auto-generated method stub
+
+  }
+
+  @Override
+  public int getColumnCount() {
+    // The timeline items are a column + every textual data source is a column
+    return 1 + graph.getTextualDSs().size();
+  }
+
+  @Override
+  public int getRowCount() {
+    return 1;
+  }
+
+}
