### 2022-10-07 #4726 @reckart
@sdruskat Hi :) As part of reviewing the submission, I'd like to ask a couple of questions as I go along. Here are some initial ones:

* Are there any community guidelines that go along with Hexatomic? Can you provide a URL?
* Are there any examples / use-cases / tutorials? Can you provide a URL?
* Could you provide a list of all depenencies of hexatomic and their respective licences please?

I have downloaded the MacOS build and tried to run it, but the application crashes with this dialog:

Is this a known issue?
