### 2022-10-10 hexatomic/hexatomic#409 @sdruskat
Hexatomic crashes on MacOS
**Describe the bug**

As [reported in the JOSS review](https://github.com/openjournals/joss-reviews/issues/4825#issuecomment-1272056131), Hexatomic crashes on MacOS.

This issue tracks that bug. @reckart, could you please complete the missing info in the sections below, so that we can reproduce and fix this? Many thanks 🙏!

**To reproduce the bug**
Steps to reproduce the behavior:
1. Go to '...'
2. Click on '....'
3. Scroll down to '....'
4. See error

**Expected behavior**

- Hexatomic runs fine on MacOS.

**Screenshots**

![](https://user-images.githubusercontent.com/1410238/194635304-9e366fac-0c60-4b39-aab4-1fde1193e37f.png).

 - Operating System: MacOS **[version?]**
 - Hexatomic Version: 1.0.1
 - Java Version: **[Version? Using the bundled JRE?]**

**Additional context**
Add any other context about the problem here.
