### 2022-10-10 hexatomic/hexatomic#409 @reckart
* Download hexatomic-1.0.1-macosx.cocoa.x86_64.zip
* Extract it
* Double-click on Hexatomic.app
* See error
