### 2022-10-10 hexatomic/hexatomic#409 @reckart
If I start from the command line, I get a different behavior:

```
% ./hexatomic
12:13:10.749 [main] INFO  o.c.h.c.ApplicationLifecycle - Logging configured from logback.xml in the Hexatomic root folder
12:13:10.857 [main] INFO  o.c.h.f.Activator - Using /.../Downloads/hexatomic.app/Contents/Eclipse/configuration/org.eclipse.osgi/38/0/.cp/pepper-plugins as Pepper module resource folder
12:13:10.894 [main] INFO  o.c.h.f.Activator - Installing Pepper module from file:/.../Downloads/hexatomic.app/Contents/Eclipse/configuration/org.eclipse.osgi/38/0/.cp/pepper-plugins/exmaralda-emf-api-1.2.2.jar
12:13:11.003 [main] INFO  o.c.h.f.Activator - Installing Pepper module from file:/.../Downloads/hexatomic.app/Contents/Eclipse/configuration/org.eclipse.osgi/38/0/.cp/pepper-plugins/gson-2.8.7.jar
12:13:11.023 [main] INFO  o.c.h.f.Activator - Installing Pepper module from file:/.../Downloads/hexatomic.app/Contents/Eclipse/configuration/org.eclipse.osgi/38/0/.cp/pepper-plugins/pepperModules-EXMARaLDAModules-1.3.1.jar
12:13:11.088 [main] INFO  o.c.h.f.Activator - Installing Pepper module from file:/.../Downloads/hexatomic.app/Contents/Eclipse/configuration/org.eclipse.osgi/38/0/.cp/pepper-plugins/pepperModules-GraphAnnoModules-2.0.1.jar
12:13:11.117 [main] INFO  o.c.h.f.Activator - Installing Pepper module from file:/.../Downloads/hexatomic.app/Contents/Eclipse/configuration/org.eclipse.osgi/38/0/.cp/pepper-plugins/pepperModules-PAULAModules-1.3.0.jar
12:13:12.069 [Worker-0: Update Job] INFO  o.c.h.c.u.UpdateRunner - Provisioning session created
```

The app window then also appears.

Closing the app and trying with a double-click on Hexatomic.app again brings up the error from before about `.metadata` being read-only.
