### 2022-10-10 #4825 @sdruskat
> @sdruskat Hi :) As part of reviewing the submission, I'd like to ask a couple of questions as I go along. Here are some initial ones:
>
>     * Are there any community guidelines that go along with Hexatomic? Can you provide a URL?

There are the community guidelines in the GitHub repository itself:

- Code of Conduct: https://github.com/hexatomic/hexatomic/blob/develop/CODE_OF_CONDUCT.md
- Contribution guidelines: https://github.com/hexatomic/hexatomic/blob/develop/CONTRIBUTING.md
- And the README linking out to the relevant documentation: https://github.com/hexatomic/hexatomic/blob/develop/README.md

Documentation:

- For users: https://hexatomic.github.io/hexatomic/user/
- For developers and maintainers: https://hexatomic.github.io/hexatomic/dev/

Please let us know if you're missing anything.

>     * Are there any examples / use-cases / tutorials? Can you provide a URL?

There is a brief ["how to create a corpus from scratch"](https://hexatomic.github.io/hexatomic/user/v1.0/howtos/corpus-creation.html) in the [user documentation](https://hexatomic.github.io/hexatomic/user/v1.0/).

>     * Could you provide a list of all depenencies of hexatomic and their respective licences please?

All dependencies and their licenses are listed

- in the [THIRD-PARTY](https://github.com/hexatomic/hexatomic/tree/develop/THIRD-PARTY) folder in the repository,
- Listed as references (with license information) in the [`CITATION.cff` file](https://github.com/hexatomic/hexatomic/blob/develop/CITATION.cff#L40-L1061).

I'll also supply a list of dependencies with licenses in a new comment.

> I have downloaded the MacOS build from https://github.com/hexatomic/hexatomic/releases/tag/v1.0.1 and tried to run it, but the application crashes with this dialog:
>
> <img alt="Screenshot 2022-10-07 at 20 59 50" width="558" src="https://user-images.githubusercontent.com/1410238/194635304-9e366fac-0c60-4b39-aab4-1fde1193e37f.png">
>
> Is this a known issue?

Hi @reckart, thanks for flagging this. We're tracking the bug in https://github.com/hexatomic/hexatomic/issues/409, and it'd be great if you helped us complete the missing info in the bug so that we can reproduce and fix this bug. Thanks 🙏!
