### 2022-10-10 #4825 @sdruskat
Dependencies and licenses:

|Dependency|License|
| --- | --- |
|SWT Resource Manager|EPL-1.0|
|ANTLR 4 Runtime|BSD-3-Clause|
|Guava: Google Core Libraries for Java|Apache-2.0|
|ICU4J|EPL-1.0|
|JSON in Java|JSON|
|Neo Visionaries Internationalization Package|Apache-2.0|
|Simple Logging Facade for Java|MIT|
|Stax2 API|BSD-2-Clause|
|Woodstox|Apache-2.0|
|ch.qos.logback.classic|EPL-1.0|
|ch.qos.logback.core|EPL-1.0|
|com.sun.jna|(Apache-2.0 OR LGPL-2.1 -> Apache-2.0)|
|com.sun.jna.platform|Apache-2.0|
|javax.annotation|EPL-1.0|
|javax.inject|EPL-1.0|
|org.apache.batik.constants|Apache-2.0|
|org.apache.batik.css|Apache-2.0|
|org.apache.batik.i18n|Apache-2.0|
|org.apache.batik.util|Apache-2.0|
|org.apache.commons.io|Apache-2.0|
|org.apache.commons.jxpath|Apache-2.0|
|org.apache.commons.lang3|Apache-2.0|
|org.apache.commons.logging|Apache-2.0|
|org.apache.felix.scr|Apache-2.0|
|org.apache.xmlgraphics|Apache-2.0|
|org.bouncycastle.bcpg|MIT-like (https://www.bouncycastle.org/licence.html)|
|org.bouncycastle.bcprov|MIT-like (https://www.bouncycastle.org/licence.html)|
|org.corpus-tools.pepper-framework|Apache-2.0|
|org.corpus-tools.salt-api|Apache-2.0|
|org.eclipse.core.commands|EPL-2.0|
|org.eclipse.core.contenttype|EPL-2.0|
|org.eclipse.core.databinding|EPL-2.0|
|org.eclipse.core.databinding.observable|EPL-2.0|
|org.eclipse.core.databinding.property|EPL-2.0|
|org.eclipse.core.expressions|EPL-2.0|
|org.eclipse.core.jobs|EPL-2.0|
|org.eclipse.core.runtime|EPL-2.0|
|org.eclipse.draw2d|EPL-2.0|
|org.eclipse.e4.core.commands|EPL-2.0|
|org.eclipse.e4.core.contexts|EPL-2.0|
|org.eclipse.e4.core.di|EPL-2.0|
|org.eclipse.e4.core.di.annotations|EPL-2.0|
|org.eclipse.e4.core.di.extensions|EPL-2.0|
|org.eclipse.e4.core.di.extensions.supplier|EPL-2.0|
|org.eclipse.e4.core.services|EPL-2.0|
|org.eclipse.e4.emf.xpath|EPL-2.0|
|org.eclipse.e4.ui.bindings|EPL-2.0|
|org.eclipse.e4.ui.css.core|EPL-2.0|
|org.eclipse.e4.ui.css.swt|EPL-2.0|
|org.eclipse.e4.ui.css.swt.theme|EPL-2.0|
|org.eclipse.e4.ui.di|EPL-2.0|
|org.eclipse.e4.ui.dialogs|EPL-2.0|
|org.eclipse.e4.ui.model.workbench|EPL-2.0|
|org.eclipse.e4.ui.services|EPL-2.0|
|org.eclipse.e4.ui.widgets|EPL-2.0|
|org.eclipse.e4.ui.workbench|EPL-2.0|
|org.eclipse.e4.ui.workbench.addons.swt|EPL-2.0|
|org.eclipse.e4.ui.workbench.renderers.swt|EPL-2.0|
|org.eclipse.e4.ui.workbench.swt|EPL-2.0|
|org.eclipse.e4.ui.workbench3|EPL-2.0|
|org.eclipse.emf.common|EPL-2.0|
|org.eclipse.emf.ecore|EPL-2.0|
|org.eclipse.emf.ecore.change|EPL-2.0|
|org.eclipse.emf.ecore.xmi|EPL-2.0|
|org.eclipse.equinox.app|EPL-2.0|
|org.eclipse.equinox.common|EPL-2.0|
|org.eclipse.equinox.event|EPL-2.0|
|org.eclipse.equinox.p2.artifact.repository|EPL-2.0|
|org.eclipse.equinox.p2.core|EPL-2.0|
|org.eclipse.equinox.p2.director|EPL-2.0|
|org.eclipse.equinox.p2.engine|EPL-2.0|
|org.eclipse.equinox.p2.jarprocessor|EPL-2.0|
|org.eclipse.equinox.p2.metadata|EPL-2.0|
|org.eclipse.equinox.p2.metadata.repository|EPL-2.0|
|org.eclipse.equinox.p2.operations|EPL-2.0|
|org.eclipse.equinox.p2.repository|EPL-2.0|
|org.eclipse.equinox.p2.ui|EPL-2.0|
|org.eclipse.equinox.preferences|EPL-2.0|
|org.eclipse.equinox.registry|EPL-2.0|
|org.eclipse.equinox.security|EPL-2.0|
|org.eclipse.equinox.security.ui|EPL-2.0|
|org.eclipse.equinox.simpleconfigurator|EPL-2.0|
|org.eclipse.help|EPL-2.0|
|org.eclipse.jface|EPL-2.0|
|org.eclipse.jface.databinding|EPL-2.0|
|org.eclipse.jface.notifications|EPL-2.0|
|org.eclipse.jface.text|EPL-2.0|
|org.eclipse.nebula.widgets.chips|EPL-2.0|
|org.eclipse.nebula.widgets.nattable.core|EPL-2.0|
|org.eclipse.nebula.widgets.opal.commons|EPL-2.0|
|org.eclipse.osgi|EPL-2.0|
|org.eclipse.osgi.services|EPL-2.0|
|org.eclipse.osgi.util|EPL-2.0|
|org.eclipse.swt|EPL-2.0|
|org.eclipse.swt.cocoa.macosx.x86_64|EPL-1.0|
|org.eclipse.swt.gtk.linux.x86_64|EPL-2.0|
|org.eclipse.swt.win32.win32.x86_64|EPL-1.0|
|org.eclipse.text|EPL-2.0|
|org.eclipse.ui|EPL-2.0|
|org.eclipse.ui.workbench|EPL-2.0|
|org.eclipse.urischeme|EPL-2.0|
|org.eclipse.zest.core|EPL-1.0|
|org.eclipse.zest.layouts|EPL-1.0|
|org.hamcrest.core|BSD-2-Clause|
|org.hamcrest.library|BSD-2-Clause|
|org.sat4j.core|LGPL-2.1|
|org.sat4j.pb|EPL-1.0|
|org.tukaani.xz|CC0-1.0|
|org.w3c.css.sac|W3C|
|org.w3c.dom.events|EPL-1.0|
|org.w3c.dom.smil|EPL-1.0|
|org.w3c.dom.svg|EPL-1.0|
|pepper-framework|Apache-2.0|
|salt-api|Apache-2.0|
|salt-extensions|Apache-2.0
