### 2022-10-10 hexatomic/hexatomic#409 @sdruskat
Thanks very much for these info, @reckart! We'll get this fixed as soon as possible.

@thomaskrause, can you get your hands on a MacOS 12.6 (VM?) to test this behaviour?
