### 2022-10-11 hexatomic/hexatomic#409 @thomaskrause
@reckart Did you extract the ZIP file to the `Applications` folder or to some other location? The error message about the `.metadata` file being read-only was caused by some newer MacOS sandbox mechanism.

In the [installation instructions](https://hexatomic.github.io/hexatomic/user/v1.0/installation.html), we had to add specific instructions for MacOS because of this sandbox and other, older security mechanisms. Can you check if the problem persists when you move the downloaded ZIP-file to the `Applications` folder? Manual unzipping should not be necessary.

We might want to extend the release notes to mention the different installation process on MacOS more prominently. I also find some of the wording in the instructions needs some clarification to not further confuse users. Like "For Mac OS X, we provide an .app file called hexatomic.app.", when the user never sees the ".app" extensions when following the instructions, but only the installed Icon and application name.
