### 2022-10-11 hexatomic/hexatomic#409 @reckart
If I move the application from the "Downloads" folder to the "Applications" folder, it starts up. I still get the "Couldn't find ProvisioningJob" dialog though.

TBH, I don't tend to read documentation before installing stuff. I just went to the [downloads page,](https://github.com/hexatomic/hexatomic/releases/tag/v1.0.1), downloaded the MacOS zip, extracted it and ran it.
