### 2022-10-11 hexatomic/hexatomic#409 @reckart
If you could provide the OS X version as a DMG with one of these nice "Please drag XXX to the Application folder" display, I think it would be a better user experience for OS X users.
