### 2022-10-11 hexatomic/hexatomic#409 @thomaskrause
> If I move the application from the "Downloads" folder to the "Applications" folder, it starts up. I still get the "Couldn't find ProvisioningJob" dialog though.
>
> TBH, I don't tend to read documentation before installing stuff. I just went to the [downloads page,](https://github.com/hexatomic/hexatomic/releases/tag/v1.0.1), downloaded the MacOS zip, extracted it and ran it.

Yes, that's totally understandable since the downloaded format (what am I downloading and how as I typically handle this kind of files) should be self-explaining. That's why when there is any deviation from the default way of doing things we should make it very clear at the location of the download (like the release notes), that something is different.

Also, we should definitely look into providing Hexatomic as DMG instead of a ZIP file. @sdruskat We should make a separate issue about that.

Thanks also for the report of the "Did you start Hexatomic out of the Eclipse IDE message". I get it to on MacOS after the first start where I had to confirm that I still trust the authors, too. This might be some sandbox-effect from a blocked network connection or an internal bug in the logic.  But the error message is not helpful in any way and too obstructive. I added an issue for that #410
