### 2022-10-11 6931ff9
Clarify Mac OS installation instructions in user documentation.
- No Unzip needed on Safari
- Command line run is confusing for normal users and additional information they need to parse, with unknown terms. Moved into a footnote
