### 2022-10-11 #4825 @reckart
I would recommend including a NOTICE file at the top-level of the repo or at least at the level of the bundles mentioning any embedded third-party-copyrighted stuff, e.g. https://github.com/hexatomic/hexatomic/blob/develop/bundles/org.corpus_tools.hexatomic.core/icons/fontawesome/LICENSE

And/or add the respective licenses and mentions to the top-level license file.

The license overview provided above did not include a mention of fontawesome - are other resource licenses missing as well?
