### 2022-10-11 hexatomic/hexatomic#414 @sdruskat
Add missing licenses to documentation
**Describe the bug**

https://github.com/openjournals/joss-reviews/issues/4825#issuecomment-1274393159 mentions that the license for fontawesome is missing from the documentation, and perhaps other resource licenses as well.

**To reproduce the bug**
Steps to reproduce the behavior:
1. `CITATION.cff` doesn't include a license for fontawesome
2. The top-level license file doesn't include the license either

**Expected behavior**

The licenses are clearly documented.

One solution for this would be to include a NOTICE/BOM file (or similar), or add the to the top-level license file.
