### 2022-10-11 hexatomic/hexatomic#416 @reckart
About dialog does not provide information on included OSS software
**Describe the bug**
The about dialog claims that the software is fully under Apache License 2.0 - however, as seen [here](https://github.com/openjournals/joss-reviews/issues/4825#issuecomment-1273098258), there are many third party libraries and resources under different licenses included.

**To reproduce the bug**
1. Click on `org.corpus_tools.hexatomic`-> About
2. See dialog

**Expected behavior**
A list of the third-party libraries / resources should be available with licenses named.

**Screenshots**

 - Operating System: MacOS 12.6
 - Hexatomic Version: 1.0.1
