### 2022-10-11 hexatomic/hexatomic#417 @reckart
Build fails with NoClassDefFound
**Describe the bug**
Build fails (see error below).

**To reproduce the bug**
Steps to reproduce the behavior:
1. Clone repo
2. `mvn clean install`
3. See error

**Expected behavior**
Build should complete successfully.

**Screenshots**
```
[INFO] --- tycho-surefire-plugin:2.7.3:test (default-test) @ org.corpus_tools.hexatomic.core.tests ---
[INFO] Could not find a java toolchain of type jdk, using java from JAVA_HOME instead (/Library/Java/JavaVirtualMachines/zulu-18.jdk/Contents/Home/bin/java)
[INFO] Executing Test Runtime with timeout 0, logs (if any) will be placed at: /.../git/hexatomic/tests/org.corpus_tools.hexatomic.core.tests/target/work/data/.metadata/.log
[INFO] Command line:
	[/Library/Java/JavaVirtualMachines/zulu-18.jdk/Contents/Home/bin/java, -Dosgi.noShutdown=false, -Dosgi.os=macosx, -Dosgi.ws=cocoa, -Dosgi.arch=aarch64, -Dosgi.clean=true, -jar, /.../.m2/repository/p2/osgi/bundle/org.eclipse.equinox.launcher/1.6.400.v20210924-0641/org.eclipse.equinox.launcher-1.6.400.v20210924-0641.jar, -data, /.../git/hexatomic/tests/org.corpus_tools.hexatomic.core.tests/target/work/data, -install, /.../git/hexatomic/tests/org.corpus_tools.hexatomic.core.tests/target/work, -configuration, /.../git/hexatomic/tests/org.corpus_tools.hexatomic.core.tests/target/work/configuration, -application, org.eclipse.tycho.surefire.osgibooter.headlesstest, -testproperties, /.../git/hexatomic/tests/org.corpus_tools.hexatomic.core.tests/target/surefire.properties]
Running org.eclipse.wb.swt.TestSwtResourceManager
Tests run: 2, Failures: 0, Errors: 2, Skipped: 0, Time elapsed: 0.016 s <<< FAILURE! - in org.eclipse.wb.swt.TestSwtResourceManager
testCacheAndDisposeColors  Time elapsed: 0.01 s  <<< ERROR!
java.lang.NoClassDefFoundError: org/eclipse/swt/graphics/Device
	at org.eclipse.wb.swt.TestSwtResourceManager.testCacheAndDisposeColors(TestSwtResourceManager.java:69)
Caused by: java.lang.ClassNotFoundException: org.eclipse.swt.graphics.Device cannot be found by org.corpus_tools.hexatomic.core_1.1.0.202210110942
	at org.eclipse.wb.swt.TestSwtResourceManager.testCacheAndDisposeColors(TestSwtResourceManager.java:69)

testCacheAndDisposeFonts  Time elapsed: 0.001 s  <<< ERROR!
java.lang.NoClassDefFoundError: org/eclipse/swt/graphics/Device
	at org.eclipse.wb.swt.TestSwtResourceManager.testCacheAndDisposeFonts(TestSwtResourceManager.java:19)
Caused by: java.lang.ClassNotFoundException: org.eclipse.swt.graphics.Device cannot be found by org.corpus_tools.hexatomic.core_1.1.0.202210110942
	at org.eclipse.wb.swt.TestSwtResourceManager.testCacheAndDisposeFonts(TestSwtResourceManager.java:19)

Running org.eclipse.wb.swt.TestResourceManager
Tests run: 1, Failures: 0, Errors: 1, Skipped: 0, Time elapsed: 0.001 s <<< FAILURE! - in org.eclipse.wb.swt.TestResourceManager
testCacheAndDisposeImages  Time elapsed: 0 s  <<< ERROR!
java.lang.NoClassDefFoundError: org/eclipse/swt/graphics/Device
	at org.eclipse.wb.swt.TestResourceManager.testCacheAndDisposeImages(TestResourceManager.java:18)
Caused by: java.lang.ClassNotFoundException: org.eclipse.swt.graphics.Device cannot be found by org.corpus_tools.hexatomic.core_1.1.0.202210110942
	at org.eclipse.wb.swt.TestResourceManager.testCacheAndDisposeImages(TestResourceManager.java:18)

Running org.corpus_tools.hexatomic.core.handler.TestSaveAsHandler
Tests run: 1, Failures: 0, Errors: 1, Skipped: 0, Time elapsed: 0 s <<< FAILURE! - in org.corpus_tools.hexatomic.core.handler.TestSaveAsHandler
org.corpus_tools.hexatomic.core.handler.TestSaveAsHandler  Time elapsed: 0 s  <<< ERROR!
java.lang.NoClassDefFoundError: org/eclipse/swt/widgets/Shell
Caused by: java.lang.ClassNotFoundException: org.eclipse.swt.widgets.Shell cannot be found by org.corpus_tools.hexatomic.core_1.1.0.202210110942

Running org.corpus_tools.hexatomic.core.TestSaltNotifications
SLF4J: Failed to load class "org.slf4j.impl.StaticLoggerBinder".
SLF4J: Defaulting to no-operation (NOP) logger implementation
SLF4J: See http://www.slf4j.org/codes.html#StaticLoggerBinder for further details.
Tests run: 3, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.361 s - in org.corpus_tools.hexatomic.core.TestSaltNotifications
testCreateAllAnnotationTypesFromFactory  Time elapsed: 0.323 s
testExampleGraph  Time elapsed: 0.031 s
testRemoveAll  Time elapsed: 0.006 s
Running org.corpus_tools.hexatomic.core.SaltHelperTest
Tests run: 2, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.003 s - in org.corpus_tools.hexatomic.core.SaltHelperTest
testResolveDelegation  Time elapsed: 0.001 s
testGetGraphForObject  Time elapsed: 0.002 s
Running org.corpus_tools.hexatomic.core.TestProjectManager
Tests run: 6, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.096 s - in org.corpus_tools.hexatomic.core.TestProjectManager
testUndoDeletion  Time elapsed: 0.045 s
testEventOnOpen  Time elapsed: 0.003 s
testDocumentCleanup  Time elapsed: 0.024 s
testNoRedoAfterManualCheckpoint  Time elapsed: 0.006 s
testUndoRedo  Time elapsed: 0.005 s
testOpenInvalid  Time elapsed: 0.002 s

Results:

Errors:
  TestSaveAsHandler » NoClassDefFound org/eclipse/swt/widgets/Shell
  TestResourceManager.testCacheAndDisposeImages:18 NoClassDefFound org/eclipse/s...
  TestSwtResourceManager.testCacheAndDisposeColors:69 NoClassDefFound org/eclips...
  TestSwtResourceManager.testCacheAndDisposeFonts:19 NoClassDefFound org/eclipse...

Tests run: 15, Failures: 0, Errors: 4, Skipped: 0

[INFO] ------------------------------------------------------------------------
[INFO] Reactor Summary for org.corpus_tools.hexatomic.root 1.1.0-SNAPSHOT:
[INFO]
[INFO] org.corpus_tools.hexatomic.root .................... SUCCESS [  5.707 s]
[INFO] org.corpus_tools.hexatomic.bundles ................. SUCCESS [  5.427 s]
[INFO] [bundle] Hexatomic Core Plugin ..................... SUCCESS [ 13.477 s]
[INFO] [bundle] Corpus Structure Editor ................... SUCCESS [  1.927 s]
[INFO] [bundle] org.corpus_tools.hexatomic.textviewer ..... SUCCESS [  1.471 s]
[INFO] org.corpus_tools.hexatomic.graph ................... SUCCESS [  5.332 s]
[INFO] [bundle] Hexatomic Grid Editor ..................... SUCCESS [  2.561 s]
[INFO] [bundle] Corpus file formats ....................... SUCCESS [  1.861 s]
[INFO] org.corpus_tools.hexatomic.features ................ SUCCESS [  0.052 s]
[INFO] org.corpus_tools.hexatomic ......................... SUCCESS [  0.233 s]
[INFO] org.corpus_tools.hexatomic.releng .................. SUCCESS [  0.063 s]
[INFO] org.corpus_tools.hexatomic.update .................. SUCCESS [  1.070 s]
[INFO] org.corpus_tools.hexatomic.product ................. SUCCESS [01:01 min]
[INFO] org.corpus_tools.hexatomic.target .................. SUCCESS [  0.057 s]
[INFO] org.corpus_tools.hexatomic.tests ................... SUCCESS [  0.050 s]
[INFO] [test-bundle] Hexatomic Core Tests ................. FAILURE [  3.417 s]
[INFO] [test-bundle] Graph Tests .......................... SKIPPED
[INFO] [test-bundle] Grid Editor Tests .................... SKIPPED
[INFO] [test-bundle] Importer and Exporter Tests .......... SKIPPED
[INFO] org.corpus_tools.hexatomic.it.tests ................ SKIPPED
[INFO] Aggregate Test Report .............................. SKIPPED
```

 - Operating System: MacOS 12.6 (arm)
 - Hexatomic Version: commit 7baf259cb1931044e2d0a9d39eea4c5e42b37bb3
 - Java Version: OpenJDK 64-Bit Server VM Zulu18.30+11-CA (build 18.0.1+10, mixed mode, sharing)
