### 2022-10-11 bf7dc11
Show more explanations when update check fails (#410)
This also limits the information a user needs to see.
For background update checks, only a status message is shown.
For manual checks, a dialog (with error log!) is shown and has the information about running from the Eclipse IDE only if this is actually run using the provided launch configuration.

This is result to limit the effects when there is some unknown connection problems (possible from a sandbox or other security mechanism #409).
