### 2022-10-11 2b83dea
Remove debug code in if-statement and fix code style
