### 2022-10-11 082c741
Update check on startup was accidentally disabled because there is no shell available for auto-inject
