+++ bundles/org.corpus_tools.hexatomic.timeline/src/main/java/org/corpus_tools/hexatomic/timeline/TimelineTokenAligner.java
@@ -27,7 +27,7 @@ import org.eclipse.swt.SWT;
 import org.eclipse.swt.layout.GridLayout;
 import org.eclipse.swt.widgets.Composite;
 
-public class TimelineAlignEditor {
+public class TimelineTokenAligner {
 
   @Inject
   private ProjectManager projectManager;
