### 2022-10-13 797e3cd
Refresh table when the graph was changed
