### 2022-10-13 3b6a783
Allow to add new token to the end of the textual data
