### 2022-10-14 458a023
Fix code smell
