### 2022-10-18 03b591e
Add diplay converter so we can directly return SToken objects
