### 2022-10-18 d409f00
Define when to span grid cells
