### 2022-10-18 9310598
Index token position in timeline when the graph changes, so scrolling becomes much faster
