+++ bundles/org.corpus_tools.hexatomic.timeline/src/main/java/org/corpus_tools/hexatomic/timeline/internal/data/TimelineTokenDataProvider.java
@@ -1,9 +1,35 @@
+/*-
+ * #%L
+ * [bundle] Timeline Editor
+ * %%
+ * Copyright (C) 2018 - 2022 Stephan Druskat, Thomas Krause
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+
 package org.corpus_tools.hexatomic.timeline.internal.data;
 
-import java.util.List;
+import java.util.LinkedHashMap;
+import java.util.Map;
 import java.util.Optional;
-import java.util.stream.Collectors;
+import java.util.TreeMap;
+import java.util.stream.Stream;
+import javax.inject.Inject;
+import org.corpus_tools.hexatomic.core.ProjectManager;
 import org.corpus_tools.hexatomic.core.SaltHelper;
+import org.corpus_tools.hexatomic.core.Topics;
+import org.corpus_tools.hexatomic.core.undo.ChangeSet;
 import org.corpus_tools.salt.SaltFactory;
 import org.corpus_tools.salt.common.SDocumentGraph;
 import org.corpus_tools.salt.common.STextualDS;
@@ -11,12 +37,19 @@ import org.corpus_tools.salt.common.STextualRelation;
 import org.corpus_tools.salt.common.STimelineRelation;
 import org.corpus_tools.salt.common.SToken;
 import org.eclipse.e4.core.di.annotations.Creatable;
+import org.eclipse.e4.ui.di.UIEventTopic;
 import org.eclipse.nebula.widgets.nattable.data.IDataProvider;
 
 @Creatable
 public class TimelineTokenDataProvider implements IDataProvider {
 
   private SDocumentGraph graph;
+
+  @Inject
+  private ProjectManager projectManager;
+
+  private final Map<STextualDS, TreeMap<Integer, SToken>> tokenByPosition =
+      new LinkedHashMap<STextualDS, TreeMap<Integer, SToken>>();
 
   public void setGraph(SDocumentGraph graph) {
     this.graph = graph;
@@ -27,6 +60,8 @@ public class TimelineTokenDataProvider implements IDataProvider {
     if (this.graph.getTimeline().getData() == null) {
       this.graph.getTimeline().increasePointOfTime();
     }
+
+    updateTokenPosition();
   }
 
   @Override
@@ -41,22 +76,60 @@ public class TimelineTokenDataProvider implements IDataProvider {
   }
 
   /** Find the tokens that are connected to TLI belonging to a data source */
-  private List<SToken> getTokenForTli(int columnIndex, int rowIndex) {
-    // Find the tokens that is connected to this TLI
-    List<SToken> token = graph.getTimelineRelations().stream()
-        .filter(rel -> rel.getStart() <= rowIndex && rel.getEnd() > rowIndex)
-        .map(rel -> rel.getSource()).collect(Collectors.toList());
-    // Reduce tokens to the ones belonging to the selected data source
-    STextualDS ds = graph.getTextualDSs().get(columnIndex);
-    return token.stream().flatMap(t -> t.getOutRelations().stream()).map(rel -> {
-          if (rel instanceof STextualRelation) {
-            STextualRelation textRel = (STextualRelation) rel;
-            if (textRel.getTarget() == ds) {
-              return textRel.getSource();
-            }
[…]
