### 2022-10-18 43c38c1
Perform index creation in background job
