### 2022-10-30 #4825 @reckart
The website structure of https://hexatomic.github.io is confusing. In the left sidebar, it says that is has documentation but then is only about sustainability and tooling, no actual tool documentation. There is a link to the actual documentation on the bottom of the page, but it is not easy to find - the user may have to scroll down to be able to see it. It would be better if the documentation was hosted properly at https://hexatomic.github.io - the information about how the project was planned and conducted could be integrated into the documentation as a subsection or appendix.
