### 2022-10-30 #4825 @reckart
When I press "new salt project", nothing happens.
