### 2022-10-30 #4825 @reckart
After importing a folder with text documents, clicking on a document does nothing. Double-clicking starts renaming of the document. I would expect that either single-click or double-click would open the document for editing.

Instead, one has to right-click on the document to choose some editor.
