### 2022-10-30 #4825 @reckart
There seems to be no way to edit the graph in the graph editor. I have opened a document and the only thing I seem to be able to do with the mouse is to drag tokens around.
