### 2022-10-30 hexatomic/hexatomic#423 @reckart
No merging in grid view?
**Describe the bug**
I don't see the option for merging cells in the grid view when selecting two adjacent cells and right-clicking.

**To reproduce the bug**
Steps to reproduce the behavior:
1. Import text
2. Open grid view
3. Add new column
4. select two cells in the new column
5. right-click

**Expected behavior**
According to the documentation, there should be an option to merge cells.

**Screenshots**
![Screenshot 2022-10-30 at 12 20 09](https://user-images.githubusercontent.com/1410238/198875804-c6b27944-0650-4869-b862-c48498b29047.png)

 - Operating System: OS X 12.6
 - Hexatomic Version: 1.0.1
 - Java Version: Java 18
