### 2022-10-30 #4825 @reckart
> Do the authors clearly state what problems the software is designed to solve and who the target audience is?

The [documentation](https://hexatomic.github.io/hexatomic/user/v1.0/index.html) has a short description of what it aims to do. However, IMHO Hexatomic has a very specific type of user interaction via commands and a quite specific way of treating documents and annotations that likely has a very specific target audience which may like Hexatomic better than the other many annotation tools out there. From the documentation introduction, it is not clear though which are the main selling points of Hexatomic and when/why a user should prefer it over other tools.
