### 2022-10-30 #4825 @reckart
> Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

There are no readily usable examples either coming with the application or being easily downloadable and importable from the website.

The authors state:

> There is a brief ["how to create a corpus from scratch"](https://hexatomic.github.io/hexatomic/user/v1.0/howtos/corpus-creation.html) in the [user documentation](https://hexatomic.github.io/hexatomic/user/v1.0/).

Having a few importable examples in the different supported formats and highlighting different aspects of how the tool can be used would be more helpful
