### 2022-10-30 #4825 @reckart
> Installation: Does installation proceed as outlined in the documentation?

The installation works as described in the documentation, but at least on OS X, it does not work as users (or at least I) would normally expect. https://github.com/hexatomic/hexatomic/issues/409
