### 2022-10-30 #4825 @reckart
>  Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

I did not notice any performance claims.
