### 2022-10-30 #4825 @reckart
>  Functionality: Have the functional claims of the software been confirmed?

* The documentation of the grid editors says that merging (and splitting) cells would be possible, but I could not confirm this: https://github.com/hexatomic/hexatomic/issues/423
* The paper states that support for Git is integrated, but I did not find any documentation about this. It appears the authors's claim extends only to being able to manage the data folders using an external git which has no integration at all with the editor (i.e. no history view, no conflict resolution, no way to view old versions in the editors, etc.). However, when I read the claims about git support in the paper, I would have expected a more integrated support. After all, it is possible with many tools to simply manage their data folders with an external git. I believe one could for example do that with brat and thereby also support the "field linguist" use-case mentioned in the paper. One can also use INCEpTION in a standalone mode without internet connection and there is also an experimental [git integration](https://inception-project.github.io/releases/25.2/docs/user-guide.html#sect_versioning) - also without history browsing or conflict resolution, but with integration into the project settings.
