### 2022-10-30 #4825 @reckart
>  Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?

The software paper's introduction describes linguists as the target group and highlights a few particular types of annotations which the tool is intended for. This section could be transferred to the documentation which does currently not explain well who should be using this tool and why.
