### 2022-10-30 #4825 @reckart
>  A statement of need: Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?

The section starts by highlighting flexible conversion capabilities and the ability to integrate data from different sources. It appears to me though that this is a feature of earlier work (Salt'n'Pepper) rather than of Hexatomic itself. Of course Hexatomic uses this earlier work as a software library.

The section continues saying that comparable tools are often run on centralized servers which would induce additional complications. However, the section fails to mention that the other cited tools (WebAnno, INCEpTION, GATE) **can** also be run as local applications and be used without internet access. Hexatomic, however, cannot be used via a server. So while other tools support both modes of operation (local and server), Hexatomic only supports local deployment.

This inablilty of a server-based installation is addressed by arguing that git can be used to synchronize data and to collaborate. See https://github.com/openjournals/joss-reviews/issues/4825#issuecomment-1296229119 regarding the integration of git with Hexatomic and other tools.

Finally the section mentions future work to be done in the form for specialized editors for text alignment and support for audio/video editing. There is no mention how these functionalities would compare with other tools like Exmaralda. Also, the features are not yet available, so they could not be inspected. I think future work should be discussed in a "future work" section and not in a "statement of need" section.
