### 2022-10-30 #4825 @reckart
> State of the field: Do the authors describe how this software compares to other commonly-used packages?

There is no such section in the paper. There is a paragraph about comparable packages embedded into the "statement of need" section, but it is rather superficial (see https://github.com/openjournals/joss-reviews/issues/4825#issuecomment-1296233690).
