### 2022-10-30 #4825 @reckart
> Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

There is no dedicated "State of the field" section. Also, a "future work" section has been embedded into the "statement of need" section.

In terms of language and readability, the paper looks good.
