### 2022-10-30 #4825 @reckart
These items remain unchecked on my list:

**General checks**

- [ ]  Reproducibility: If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item. -- I cannot build the repository. https://github.com/hexatomic/hexatomic/issues/417

**Functionality**
A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is? -- https://github.com/hexatomic/hexatomic/issues/423 , no documentation on git integration claimed in paper

**Documentation**
- [ ] A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is? -- statement of need from paper should be transferred to documentation
- [ ] Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems). -- Easily downloadable examples in different formats and for different use-cases should be provided

**Software paper**

- [ ] A statement of need: Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work? -- see https://github.com/openjournals/joss-reviews/issues/4825#issuecomment-1296233690
- [ ] State of the field: Do the authors describe how this software compares to other commonly-used packages? -- Comparison is very superficial and not in a dedicated section.
- [ ] Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)? -- Sections for "State of the field" and "Future Work" should be extracted. Otherwise ok.
