### 2022-11-25 hexatomic/hexatomic#416 @thomaskrause
We are currently distributing the license information in the https://github.com/hexatomic/hexatomic/blob/develop/CITATION.cff citation file and the corresponding `THIRD-PARTY` folder in case the license requires to ship the license text.

The simplest implementation would probably be to add a link to the tagged `CITATION.cff` and the folder in GitHub to the About dialog (because dependencies might be different between the develop branch and the used version of Hexatomic).

I added a separate issue to add the non P2 dependencies to the CITATION.cff template manually ( #425)
