### 2022-11-25 hexatomic/hexatomic#417 @thomaskrause
Can you check that the build works with disabled tests?

```
mvn -DskipTests clean install
```
From the error message, it looks like there are issues with the user interface integration tests. There were known issues on macOS with the test, which we started to address in a separate PR (#426). I will get back when the tests have been fixed on macOS.
