### 2022-11-25 hexatomic/hexatomic#409 @thomaskrause
I added the DMG feature as a new issue #427

As far as I understand, with the fixes about the update notification and the installation by moving the application to the `Applications` folder (and documenting this more clearly) this issue is resolved. Please re-open if you disagree.
