+++ tests/org.corpus_tools.hexatomic.it.tests/src/main/java/org/corpus_tools/hexatomic/it/tests/TestHelper.java
@@ -107,6 +107,7 @@ public class TestHelper {
     } else {
       return "Shift+";
     }
+  }
 
   /**
    * Recursively delete a directory.
