### 2022-11-25 22707e4
Use Apple System font name in test
