+++ tests/org.corpus_tools.hexatomic.core.tests/src/main/java/org/eclipse/wb/swt/TestSwtResourceManager.java
@@ -27,7 +27,7 @@ class TestSwtResourceManager {
     assertEquals(SWT.NONE, font1.getFontData()[0].getStyle());
     assertEquals(1, SwtResourceManager.fontMap.size());
 
-    
+
     // Add second font (a bold version of the same font)
     Font font2 = SwtResourceManager.getFont(getDefaultFontName(), 12, SWT.BOLD);
     assertNotNull(font2);
@@ -35,7 +35,7 @@ class TestSwtResourceManager {
     assertEquals(12, font2.getFontData()[0].getHeight());
     assertEquals(SWT.BOLD, font2.getFontData()[0].getStyle());
     assertEquals(2, SwtResourceManager.fontMap.size());
-    
+
     // Add third font (a larger version of the same font)
     Font font3 = SwtResourceManager.getFont(getDefaultFontName(), 18, SWT.NONE);
     assertNotNull(font3);
@@ -48,7 +48,7 @@ class TestSwtResourceManager {
     Font cachedFont = SwtResourceManager.getFont(getDefaultFontName(), 12, SWT.BOLD);
     assertEquals(font2, cachedFont);
     assertEquals(3, SwtResourceManager.fontMap.size());
-    
+
     // Add a bold version of the font using the helper method
     assertEquals(0, SwtResourceManager.fontToBoldFontMap.size());
     Font boldFont = SwtResourceManager.getBoldFont(font1);
@@ -61,7 +61,7 @@ class TestSwtResourceManager {
     assertEquals(0, SwtResourceManager.fontMap.size());
     assertEquals(0, SwtResourceManager.fontToBoldFontMap.size());
   }
-  
+
   /**
    * Tests correct caching and disposal of colors.
    */
@@ -76,7 +76,7 @@ class TestSwtResourceManager {
     assertEquals(50, color1.getGreen());
     assertEquals(255, color1.getBlue());
     assertEquals(1, SwtResourceManager.colorMap.size());
-    
+
     // Add second color
     Color color2 = SwtResourceManager.getColor(new RGB(100, 128, 0));
     assertNotNull(color2);
@@ -84,25 +84,25 @@ class TestSwtResourceManager {
     assertEquals(128, color2.getGreen());
     assertEquals(0, color2.getBlue());
     assertEquals(2, SwtResourceManager.colorMap.size());
-    
+
     // Get same color again
     Color cachedColor = SwtResourceManager.getColor(new RGB(128, 50, 255));
     assertEquals(color1, cachedColor);
     assertEquals(2, SwtResourceManager.colorMap.size());
-        
+
 
     // Dispose color resources and check that the map is empty
     SwtResourceManager.dispose();
     assertEquals(0, SwtResourceManager.colorMap.size());
   }
 
-  
+
   private String getDefaultFontName() {
-    if(SystemUtils.IS_OS_MAC_OSX) {
+    if (SystemUtils.IS_OS_MAC_OSX) {
       return ".AppleSystemUIFont";
     } else {
       return "Sans";
     }
   }
-  
+
 }
