### 2022-11-28 hexatomic/hexatomic#423 @sdruskat
Hi @reckart, thanks for opening this issue, and sorry for the late reply.

The [user documentation notes in the section on merging spans](https://hexatomic.github.io/hexatomic/user/v1.0/usage/grid-editor/editing-values.html#merging-spans), that

> You can merge spans that are in the same annotation column and **have the same annotation value**: ...

This is why your selection across two _different_ annotation values does not display the option to merge spans.

In this basic implementation of merging spans, no option to pick one/more/all of many annotation values across selected spans is given. We may implement this if users request it.

Closing this now, but please feel free to continue the discussion, or re-open if you don't agree that this is not a bug.
