### 2022-12-02 745823c
Fix graph editor tests on macOS
