### 2023-01-06 e8dbcbd
Update AddColumnCommandHandler.java
Even if the user aborts the "Add Column" handler, the command has been handled.

Otherwise, the handler might be called multiple times.
