### 2023-01-09 6e20e73
Click on the context menu in tests to ensure they are closed
