+++ tests/org.corpus_tools.hexatomic.it.tests/src/main/java/org/corpus_tools/hexatomic/it/tests/TestGridEditor.java
@@ -72,10 +72,13 @@ import org.eclipse.swtbot.swt.finder.widgets.SWTBotCombo;
 import org.eclipse.swtbot.swt.finder.widgets.SWTBotMenu;
 import org.eclipse.swtbot.swt.finder.widgets.SWTBotRootMenu;
 import org.eclipse.swtbot.swt.finder.widgets.SWTBotShell;
+import org.eclipse.swtbot.swt.finder.widgets.SWTBotToolbarButton;
 import org.eclipse.swtbot.swt.finder.widgets.SWTBotTreeItem;
 import org.eclipse.swtbot.swt.finder.widgets.TimeoutException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.parallel.Execution;
+import org.junit.jupiter.api.parallel.ExecutionMode;
 
 /**
  * Integration tests for the grid editor.
@@ -83,6 +86,7 @@ import org.junit.jupiter.api.Test;
  * @author Stephan Druskat (mail@sdruskat.net)
  *
  */
+@Execution(ExecutionMode.SAME_THREAD)
 public class TestGridEditor {
 
 
@@ -135,7 +139,7 @@ public class TestGridEditor {
 
 
 
-  private SWTWorkbenchBot bot = new SWTWorkbenchBot(TestHelper.getEclipseContext());
+  private final SWTWorkbenchBot bot = new SWTWorkbenchBot(TestHelper.getEclipseContext());
 
   private URI exampleProjectUri;
   private URI overlappingExampleProjectUri;
@@ -190,11 +194,16 @@ public class TestGridEditor {
     twoDsExampleProjectUri = URI.createFileURI(twoDsExampleProjectDirectory.getAbsolutePath());
     scrollingExampleProjectUri =
         URI.createFileURI(scrollingExampleProjectDirectory.getAbsolutePath());
+
   }
 
   SWTBotView openEditorForDefaultDocument() {
+
+    SWTBotView corpusStructurePart = bot.partByTitle("Corpus Structure");
+
     // Select the first example document
-    SWTBotTreeItem docMenu = bot.tree().expandNode("corpusGraph1").expandNode("rootCorpus")
+    SWTBotTreeItem docMenu = corpusStructurePart.bot().tree().expandNode("corpusGraph1")
+        .expandNode("rootCorpus")
         .expandNode("subCorpus1").expandNode("doc2");
 
     // select and open the editor
@@ -230,10 +239,12 @@ public class TestGridEditor {
 
   private SWTBotView openEditorOnSameDocument() {
     // Activate corpus structure editor
-    bot.partById("org.corpus_tools.hexatomic.corpusedit.part.corpusstructure").show();
+    SWTBotView corpusStructurePart = bot.partByTitle("Corpus Structure");
+    corpusStructurePart.show();
 
 
-    SWTBotTreeItem docMenu = bot.tree().expandNode("corpusGraph1").expandNode("rootCorpus")
+    SWTBotTreeItem docMenu = corpusStructurePart.bot().tree().expandNode("corpusGraph1")
+        .expandNode("rootCorpus")
         .expandNode("subCorpus1").expandNode("doc2");
 
     // select and open the editor
@@ -256,9 +267,11 @@ public class TestGridEditor {
   SWTBotView openOverlapExample() {
     // Programmatically open the example corpus
     openExample(overlappingExampleProjectUri);
+    SWTBotView corpusStructurePart = bot.partByTitle("Corpus Structure");
     // Select the first example document
     SWTBotTreeItem docMenu =
-        bot.tree().expandNode("corpus-graph").expandNode(CORPUS).expandNode(DOC);
[…]
