### 2023-01-14 cc6b9cf
Remove unused variable
