### 2023-01-14 d90ad6a
Downgrade to 2022-06 platform
