### 2023-01-19 8264fb3
Use current display to get the active shell
