### 2023-01-19 ddb1289
Add explicit click on context menu entry to more grid tests
