### 2023-01-19 87170a4
Set text to avoid cutting of the string in half
