### 2023-02-19 2305e99
Add temporary grid test file
This has the same tests but a lot of them are commented out. There are several tests that only fail because a previous test did not clean up its resources and this way it is easer to debug this.
