### 2023-02-20 bcc13a4
More context menu cleanup
