+++ tests/org.corpus_tools.hexatomic.it.tests/src/main/java/org/corpus_tools/hexatomic/it/tests/TestGridEditorNew.java
@@ -12,6 +12,7 @@ import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.concurrent.Callable;
 import java.util.concurrent.ExecutionException;
@@ -58,6 +59,7 @@ import org.eclipse.swtbot.nebula.nattable.finder.widgets.SWTBotNatTable;
 import org.eclipse.swtbot.swt.finder.finders.UIThreadRunnable;
 import org.eclipse.swtbot.swt.finder.keyboard.Keyboard;
 import org.eclipse.swtbot.swt.finder.keyboard.KeyboardFactory;
+import org.eclipse.swtbot.swt.finder.keyboard.Keystrokes;
 import org.eclipse.swtbot.swt.finder.waits.Conditions;
 import org.eclipse.swtbot.swt.finder.waits.DefaultCondition;
 import org.eclipse.swtbot.swt.finder.widgets.SWTBotCombo;
@@ -624,64 +626,64 @@ public class TestGridEditorNew {
     assertTrue(table.hasConfigLabel(1, 5, StyleConfiguration.EMPTY_CELL_STYLE));
   }
 
-  // /**
-  // * Tests editing cells by activating the cell editor with a double-click on the cell.
-  // */
-  // @Test
-  // void testEditCellOnDoubleClick() {
-  // openDefaultExample();
-  //
-  // SWTNatTableBot tableBot = new SWTNatTableBot();
-  // SWTBotNatTable table = tableBot.nattable();
-  //
-  // table.doubleclick(2, 2);
-  // typeTextPressReturn(table);
-  //
-  // Object nodeObj = table.widget.getDataValueByPosition(2, 2);
-  // assertTrue(nodeObj instanceof SNode);
-  // SNode node = (SNode) nodeObj;
-  // assertEquals(TEST_ANNOTATION_VALUE,
-  // node.getAnnotation(table.getCellDataValueByPosition(0, 2)).getValue());
-  // }
-  //
-  // /**
-  // * Tests editing cells by activating the cell editor by just typing away.
-  // */
-  // @Test
-  // void testEditCellTyping() {
-  // openDefaultExample();
-  //
-  // SWTNatTableBot tableBot = new SWTNatTableBot();
-  // SWTBotNatTable table = tableBot.nattable();
-  //
-  // table.click(2, 2);
-  // typeTextPressReturn(table);
-  // Object nodeObj = table.widget.getDataValueByPosition(2, 2);
-  // assertTrue(nodeObj instanceof SNode);
-  // SNode node = (SNode) nodeObj;
-  // assertEquals(TEST_ANNOTATION_VALUE,
-  // node.getAnnotation(table.getCellDataValueByPosition(0, 2)).getValue());
-  // }
-  //
-  // /**
-  // * Tests editing cells by activating the cell editor by just typing away.
-  // */
-  // @Test
-  // void testEditCellSpaceActivatedTyping() {
-  // openDefaultExample();
-  //
-  // SWTNatTableBot tableBot = new SWTNatTableBot();
-  // SWTBotNatTable table = tableBot.nattable();
-  //
-  // table.click(2, 2);
-  // keyboard.pressShortcut(Keystrokes.SPACE);
-  // typeTextPressReturn(table);
-  // Object nodeObj = table.widget.getDataValueByPosition(2, 2);
-  // assertTrue(nodeObj instanceof SNode);
-  // SNode node = (SNode) nodeObj;
-  // assertEquals(TEST_ANNOTATION_VALUE,
-  // node.getAnnotation(table.getCellDataValueByPosition(0, 2)).getValue());
-  // }
+  /**
[…]
