### 2023-02-20 1f5ea65
Use "test" instead of "TEST" as value
There are just too many issues with mapping upper case characters when typing with a SWTbot keyboard under different systems.
