### 2023-02-20 df71578
Use generic shortcut generator helper function
