+++ tests/org.corpus_tools.hexatomic.it.tests/src/main/java/org/corpus_tools/hexatomic/it/tests/TestGridEditorNew.java
@@ -2,6 +2,7 @@ package org.corpus_tools.hexatomic.it.tests;
 
 import static org.eclipse.swtbot.swt.finder.matchers.WidgetMatcherFactory.widgetOfType;
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -684,40 +685,40 @@ public class TestGridEditorNew {
     assertEquals(TEST_ANNOTATION_VALUE,
         node.getAnnotation(table.getCellDataValueByPosition(0, 2)).getValue());
   }
-  //
-  // @Test
-  // void testCreateAnnotationOnEmptySpanCell() {
-  // openOverlapExample();
-  //
-  // SWTNatTableBot tableBot = new SWTNatTableBot();
-  // SWTBotNatTable table = tableBot.nattable();
-  //
-  // assertEquals("", table.getCellDataValueByPosition(1, 5));
-  // table.click(1, 5);
-  // typeTextPressReturn(table);
-  // Object nodeObj = table.widget.getDataValueByPosition(5, 1);
-  // assertTrue(nodeObj instanceof SNode);
-  // SNode node = (SNode) nodeObj;
-  // assertEquals(TEST_ANNOTATION_VALUE,
-  // node.getAnnotation(table.getCellDataValueByPosition(0, 5)).getValue());
-  // }
-  //
-  // @Test
-  // void testCreateAnnotationOnEmptyTokenCell() {
-  // openOverlapExample();
-  //
-  // SWTNatTableBot tableBot = new SWTNatTableBot();
-  // SWTBotNatTable table = tableBot.nattable();
-  //
-  // assertEquals("", table.getCellDataValueByPosition(2, 2));
-  // table.click(2, 2);
-  // typeTextPressReturn(table);
-  // Object nodeObj = table.widget.getDataValueByPosition(2, 2);
-  // assertTrue(nodeObj instanceof SNode);
-  // SNode node = (SNode) nodeObj;
-  // assertEquals(TEST_ANNOTATION_VALUE,
-  // node.getAnnotation(table.getCellDataValueByPosition(0, 2)).getValue());
-  // }
+
+  @Test
+  void testCreateAnnotationOnEmptySpanCell() {
+    openOverlapExample();
+
+    SWTNatTableBot tableBot = new SWTNatTableBot();
+    SWTBotNatTable table = tableBot.nattable();
+
+    assertEquals("", table.getCellDataValueByPosition(1, 5));
+    table.click(1, 5);
+    typeTextPressReturn(table);
+    Object nodeObj = table.widget.getDataValueByPosition(5, 1);
+    assertTrue(nodeObj instanceof SNode);
+    SNode node = (SNode) nodeObj;
+    assertEquals(TEST_ANNOTATION_VALUE,
+        node.getAnnotation(table.getCellDataValueByPosition(0, 5)).getValue());
+  }
+
+  @Test
+  void testCreateAnnotationOnEmptyTokenCell() {
+    openOverlapExample();
+
+    SWTNatTableBot tableBot = new SWTNatTableBot();
+    SWTBotNatTable table = tableBot.nattable();
+
+    assertEquals("", table.getCellDataValueByPosition(2, 2));
+    table.click(2, 2);
+    typeTextPressReturn(table);
+    Object nodeObj = table.widget.getDataValueByPosition(2, 2);
[…]
