### 2023-02-20 9071a1c
Enabling more tests
