+++ tests/org.corpus_tools.hexatomic.it.tests/src/main/java/org/corpus_tools/hexatomic/it/tests/TestGridEditorNew.java
@@ -3,6 +3,7 @@ package org.corpus_tools.hexatomic.it.tests;
 import static org.eclipse.swtbot.swt.finder.matchers.WidgetMatcherFactory.widgetOfType;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -52,6 +53,7 @@ import org.eclipse.swt.widgets.Composite;
 import org.eclipse.swt.widgets.Control;
 import org.eclipse.swt.widgets.Display;
 import org.eclipse.swt.widgets.Label;
+import org.eclipse.swt.widgets.Shell;
 import org.eclipse.swt.widgets.Text;
 import org.eclipse.swtbot.e4.finder.widgets.SWTBotView;
 import org.eclipse.swtbot.e4.finder.widgets.SWTWorkbenchBot;
@@ -1136,294 +1138,284 @@ public class TestGridEditorNew {
         token.getAnnotation(SaltUtil.SALT_NAMESPACE, TEST_ANNOTATION_VALUE).getValue());
   }
 
-  // /**
-  // * Tests that when a single cell is selected and the annotation rename dialog is cancelled, that
-  // * everything has stayed the same.
-  // *
-  // * @throws Exception if the thread extracting the information from a given dialog is
-  // interrupted,
-  // * or if there is an error during execution. Can be either {@link InterruptedException} or
-  // * {@link ExecutionException}.
-  // */
-  // @Test
-  // void testCancelChangeAnnotationNameSingleCell() throws InterruptedException, ExecutionException
-  // {
-  // openDefaultExample();
-  //
-  // SWTNatTableBot tableBot = new SWTNatTableBot();
-  // SWTBotNatTable table = tableBot.nattable();
-  //
-  // assertTrue(table.widget.getDataValueByPosition(2, 3) instanceof SToken);
-  // SToken token = (SToken) table.widget.getDataValueByPosition(2, 3);
-  // assertEquals(EXAMPLE_VALUE,
-  // token.getAnnotation(SaltUtil.SALT_NAMESPACE, LEMMA_NAME).getValue());
-  // // Select and change name
-  // table.click(3, 2);
-  // table.contextMenu(3,
-  // 2).contextMenu(GridEditor.CHANGE_ANNOTATION_NAME_POPUP_MENU_LABEL).click();
-  // SWTBotShell dialog = tableBot.shell(RENAME_DIALOG_TITLE);
-  // assertNotNull(dialog);
-  // // Check that the fields are pre-filled
-  // assertDialogTexts(dialog, SaltUtil.SALT_NAMESPACE + SaltUtil.NAMESPACE_SEPERATOR + LEMMA_NAME);
-  // keyboard.typeText(TEST_ANNOTATION_VALUE);
-  // tableBot.button(CANCEL).click();
-  // bot.waitUntil(Conditions.shellCloses(dialog));
-  // // Assert names and positions have not changed
-  // assertEquals(token, table.widget.getDataValueByPosition(2, 3));
-  // token = (SToken) table.widget.getDataValueByPosition(2, 3);
-  // assertEquals(EXAMPLE_VALUE,
-  // token.getAnnotation(SaltUtil.SALT_NAMESPACE, LEMMA_NAME).getValue());
-  //
-  // }
-  //
-  //
-  // /**
-  // * Tests that when multiple cells in the same column are selected, that the annotation renaming
-  // * works for all of these cells.
-  // *
[…]
