+++ tests/org.corpus_tools.hexatomic.it.tests/src/main/java/org/corpus_tools/hexatomic/it/tests/TestGridEditor.java
@@ -20,6 +20,7 @@ import java.util.Optional;
 import java.util.concurrent.Callable;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.FutureTask;
+import org.apache.commons.lang3.SystemUtils;
 import org.apache.commons.lang3.tuple.Pair;
 import org.corpus_tools.hexatomic.core.CommandParams;
 import org.corpus_tools.hexatomic.core.ProjectManager;
@@ -86,6 +87,8 @@ import org.junit.jupiter.api.Test;
  */
 public class TestGridEditor {
 
+
+  private static final String AUTO_RESIZE_ROW_S = "Auto-resize row(s)";
 
   private static final String AUTO_RESIZE_COLUMN_S = "Auto-resize column(s)";
 
@@ -519,9 +522,14 @@ public class TestGridEditor {
     combo.setSelection(0);
     assertEquals(TEXT1_CAPTION, combo.getText());
 
-    combo.pressShortcut(KeyStroke.getInstance(SWT.ARROW_DOWN));
-    assertEquals("Token annotations only", combo.getText());
+    // Using the keyboard arrows only works on some systems, e.g. on macOS you can only click on the
+    // Combo Box to select the item
+    if (SystemUtils.IS_OS_LINUX) {
+      combo.pressShortcut(KeyStroke.getInstance(SWT.ARROW_DOWN));
+      assertEquals("Token annotations only", combo.getText());
+    }
   }
+
 
   @Test
   void testDropdownWithTextButNoToken() {
@@ -570,12 +578,16 @@ public class TestGridEditor {
     SWTNatTableBot tableBot = new SWTNatTableBot();
     SWTBotNatTable table = tableBot.nattable();
 
-    List<String> columnItems = table.contextMenu(0, 1).menuItems();
+    SWTBotRootMenu contextMenu = table.contextMenu(0, 1);
+    List<String> columnItems = contextMenu.menuItems();
     assertEquals("Hide column(s)", columnItems.get(0));
     assertEquals(SHOW_ALL_COLUMNS, columnItems.get(1));
     assertEquals(AUTO_RESIZE_COLUMN_S, columnItems.get(3));
     assertEquals("Set column freeze", columnItems.get(5));
     assertEquals("Toggle freeze", columnItems.get(6));
+
+    // Click on an item to ensure the context menu is closed
+    contextMenu.menu(AUTO_RESIZE_COLUMN_S).click();
   }
 
   @Test
@@ -585,10 +597,14 @@ public class TestGridEditor {
     SWTNatTableBot tableBot = new SWTNatTableBot();
     SWTBotNatTable table = tableBot.nattable();
 
-    List<String> columnItems = table.contextMenu(1, 0).menuItems();
-    assertEquals("Auto-resize row(s)", columnItems.get(0));
+    SWTBotRootMenu contextMenu = table.contextMenu(1, 0);
+    List<String> columnItems = contextMenu.menuItems();
+    assertEquals(AUTO_RESIZE_ROW_S, columnItems.get(0));
     assertEquals("Set row freeze", columnItems.get(2));
     assertEquals("Toggle freeze", columnItems.get(3));
+
+    // Click on an item to ensure the context menu is closed
+    contextMenu.menu(AUTO_RESIZE_ROW_S).click();
   }
 
   @Test
@@ -715,7 +731,7 @@ public class TestGridEditor {
   /**
    * Types the value of TEST_ANNOTATION_VALUE, then Return, then waits until the tableToTest has no
    * active cell editors, up to 1000ms.
-   * 
+   *
    * @param tableToTest The {@link NatTable} to operate on
[…]
