### 2023-02-21 4696971
Use NATTable bot to trigger shortcut when adding the column
