### 2023-02-27 b2eab13
Simplify clicking the context menu and use ESC before typing on cell on Mac
