### 2023-02-28 c6121e9
Describe workaround and use constant for test annotation name
