### 2023-02-28 625ea51
Close all open grid editors parts in graph editor test
