### 2023-02-28 6b27cb8
Add tolerance when checking if view point has been reached
