+++ tests/org.corpus_tools.hexatomic.it.tests/src/main/java/org/corpus_tools/hexatomic/it/tests/ViewLocationReachedCondition.java
@@ -10,6 +10,8 @@ class ViewLocationReachedCondition extends DefaultCondition {
   private final Point expected;
   private Point lastChecked;
 
+  private final static int TOLERANCE_IN_PIXELS = 2;
+
   public ViewLocationReachedCondition(Viewport viewport, Point expected) {
     super();
     this.viewport = viewport;
@@ -19,7 +21,9 @@ class ViewLocationReachedCondition extends DefaultCondition {
   @Override
   public boolean test() throws Exception {
     lastChecked = viewport.getViewLocation();
-    return viewport.getViewLocation().equals(expected.x, expected.y);
+    int xDiff = Math.abs(expected.x - viewport.getViewLocation().x);
+    int yDiff = Math.abs(expected.y - viewport.getViewLocation().y);
+    return xDiff <= TOLERANCE_IN_PIXELS && yDiff <= TOLERANCE_IN_PIXELS;
   }
 
   @Override
