+++ tests/org.corpus_tools.hexatomic.it.tests/src/main/java/org/corpus_tools/hexatomic/it/tests/TestGridEditor.java
@@ -746,7 +746,7 @@ public class TestGridEditor {
       keyboard.pressShortcut(Keystrokes.ESC);
     }
 
-    keyboard.typeText(TEST_ANNOTATION_VALUE);
+    keyboard.typeText(TEST_ANNOTATION_VALUE, 10);
     keyboard.pressShortcut(Keystrokes.CR);
     bot.waitUntil(new DefaultCondition() {
 
@@ -957,7 +957,7 @@ public class TestGridEditor {
 
     table.contextMenu(0, 2).contextMenu(GridEditor.CHANGE_ANNOTATION_NAME_POPUP_MENU_LABEL).click();
     SWTBotShell dialog = tableBot.shell(RENAME_DIALOG_TITLE);
-    keyboard.typeText(TEST_ANNOTATION_VALUE);
+    keyboard.typeText(TEST_ANNOTATION_VALUE, 10);
     tableBot.button("OK").click();
     bot.waitUntil(Conditions.shellCloses(dialog));
     assertEquals(NAMESPACE + TEST_ANNOTATION_VALUE, table.getCellDataValueByPosition(0, 2));
@@ -977,7 +977,7 @@ public class TestGridEditor {
 
     table.contextMenu(0, 2).contextMenu(GridEditor.CHANGE_ANNOTATION_NAME_POPUP_MENU_LABEL).click();
     SWTBotShell dialog = tableBot.shell(RENAME_DIALOG_TITLE);
-    keyboard.typeText(TEST_ANNOTATION_VALUE);
+    keyboard.typeText(TEST_ANNOTATION_VALUE, 10);
     keyboard.pressShortcut(Keystrokes.CR);
     bot.waitUntil(Conditions.shellCloses(dialog));
     assertEquals(NAMESPACE + TEST_ANNOTATION_VALUE, table.getCellDataValueByPosition(0, 2));
@@ -997,7 +997,7 @@ public class TestGridEditor {
 
     table.contextMenu(0, 2).contextMenu(GridEditor.CHANGE_ANNOTATION_NAME_POPUP_MENU_LABEL).click();
     SWTBotShell dialog = tableBot.shell(RENAME_DIALOG_TITLE);
-    keyboard.typeText(TEST_ANNOTATION_VALUE);
+    keyboard.typeText(TEST_ANNOTATION_VALUE, 10);
     tableBot.button(CANCEL).click();
     bot.waitUntil(Conditions.shellCloses(dialog));
     assertEquals(NAMESPACE + LEMMA_NAME, table.getCellDataValueByPosition(0, 2));
@@ -1038,7 +1038,7 @@ public class TestGridEditor {
 
     table.contextMenu(0, 2).contextMenu(GridEditor.CHANGE_ANNOTATION_NAME_POPUP_MENU_LABEL).click();
     SWTBotShell dialog = tableBot.shell(RENAME_DIALOG_TITLE);
-    keyboard.typeText(TEST_ANNOTATION_VALUE);
+    keyboard.typeText(TEST_ANNOTATION_VALUE, 10);
     tableBot.button("OK").click();
     bot.waitUntil(Conditions.shellCloses(dialog));
     assertTrue(projectManager.isDirty());
@@ -1093,7 +1093,7 @@ public class TestGridEditor {
     // Change annotation name in first overlapping column (currently at position 3)
     table.contextMenu(0, 3).contextMenu(GridEditor.CHANGE_ANNOTATION_NAME_POPUP_MENU_LABEL).click();
     SWTBotShell dialog = tableBot.shell(RENAME_DIALOG_TITLE);
-    keyboard.typeText(TEST_ANNOTATION_VALUE);
+    keyboard.typeText(TEST_ANNOTATION_VALUE, 10);
     tableBot.button("OK").click();
[…]
