### 2023-02-28 ea601c9
Adjust setup/tear down logic in TestTextViewerr to the other integration tests
