+++ tests/org.corpus_tools.hexatomic.it.tests/src/main/java/org/corpus_tools/hexatomic/it/tests/TestTextViewer.java
@@ -17,11 +17,11 @@ import org.eclipse.e4.core.commands.ECommandService;
 import org.eclipse.e4.core.commands.EHandlerService;
 import org.eclipse.e4.core.contexts.ContextInjectionFactory;
 import org.eclipse.e4.core.contexts.IEclipseContext;
-import org.eclipse.e4.ui.workbench.modeling.EPartService;
 import org.eclipse.emf.common.util.URI;
 import org.eclipse.swtbot.e4.finder.widgets.SWTBotView;
 import org.eclipse.swtbot.e4.finder.widgets.SWTWorkbenchBot;
 import org.eclipse.swtbot.swt.finder.widgets.SWTBotTreeItem;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.MethodOrderer.OrderAnnotation;
 import org.junit.jupiter.api.Test;
@@ -38,12 +38,13 @@ class TestTextViewer {
 
   private ProjectManager projectManager;
 
+  private static final String CORPUS_EDITOR_PART_ID =
+      "org.corpus_tools.hexatomic.corpusedit.part.corpusstructure";
 
 
   @BeforeEach
   void setup() {
     TestHelper.setKeyboardLayout();
-
     IEclipseContext ctx = TestHelper.getEclipseContext();
 
     projectManager = ContextInjectionFactory.make(ProjectManager.class, ctx);
@@ -53,28 +54,21 @@ class TestTextViewer {
 
     handlerService = ctx.get(EHandlerService.class);
     assertNotNull(handlerService);
-    
-    EPartService partService = ctx.get(EPartService.class);
-    assertNotNull(partService);
 
     File exampleProjectDirectory = new File("../org.corpus_tools.hexatomic.core.tests/"
         + "src/main/resources/org/corpus_tools/hexatomic/core/example-corpus/");
     assertTrue(exampleProjectDirectory.isDirectory());
-
     exampleProjectUri = URI.createFileURI(exampleProjectDirectory.getAbsolutePath());
+  }
 
-    // Programmatically start a new salt project to get a clean state
-    Map<String, String> params = new HashMap<>();
-    params.put(CommandParams.FORCE_CLOSE, "true");
-    ParameterizedCommand cmd = commandService
-        .createCommand("org.corpus_tools.hexatomic.core.command.new_salt_project", params);
-    handlerService.executeHandler(cmd);
+  @AfterEach
+  void close() {
+    // Programmatically close the example corpus by opening a new fresh one
+    TestHelper.executeNewProjectCommand(commandService, handlerService);
   }
 
 
-  void openDefaultExample() {
-
-    // Programmatically open the example corpus
+  SWTBotView openDefaultExample() {
     Map<String, String> params = new HashMap<>();
     params.put(CommandParams.LOCATION, exampleProjectUri.toFileString());
     params.put(CommandParams.FORCE_CLOSE, "true");
@@ -87,12 +81,22 @@ class TestTextViewer {
 
     // Select the first example document
     SWTBotTreeItem docMenu = bot.tree().expandNode("corpusGraph1").expandNode("rootCorpus")
-        .expandNode("subCorpus1").expandNode("doc1");
+        .expandNode("subCorpus1").expandNode("doc2");
 
     // select and open the editor
     docMenu.click();
     assertNotNull(docMenu.contextMenu("Open with Text Viewer").click());
 
+    SWTBotView view = bot.partByTitle("doc2 (Text Viewer)");
+    assertNotNull(view);
+
+    // Use all available windows space (the tableToTest needs to be fully visible for some of the
+    // tests)
+    bot.waitUntil(new PartActiveCondition(view.getPart()));
[…]
