### 2023-02-28 bfd49a9
Cleanup after execution of TestCorpusStructure
