### 2023-02-28 e92e82b
Use an AWT keyboard in GraphEditor test when typing on the NatTable cell
