+++ tests/org.corpus_tools.hexatomic.it.tests/src/main/java/org/corpus_tools/hexatomic/it/tests/TestGraphEditor.java
@@ -19,6 +19,8 @@ import java.util.Objects;
 import java.util.Optional;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
+
+import org.apache.commons.lang3.SystemUtils;
 import org.corpus_tools.hexatomic.core.CommandParams;
 import org.corpus_tools.hexatomic.core.ProjectManager;
 import org.corpus_tools.hexatomic.core.UiStatusReport;
@@ -77,1052 +79,1038 @@ import org.junit.jupiter.api.TestMethodOrder;
 @TestMethodOrder(OrderAnnotation.class)
 class TestGraphEditor {
 
-  private static final String DOC1_SALT_ID = "salt:/rootCorpus/subCorpus1/doc1";
-  private static final String DOC1_TITLE = "doc1 (Graph Editor)";
-  private static final String INF_STRUCT = "Inf-Struct";
-  private static final String CONST = "const";
-  private static final String SEARCH = "Search";
-  private static final String ANNOTATION_NAME = "Node Annotations";
-  private static final String SPANS = "Spans";
-  private static final String FILTER_VIEW = "Filter View";
-  private static final String GRAPH_LAYOUT = "Graph Layout";
-  private static final String ANNOTATION_TYPES = "Annotation Types";
-  private static final String CORPUS_EDITOR_PART_ID =
-      "org.corpus_tools.hexatomic.corpusedit.part.corpusstructure";
-  private static final String TEST_TOKEN = "abc";
-  private static final String ADD_POINTING_COMMMAND = "e #structure3 -> #structure5";
-  private static final String ANOTHER_TEXT = "Another text";
-  private final SWTWorkbenchBot bot = new SWTWorkbenchBot(TestHelper.getEclipseContext());
-  private static final String GET_VIEWPORT = "getViewport";
-
-  private static final String GET_VIEW_LOCATION = "getViewLocation";
-
-  private URI exampleProjectUri;
-  private ECommandService commandService;
-  private EHandlerService handlerService;
-
-  private ErrorService errorService;
-  private ProjectManager projectManager;
-  private UiStatusReport uiStatus;
-
-  private final Keyboard keyboard = KeyboardFactory.getSWTKeyboard();
-
-  private final class HorizontalNodeDistanceCondition extends DefaultCondition {
-    private final SNode leftNode;
-    private final SNode rightNode;
-    private final double expected;
-    private final Graph graphWidget;
-
-    private HorizontalNodeDistanceCondition(SNode leftNode, SNode rightNode, double expected,
-        Graph graphWidget) {
-      this.leftNode = leftNode;
-      this.rightNode = rightNode;
-      this.expected = expected;
-      this.graphWidget = graphWidget;
-    }
-
-    private double getDistance() {
-      Rectangle l = getGraphNodeForSalt(bot, graphWidget, leftNode).getNodeFigure().getBounds();
-      Rectangle r = getGraphNodeForSalt(bot, graphWidget, rightNode).getNodeFigure().getBounds();
-      return Math.abs(l.getRight().x - r.getLeft().x);
-    }
-
-    @Override
-    public boolean test() throws Exception {
-      double diff = Math.abs(getDistance() - expected);
-      // Allow for 20% tolerance, but at least 1 point
-      return diff <= Math.max(1.0, expected * 0.2);
-    }
-
-    @Override
-    public String getFailureMessage() {
-      return "Horizontal distance between node " + leftNode.getName() + " and "
-          + rightNode.getName() + " should have been " + expected + " but was " + getDistance()
-          + ".";
-    }
-  }
-
[…]
