+++ tests/org.corpus_tools.hexatomic.it.tests/src/main/java/org/corpus_tools/hexatomic/it/tests/TestGridEditor.java
@@ -20,7 +20,6 @@ import java.util.Optional;
 import java.util.concurrent.Callable;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.FutureTask;
-
 import org.apache.commons.lang3.SystemUtils;
 import org.apache.commons.lang3.tuple.Pair;
 import org.corpus_tools.hexatomic.core.CommandParams;
@@ -88,2411 +87,2432 @@ import org.junit.jupiter.api.Test;
  */
 public class TestGridEditor {
 
-	private static final String AUTO_RESIZE_ROW_S = "Auto-resize row(s)";
+  private static final String AUTO_RESIZE_ROW_S = "Auto-resize row(s)";
 
-	private static final String AUTO_RESIZE_COLUMN_S = "Auto-resize column(s)";
+  private static final String AUTO_RESIZE_COLUMN_S = "Auto-resize column(s)";
 
-	private static final String SHOW_ALL_COLUMNS = "Show all columns";
+  private static final String SHOW_ALL_COLUMNS = "Show all columns";
 
-	private static final String REFRESH_GRID = "Refresh grid";
+  private static final String REFRESH_GRID = "Refresh grid";
 
-	private static final String TEXT1_CAPTION = "sText1";
+  private static final String TEXT1_CAPTION = "sText1";
 
-	private static final String DOC_GRID_EDITOR = "doc (Grid Editor)";
+  private static final String DOC_GRID_EDITOR = "doc (Grid Editor)";
 
-	private static final String DOC = "doc";
+  private static final String DOC = "doc";
 
-	private static final String CORPUS = "corpus";
+  private static final String CORPUS = "corpus";
 
-	private static final String UNRENAMED_ANNOTATIONS_DIALOG_TITLE = "Some annotations were not renamed!";
+  private static final String UNRENAMED_ANNOTATIONS_DIALOG_TITLE =
+      "Some annotations were not renamed!";
 
-	private static final String TOKEN_VALUE = "Token";
-	private static final String SPAN_VALUE = "Span";
+  private static final String TOKEN_VALUE = "Token";
+  private static final String SPAN_VALUE = "Span";
 
-	private static final String OPEN_WITH_GRID_EDITOR = "Open with Grid Editor";
+  private static final String OPEN_WITH_GRID_EDITOR = "Open with Grid Editor";
 
-	private static final String TEST_ANNOTATION_NAME = "TEST";
-	private static final String TEST_ANNOTATION_VALUE = "test";
-	private static final String CONTRAST_FOCUS_VALUE = "contrast-focus";
-	private static final String MORE_VALUE = "more";
-	private static final String COMPLICATED_VALUE = "complicated";
-	private static final String IT_VALUE = "it";
-	private static final String JJ_VALUE = "JJ";
-	private static final String EXAMPLE_VALUE = "example";
-	private static final String TOPIC_VALUE = "topic";
+  private static final String TEST_ANNOTATION_NAME = "TEST";
+  private static final String TEST_ANNOTATION_VALUE = "test";
+  private static final String CONTRAST_FOCUS_VALUE = "contrast-focus";
+  private static final String MORE_VALUE = "more";
+  private static final String COMPLICATED_VALUE = "complicated";
+  private static final String IT_VALUE = "it";
+  private static final String JJ_VALUE = "JJ";
+  private static final String EXAMPLE_VALUE = "example";
+  private static final String TOPIC_VALUE = "topic";
 
[…]
