### 2023-02-28 79f76e2
Address some code style violations
