### 2023-02-28 28d3ebb
Rethrow exception instead of failing
