### 2023-02-28 1830b29
Fix some code smell reported by SonaCloud
