### 2023-02-28 5754658
Use more recent Ubuntu LTS version for GH actions
