+++ .github/workflows/release.yml
@@ -7,7 +7,7 @@ on:
 jobs:
   deploy_binaries:
     name: Deploy release Windows and Linux binaries
-    runs-on: ubuntu-20.04
+    runs-on: ubuntu-22.04
     steps:
         - name: Update packages
           run: sudo apt-get update
