+++ .github/workflows/test.yml
@@ -4,7 +4,7 @@ on: push
 jobs:
   test_jdk11:
     name: OpenJDK 11 Linux + SonarCloud code analysis
-    runs-on: ubuntu-20.04
+    runs-on: ubuntu-22.04
     steps:
       - name: Update packages
         run: sudo apt-get update
