### 2023-02-28 ac68750
Use Monterey image for tests
