### 2023-02-28 3666997
Use setup-java action for macos
