+++ .github/workflows/test.yml
@@ -35,7 +35,11 @@ jobs:
     runs-on: macos-12
     steps:
       - name: Checkout
-        uses: actions/checkout@v2
+        uses: actions/checkout@v3
+      - uses: actions/setup-java@v3
+	    with:
+	      java-version: '11'
+	      distribution: 'temurin'
       - name: Cache local Maven repository
         uses: actions/cache@v2
         with:
@@ -43,8 +47,6 @@ jobs:
           key: ${{ runner.os }}-maven-11-${{ hashFiles('**/pom.xml') }}
           restore-keys: |
             ${{ runner.os }}-maven-11-
-      - name: Set JAVA_HOME to OpenJDK 11
-        run: echo JAVA_HOME=$JAVA_HOME_11_X64 >> $GITHUB_ENV
       - name: Run Maven install (includes tests)
         run: mvn verify
         env:
@@ -52,7 +54,7 @@ jobs:
 
   test_docs:
     name: Test documentation
-    runs-on: ubuntu-20.04
+    runs-on: ubuntu-22.04
     steps:
       - name: Checkout
         uses: actions/checkout@v2
