+++ .github/workflows/test.yml
@@ -37,9 +37,9 @@ jobs:
       - name: Checkout
         uses: actions/checkout@v3
       - uses: actions/setup-java@v3
-	    with:
-	      java-version: '11'
-	      distribution: 'temurin'
+        with:
+          java-version: '11'
+          distribution: 'temurin'
       - name: Cache local Maven repository
         uses: actions/cache@v2
         with:
