### 2023-02-28 86b72b8
Use older MacOS version which might come with an older Maven version.
Maven 3.9.0 has some issues with one of the dependencies: https://github.com/mojohaus/build-helper-maven-plugin/issues/168

This is hopefully fixed im Maven 3.9.1
