+++ .github/workflows/test.yml
@@ -32,7 +32,7 @@ jobs:
           LANG: en_US.UTF-8
   test_macos:
     name: OpenJDK 11 macOS
-    runs-on: macos-11
+    runs-on: macos-12
     steps:
       - name: Checkout
         uses: actions/checkout@v3
@@ -40,6 +40,10 @@ jobs:
         with:
           java-version: '11'
           distribution: 'temurin'
+      - name: Set up Maven
+        uses: stCarolas/setup-maven@v4.5
+        with:
+          maven-version: 3.8.7
       - name: Cache local Maven repository
         uses: actions/cache@v2
         with:
