+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Running the tests on macOS would fail (#417). An GitHub action is now
+  configured to execute the tests also on macOS for each pull request.
+
 ## [1.1.0] - 2023-02-19
 
 ### Fixed
