### 2023-02-28 944bfab
Also specify Maven version on Linux build
