+++ .github/workflows/test.yml
@@ -11,10 +11,14 @@ jobs:
       - name: Install metacity window manager for background tests
         run: sudo apt-get install metacity
       - name: Checkout
-        uses: actions/checkout@v2
+        uses: actions/checkout@v3
         with:
           # Otherwise the new code can't be determined or code coverage is less informing
           fetch-depth: 0
+      - name: Set up Maven
+        uses: stCarolas/setup-maven@v4.5
+        with:
+          maven-version: 3.8.7
       - name: Cache local Maven repository
         uses: actions/cache@v2
         with:
