### 2023-02-28 e58f56b
Update CITATION.cff to newest snapshot version
