+++ CITATION.cff
@@ -34,8 +34,8 @@ identifiers:
   type: doi
   value: 10.5281/zenodo.6900689
 doi: 10.5281/zenodo.6900689
-version: 1.1.0
-date-released: 2023-02-19
+version: 1.2.0-SNAPSHOT
+date-released: 2023-02-28
 references:
 - type: software
   title: SWT Resource Manager
