+++ tests/org.corpus_tools.hexatomic.it.tests/src/main/java/org/corpus_tools/hexatomic/it/tests/TestGraphEditor.java
@@ -674,7 +674,6 @@ class TestGraphEditor {
   }
 
   @Test
-  @EnabledOnOs({OS.WINDOWS, OS.LINUX})
   void testFilterOptions() {
     openDefaultExample();
 
@@ -733,6 +732,7 @@ class TestGraphEditor {
 
   @Test
 
+  @EnabledOnOs({OS.WINDOWS, OS.LINUX})
   void testLayoutParameters() {
     openDefaultExample();
 
