### 2023-02-28 93b10fc
Do not catch interrupted exception in test
