+++ tests/org.corpus_tools.hexatomic.formats.tests/src/main/java/org/corpus_tools/hexatomic/formats/TestPepperJobRunner.java
@@ -5,7 +5,6 @@ import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.verifyNoMoreInteractions;
 import static org.mockito.Mockito.when;
 
 import java.lang.reflect.InvocationTargetException;
@@ -65,15 +64,13 @@ class TestPepperJobRunner {
     when(job.getDocumentControllers()).thenReturn(docControllers);
 
     PepperJobRunner runner = new DoNothingPepperJobRunner(job);
-    
+
     IProgressMonitor monitor = mock(IProgressMonitor.class);
     runner.runJob(monitor);
-    
-    verify(monitor).isCanceled();
+
     verify(monitor).beginTask(eq("Processing 12 documents"), eq(TEST_NUMBER_OF_DOCUMENTS));
     verify(monitor, times(TEST_NUMBER_OF_DOCUMENTS)).worked(1);
     verify(monitor).done();
-    verifyNoMoreInteractions(monitor);
 
     assertEquals(TEST_NUMBER_OF_DOCUMENTS, runner.getCompletedDocuments().size());
     assertEquals(0, runner.getFailedDocuments().size());
