### 2023-02-28 a2ad15c
Test when no documents are imported
