### 2023-02-28 14f7a55
Also test the case there are failed documents in the pepper job
