+++ tests/org.corpus_tools.hexatomic.it.tests/src/main/java/org/corpus_tools/hexatomic/it/tests/TestCorpusStructure.java
@@ -23,7 +23,7 @@ import org.eclipse.swtbot.swt.finder.keyboard.KeyboardFactory;
 import org.eclipse.swtbot.swt.finder.keyboard.Keystrokes;
 import org.eclipse.swtbot.swt.finder.utils.SWTBotPreferences;
 import org.eclipse.swtbot.swt.finder.waits.Conditions;
-import org.eclipse.swtbot.swt.finder.widgets.SWTBotTreeItem;
+import org.eclipse.swtbot.swt.finder.waits.DefaultCondition;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.MethodOrderer.OrderAnnotation;
@@ -34,6 +34,32 @@ import org.junit.jupiter.api.TestMethodOrder;
 @TestMethodOrder(OrderAnnotation.class)
 public class TestCorpusStructure {
 
+
+  private final class DocumentRenamedCondition extends DefaultCondition {
+
+    private final String documentID;
+    private final String expectedName;
+
+    public DocumentRenamedCondition(String documentID, String expectedName) {
+      this.documentID = documentID;
+      this.expectedName = expectedName;
+    }
+
+    @Override
+    public boolean test() throws Exception {
+      Optional<SDocument> doc1 = projectManager.getDocument(documentID);
+      if (doc1.isPresent()) {
+        return doc1.get().getName().equals(expectedName);
+      } else {
+        return false;
+      }
+    }
+
+    @Override
+    public String getFailureMessage() {
+      return "Document " + documentID + " has not been renamed to " + expectedName;
+    }
+  }
 
   private static final String SALT_PREFIX = "salt:/";
 
@@ -125,21 +151,20 @@ public class TestCorpusStructure {
     bot.tree().getTreeItem(CORPUS_GRAPH_1).getNode(CORPUS_1).getNode(DOCUMENT_1).select();
     bot.tree().getTreeItem(CORPUS_GRAPH_1).getNode(CORPUS_1).getNode(DOCUMENT_1).doubleClick();
     bot.text(DOCUMENT_1).setText(ABC).setFocus();
-    keyboard.pressShortcut(Keystrokes.LF);
+    keyboard.pressShortcut(Keystrokes.CR);
 
-    bot.waitUntil(Conditions.treeItemHasNode(bot.tree().getTreeItem(CORPUS_GRAPH_1), CORPUS_1));
+    bot.waitUntil(new DocumentRenamedCondition(SALT_PREFIX + CORPUS_1 + "/" + DOCUMENT_1, ABC));
 
-    SWTBotTreeItem corpusNode = bot.tree().expandNode(CORPUS_GRAPH_1).expandNode(CORPUS_1);
-    bot.waitUntil(Conditions.treeItemHasNode(corpusNode, DOCUMENT_2));
-    SWTBotTreeItem documentNode = corpusNode.expandNode(DOCUMENT_2);
 
-    documentNode.select();
-    documentNode.doubleClick();
+    bot.tree().expandNode(CORPUS_GRAPH_1).expandNode(CORPUS_1).expandNode(DOCUMENT_2);
+
+    bot.tree().getTreeItem(CORPUS_GRAPH_1).getNode(CORPUS_1).getNode(DOCUMENT_2).select();
+    bot.tree().getTreeItem(CORPUS_GRAPH_1).getNode(CORPUS_1).getNode(DOCUMENT_2).doubleClick();
     bot.text(DOCUMENT_2).setText(DEF).setFocus();
-    keyboard.pressShortcut(Keystrokes.LF);
[…]
