+++ .github/workflows/test.yml
@@ -2,8 +2,8 @@ name: Automated tests
 on: push
 
 jobs:
-  test_jdk11:
-    name: OpenJDK 11 Linux + SonarCloud code analysis
+  test_jdk:
+    name: OpenJDK 17 Linux + SonarCloud code analysis
     runs-on: ubuntu-22.04
     steps:
       - name: Update packages
@@ -26,8 +26,8 @@ jobs:
           key: ${{ runner.os }}-maven-11-${{ hashFiles('**/pom.xml') }}
           restore-keys: |
             ${{ runner.os }}-maven-11-
-      - name: Set JAVA_HOME to OpenJDK 11
-        run: echo JAVA_HOME=$JAVA_HOME_11_X64 >> $GITHUB_ENV
+      - name: Set JAVA_HOME to OpenJDK 17
+        run: echo JAVA_HOME=$JAVA_HOME_17_X64 >> $GITHUB_ENV
       - name: Run Maven Tests with enabled code coverage and upload results to SonarCloud
         run: ./releng/sh/metacity-run.sh mvn clean verify sonar:sonar -Pcoverage -Dsonar.projectKey=hexatomic_hexatomic -Dsonar.organization=hexatomic -Dsonar.host.url=https://sonarcloud.io -Dsonar.login=$SONAR_TOKEN
         env:
@@ -35,14 +35,14 @@ jobs:
           SONAR_TOKEN: ${{ secrets.SONAR_TOKEN }}
           LANG: en_US.UTF-8
   test_macos:
-    name: OpenJDK 11 macOS
+    name: OpenJDK 12 macOS
     runs-on: macos-12
     steps:
       - name: Checkout
         uses: actions/checkout@v3
       - uses: actions/setup-java@v3
         with:
-          java-version: '11'
+          java-version: '17'
           distribution: 'temurin'
       - name: Set up Maven
         uses: stCarolas/setup-maven@v4.5
