### 2023-02-28 eaabd4b
Since we now bundle the JRE, there is no need to specify which OpenJDK we test against (its just one)
