### 2023-02-28 a2406e7
Use String constants in TestGridEditor
