+++ CITATION.cff
@@ -86,6 +86,27 @@ references:
   authors:
   - name: Takahiko Kawasaki
 - type: software
+  title: SAT4J core
+  version: 2.3.6
+  abbreviation: org.ow2.sat4j:org.ow2.sat4j.core
+  license: EPL-1.0
+  notes: More license information can be found in the THIRD-PARTY/SAT4J_core directory.
+  authors:
+  - name: Daniel Le Berre
+  - name: Stephanie Roussel
+  - name: Emmanuel Lonca
+  repository-code: https://gitlab.ow2.org/sat4j/sat4j/org.ow2.sat4j.core/
+- type: software
+  title: SAT4J pseudo
+  version: 2.3.6
+  abbreviation: org.ow2.sat4j:org.ow2.sat4j.pb
+  license: EPL-1.0
+  notes: More license information can be found in the THIRD-PARTY/SAT4J_pseudo directory.
+  authors:
+  - name: Daniel Le Berre
+  - name: Anne Parrain
+  repository-code: https://gitlab.ow2.org/sat4j/sat4j/org.ow2.sat4j.pb/
+- type: software
   title: Simple Logging Facade for Java
   abbreviation: SLF4J
   version: 1.7.30.v20200204-2150
@@ -114,7 +135,7 @@ references:
   repository-code: https://github.com/FasterXML/woodstox
 - type: software
   title: ch.qos.logback.classic
-  version: 1.2.3.v20200428-2012
+  version: 1.1.2.v20171220-1825
   license: EPL-1.0
   authors:
   - name: QOS.ch
@@ -123,7 +144,7 @@ references:
     directory.
 - type: software
   title: ch.qos.logback.core
-  version: 1.2.3.v20200428-2012
+  version: 1.1.2.v20160208-0839
   license: EPL-1.0
   authors:
   - name: QOS.ch
@@ -155,7 +176,7 @@ references:
     directory.
 - type: software
   title: javax.inject
-  version: 1.0.0.v20091030
+  version: 1.0.0.v20220405-0441
   license: EPL-1.0
   authors:
   - name: atinject project
@@ -242,14 +263,14 @@ references:
     directory.
 - type: software
   title: org.bouncycastle.bcpg
-  version: 1.69.0.v20210713-1924
+  version: 1.70.0.v20220507-1208
   authors:
-  - name: The org.bouncycastle.bcpg 1.69.0.v20210713-1924 Team
+  - name: The org.bouncycastle.bcpg 1.70.0.v20220507-1208 Team
 - type: software
   title: org.bouncycastle.bcprov
-  version: 1.69.0.v20210923-1401
+  version: 1.70.0.v20220507-1208
   authors:
-  - name: The org.bouncycastle.bcprov 1.69.0.v20210923-1401 Team
+  - name: The org.bouncycastle.bcprov 1.70.0.v20220507-1208 Team
 - type: software
   title: org.corpus-tools.pepper-framework
   version: 3.6.0
@@ -263,7 +284,7 @@ references:
   - name: The org.corpus-tools.salt-api 3.4.2 Team
 - type: software
   title: org.eclipse.core.commands
-  version: 3.10.100.v20210722-1426
+  version: 3.10.200.v20220512-0851
   license: EPL-2.0
   authors:
   - name: The Eclipse Foundation
@@ -281,7 +302,7 @@ references:
     directory.
 - type: software
   title: org.eclipse.core.databinding
-  version: 1.10.100.v20200926-1123
+  version: 1.11.0.v20220118-1028
   license: EPL-2.0
   authors:
   - name: The Eclipse Foundation
@@ -290,7 +311,7 @@ references:
     directory.
 - type: software
[…]
