### 2023-02-28 e823b20
Fix filter test
