### 2023-02-28 a9af032
Make test for filter options more robust by waiting for the chips to appear
