### 2023-03-01 7d140ca
Use AWT keyboard to type inf-struct anno name (otherwiese the - might be typed as &)
