### 2023-03-01 21dd588
Adjust a grid test to run in macos terminal
