### 2023-03-01 fec7bfd
On non-mac systems editing multiple cell values does not trigger a new dialog
