### 2023-03-01 47c3f50
Use awt keyboard when typing the filter options
