### 2023-03-01 20d79c6
Don't fail if editor part is not opened
