### 2023-03-01 38de75a
Synchronize release workflow for macos with its test workflow
fixes #404
