+++ .github/workflows/release.yml
@@ -29,7 +29,7 @@ jobs:
           uses: softprops/action-gh-release@v0.1.5
           with:
             files: |
-              releng/org.corpus_tools.hexatomic.product/target/products/hexatomic-*-linux.gtk.x86_64.zip
+              releng/org.corpus_tools.hexatomic.product/target/products/hexatomic-*-linux.gtk.x86_64.tar.gz
               releng/org.corpus_tools.hexatomic.product/target/products/hexatomic-*-win32.win32.x86_64.zip  
             draft: true
           env:
