### 2023-03-02 1fbb089
Adjust target platform and manifests and reduce orbit dependencies
