+++ tests/org.corpus_tools.hexatomic.core.tests/src/main/java/org/corpus_tools/hexatomic/core/handler/TestSaveAsHandler.java
@@ -29,7 +29,7 @@ public class TestSaveAsHandler {
 
 
   @BeforeEach
-  private void setUp() {
+  public void setUp() {
     dialog = mock(DirectoryDialog.class);
 
     this.fixture = new SaveAsHandler();
@@ -47,8 +47,8 @@ public class TestSaveAsHandler {
 
     shell = mock(Shell.class);
     projectManager = mock(ProjectManager.class);
-    
-    
+
+
     this.fixture.setProjectManager(projectManager);
 
   }
