### 2023-03-02 d79ef4d
Add preconfigured non-ui tests
