+++ tests/org.corpus_tools.hexatomic.core.tests/src/main/java/org/corpus_tools/hexatomic/core/handler/TestSaveAsHandler.java
@@ -29,7 +29,7 @@ public class TestSaveAsHandler {
 
 
   @BeforeEach
-  public void setUp() {
+  void setUp() {
     dialog = mock(DirectoryDialog.class);
 
     this.fixture = new SaveAsHandler();
