+++ CITATION.cff
@@ -35,7 +35,7 @@ identifiers:
   value: 10.5281/zenodo.6900689
 doi: 10.5281/zenodo.6900689
 version: 1.2.0-SNAPSHOT
-date-released: 2023-03-01
+date-released: 2023-03-02
 references:
 - type: software
   title: SWT Resource Manager
@@ -44,12 +44,88 @@ references:
   - name: Google, Inc.
 - type: software
   title: ANTLR 4 Runtime
-  version: 4.7.2.v20200218-0804
+  version: 4.7.2.v20221112-0806
   license: BSD-3-Clause
   authors:
   - name: The ANTLR Project
   notes: More license information can be found in the THIRD-PARTY/org.antlr.runtime
     directory.
+- type: software
+  title: Animal Sniffer Annotations
+  version: '1.9'
+  abbreviation: org.codehaus.mojo:animal-sniffer-annotations
+  license: MIT
+  notes: More license information can be found in the THIRD-PARTY/Animal_Sniffer_Annotations
+    directory.
+  authors:
+  - name: Kohsuke Kaw
+  - name: Stephen Connolly
+  repository-code: http://fisheye.codehaus.org/browse/mojo/tags/animal-sniffer-parent-1.9/animal-sniffer-annotations
+- type: software
+  title: Apache Commons Lang
+  version: '3.4'
+  abbreviation: org.apache.commons:commons-lang3
+  license: Apache-2.0
+  notes: More license information can be found in the THIRD-PARTY/Apache_Commons_Lang
+    directory.
+  authors:
+  - name: Daniel Rall
+  - name: Stephen Colebourne
+  - name: Henri Yandell
+  - name: Steven Caswell
+  - name: Robert Burrell Donkin
+  - name: Gary D. Gregory
+  - name: Fredrik Westermarck
+  - name: James Carman
+  - name: Niall Pemberton
+  - name: Matt Benson
+  - name: Joerg Schaible
+  - name: Oliver Heger
+  - name: Paul Benedict
+  - name: Benedikt Ritter
+  - name: Duncan Jones
+  repository-code: http://svn.apache.org/viewvc/commons/proper/lang/tags/LANG_3_4
+- type: software
+  title: Bouncy Castle OpenPGP API
+  version: '1.72'
+  abbreviation: org.bouncycastle:bcpg-jdk18on
+  license: Apache-2.0
+  notes: More license information can be found in the THIRD-PARTY/Bouncy_Castle_OpenPGP_API
+    directory.
+  authors:
+  - name: The Legion of the Bouncy Castle Inc.
+  repository-code: https://github.com/bcgit/bc-java
+- type: software
+  title: Bouncy Castle Provider
+  version: '1.72'
+  abbreviation: org.bouncycastle:bcprov-jdk18on
+  license: MIT
+  notes: More license information can be found in the THIRD-PARTY/Bouncy_Castle_Provider
+    directory.
+  authors:
+  - name: The Legion of the Bouncy Castle Inc.
+  repository-code: https://github.com/bcgit/bc-java
+- type: software
+  title: Commons IO
+  version: '2.4'
+  abbreviation: commons-io:commons-io
+  license: Apache-2.0
+  notes: More license information can be found in the THIRD-PARTY/Commons_IO directory.
+  authors:
+  - name: Scott Sanders
+  - name: dIon Gillard
+  - name: Nicola Ken Barozzi
+  - name: Henri Yandell
+  - name: Stephen Colebourne
+  - name: Jeremias Maerki
+  - name: Matthew Hawthorne
+  - name: Martin Cooper
+  - name: Rob Oxspring
+  - name: Jochen Wiedmann
+  - name: Niall Pemberton
[…]
