### 2023-03-02 68c4d73
Allow to set textual data source name
