+++ bundles/org.corpus_tools.hexatomic.transcription/src/main/java/org/corpus_tools/hexatomic/transcription/TranscriptionEditor.java
@@ -18,7 +18,7 @@
  * #L%
  */
 
-package org.corpus_tools.hexatomic.timeline;
+package org.corpus_tools.hexatomic.transcription;
 
 import javax.annotation.PostConstruct;
 import javax.inject.Inject;
@@ -26,12 +26,12 @@ import org.corpus_tools.hexatomic.core.ProjectManager;
 import org.corpus_tools.hexatomic.core.Topics;
 import org.corpus_tools.hexatomic.core.handlers.OpenSaltDocumentHandler;
 import org.corpus_tools.hexatomic.core.undo.ChangeSet;
-import org.corpus_tools.hexatomic.timeline.internal.data.GridDisplayConverter;
-import org.corpus_tools.hexatomic.timeline.internal.data.NodeSpanningDataProvider;
-import org.corpus_tools.hexatomic.timeline.internal.data.TextualDsHeaderDataProvider;
-import org.corpus_tools.hexatomic.timeline.internal.data.TimelineTokenDataProvider;
-import org.corpus_tools.hexatomic.timeline.internal.data.TliCornerDataProvider;
-import org.corpus_tools.hexatomic.timeline.internal.data.TliRowHeaderDataProvider;
+import org.corpus_tools.hexatomic.transcription.internal.data.GridDisplayConverter;
+import org.corpus_tools.hexatomic.transcription.internal.data.NodeSpanningDataProvider;
+import org.corpus_tools.hexatomic.transcription.internal.data.TextualDsHeaderDataProvider;
+import org.corpus_tools.hexatomic.transcription.internal.data.TimelineTokenDataProvider;
+import org.corpus_tools.hexatomic.transcription.internal.data.TliCornerDataProvider;
+import org.corpus_tools.hexatomic.transcription.internal.data.TliRowHeaderDataProvider;
 import org.corpus_tools.salt.common.SDocument;
 import org.corpus_tools.salt.common.SDocumentGraph;
 import org.eclipse.e4.ui.di.UIEventTopic;
@@ -64,10 +64,10 @@ import org.eclipse.swt.layout.GridLayout;
 import org.eclipse.swt.widgets.Button;
 import org.eclipse.swt.widgets.Composite;
 
-public class TimelineTokenAligner {
+public class TranscriptionEditor {
 
   private static final org.slf4j.Logger log =
-      org.slf4j.LoggerFactory.getLogger(TimelineTokenAligner.class);
+      org.slf4j.LoggerFactory.getLogger(TranscriptionEditor.class);
 
   @Inject
   private ProjectManager projectManager;
