+++ bundles/org.corpus_tools.hexatomic.transcription/src/main/java/org/corpus_tools/hexatomic/transcription/internal/data/TextualDsHeaderDataProvider.java
@@ -18,21 +18,42 @@
  * #L%
  */
 
-package org.corpus_tools.hexatomic.timeline.internal.data;
+package org.corpus_tools.hexatomic.transcription.internal.data;
 
+import org.corpus_tools.salt.common.SDocumentGraph;
+import org.eclipse.e4.core.di.annotations.Creatable;
 import org.eclipse.nebula.widgets.nattable.data.IDataProvider;
-import org.eclipse.nebula.widgets.nattable.grid.data.DefaultCornerDataProvider;
 
-public class TliCornerDataProvider extends DefaultCornerDataProvider {
+@Creatable
+public class TextualDsHeaderDataProvider implements IDataProvider {
 
-  public TliCornerDataProvider(IDataProvider columnHeaderDataProvider,
-      IDataProvider rowHeaderDataProvider) {
-    super(columnHeaderDataProvider, rowHeaderDataProvider);
+  private SDocumentGraph graph;
+
+  public void setGraph(SDocumentGraph graph) {
+    this.graph = graph;
   }
 
   @Override
   public Object getDataValue(int columnIndex, int rowIndex) {
-    return "TLI";
+    return graph.getTextualDSs().get(columnIndex).getName();
+  }
+
+  @Override
+  public void setDataValue(int columnIndex, int rowIndex, Object newValue) {
+    if (newValue instanceof String) {
+      graph.getTextualDSs().get(columnIndex).setName((String) newValue);
+    }
+  }
+
+  @Override
+  public int getColumnCount() {
+    // Every textual data source is a column
+    return graph.getTextualDSs().size();
+  }
+
+  @Override
+  public int getRowCount() {
+    return 1;
   }
 
 }
