### 2023-03-03 hexatomic/hexatomic#432 @sdruskat
Move link to tool docs to top of project website
As requested by @reckart in https://github.com/openjournals/joss-reviews/issues/4825#issuecomment-1296210205, tool docs should be more easily/obviously accessible fro the project docs.

The minimal fix for this is to

- move the link to the top of the landing page.
- Rename "Documentation" on the left sidebar to something better suited.

Ideally, we should also change the UI of the project website to something other than mdBook as used also for tool docs (tracked in https://github.com/hexatomic/hexatomic.github.io/issues/38).
