### 2023-03-03 d4337e6
Add links to CITATION.cff and THIRD-party folder in about dialog.
