### 2023-03-03 88e5005
Use more specific label for the "OK" button
