### 2023-03-03 b162c84
Use key without pom.xml
If the pom.xml changes, we still want the maven dependencies to be available
