### 2023-03-03 0edada6
Add tests that the links are actually pointing to the correct location
