+++ tests/org.corpus_tools.hexatomic.it.tests/src/main/java/org/corpus_tools/hexatomic/it/tests/TestHelpMenu.java
@@ -3,9 +3,17 @@ package org.corpus_tools.hexatomic.it.tests;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.ArgumentMatchers.matches;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.verifyNoMoreInteractions;
 
+import org.corpus_tools.hexatomic.core.LinkOpener;
 import org.eclipse.core.runtime.preferences.ConfigurationScope;
 import org.eclipse.core.runtime.preferences.IEclipsePreferences;
+import org.eclipse.e4.core.contexts.IEclipseContext;
+import org.eclipse.swtbot.e4.finder.waits.Conditions;
 import org.eclipse.swtbot.e4.finder.widgets.SWTWorkbenchBot;
 import org.eclipse.swtbot.swt.finder.widgets.SWTBotMenu;
 import org.eclipse.swtbot.swt.finder.widgets.SWTBotShell;
@@ -15,21 +23,35 @@ import org.junit.jupiter.api.condition.EnabledOnOs;
 import org.junit.jupiter.api.condition.OS;
 
 class TestHelpMenu {
+  private static final String USER_GUIDE_REGEX =
+      "https://hexatomic.github.io/hexatomic/user/v[.0-9]+";
+  private static final String ONLINE_DOCUMENTATION = "Online Documentation";
   private IEclipsePreferences prefs =
       ConfigurationScope.INSTANCE.getNode("org.corpus_tools.hexatomic.core");
   private SWTWorkbenchBot bot = new SWTWorkbenchBot(TestHelper.getEclipseContext());
 
+  private LinkOpener linkOpener;
+
   @BeforeEach
   void setup() {
     TestHelper.setKeyboardLayout();
+
+    IEclipseContext ctx = TestHelper.getEclipseContext();
+
+    // Replace the default link opener implementation with a mock
+    linkOpener = mock(LinkOpener.class);
+    ctx.set(LinkOpener.class, linkOpener);
   }
 
 
-
   @Test
-  void testOnlineDocumentationExists() {
+  void testOnlineDocumentationLink() {
     SWTBotMenu helpMenu = bot.menu("Help");
-    helpMenu.menu("Online Documentation").isActive();
+    helpMenu.menu(ONLINE_DOCUMENTATION).isActive();
+    helpMenu.menu(ONLINE_DOCUMENTATION).click();
+
+    verify(linkOpener).open(matches(USER_GUIDE_REGEX));
+    verifyNoMoreInteractions(linkOpener);
   }
 
   @Test
@@ -40,7 +62,37 @@ class TestHelpMenu {
 
     SWTBotShell aboutShell = bot.shell("About Hexatomic");
     aboutShell.bot().label("Hexatomic").isVisible();
-    aboutShell.bot().button("OK").click();
+
+    // Click on the links and check the link opener has been involved
+    aboutShell.bot().link(0).click("https://hexatomic.github.io/");
+    verify(linkOpener).open(eq("https://hexatomic.github.io/"));
+
+    aboutShell.bot().link(1).click("Online User Documentation");
+    verify(linkOpener).open(matches(USER_GUIDE_REGEX));
+
+    aboutShell.bot().link(2).click("Hexatomic project team");
+    verify(linkOpener).open(eq("https://github.com/orgs/hexatomic/teams/project/members"));
+
[…]
