### 2023-03-03 2020df7
Add changelog entry and fix code smell
