### 2023-03-03 d563135
Fix #420 by removing unnecessary layout data
