### 2023-03-03 754dd33
Add regression test for #422
