### 2023-03-03 c04232a
Set checkpoint when deleting item from the corpus structure
