+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.1] - 2023-03-03
+
 ### Fixed
 
 - Deleting a document could not be undone (#422)
@@ -323,7 +325,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - This is the first public preview release of the Hexatomic developer platform.
 
 
-[Unreleased]: https://github.com/hexatomic/hexatomic/compare/v1.1.0...HEAD
+[Unreleased]: https://github.com/hexatomic/hexatomic/compare/v1.1.1...HEAD
+[1.1.1]: https://github.com/hexatomic/hexatomic/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/hexatomic/hexatomic/compare/v1.0.1...v1.1.0
 [1.0.1]: https://github.com/hexatomic/hexatomic/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/hexatomic/hexatomic/compare/v0.14.0...v1.0.0
