### 2023-03-03 2dc5c47
Explicitly set the Maven version in release jobs
