+++ CITATION.cff
@@ -34,7 +34,7 @@ identifiers:
   type: doi
   value: 10.5281/zenodo.6900689
 doi: 10.5281/zenodo.6900689
-version: 1.1.1
+version: 1.1.2
 date-released: 2023-03-03
 references:
 - type: software
