### 2023-03-03 bb0a7f3
Set Java version in P2 Update site workflow
