+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.3] - 2023-03-03
+
 ### Fixed
 
 - Excplicitly set the Java version in the workflow to publish the P2 update
@@ -338,7 +340,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - This is the first public preview release of the Hexatomic developer platform.
 
 
-[Unreleased]: https://github.com/hexatomic/hexatomic/compare/v1.1.2...HEAD
+[Unreleased]: https://github.com/hexatomic/hexatomic/compare/v1.1.3...HEAD
+[1.1.3]: https://github.com/hexatomic/hexatomic/compare/v1.1.2...v1.1.3
 [1.1.2]: https://github.com/hexatomic/hexatomic/compare/v1.1.1...v1.1.2
 [1.1.1]: https://github.com/hexatomic/hexatomic/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/hexatomic/hexatomic/compare/v1.0.1...v1.1.0
