### 2023-03-03 3f3aa26
Removed Maven-based target platform entries
There seems to be an issue with auto update otherwise.
