+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.5] - 2023-03-03
+
 ### Fixed
 
 - There seems to be an issue with Maven-based target platform entries and the
@@ -352,7 +354,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - This is the first public preview release of the Hexatomic developer platform.
 
 
-[Unreleased]: https://github.com/hexatomic/hexatomic/compare/v1.1.3...HEAD
+[Unreleased]: https://github.com/hexatomic/hexatomic/compare/v1.1.5...HEAD
+[1.1.5]: https://github.com/hexatomic/hexatomic/compare/v1.1.4...v1.1.5
+[1.1.4]: https://github.com/hexatomic/hexatomic/compare/v1.1.3...v1.1.4
 [1.1.3]: https://github.com/hexatomic/hexatomic/compare/v1.1.2...v1.1.3
 [1.1.2]: https://github.com/hexatomic/hexatomic/compare/v1.1.1...v1.1.2
 [1.1.1]: https://github.com/hexatomic/hexatomic/compare/v1.1.0...v1.1.1
