+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.7] - 2023-03-03
+
 ### Fixed
 
 - Add required P2 update bundles directly to core bundle so they are not lost
@@ -365,7 +367,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - This is the first public preview release of the Hexatomic developer platform.
 
 
-[Unreleased]: https://github.com/hexatomic/hexatomic/compare/v1.1.6...HEAD
+[Unreleased]: https://github.com/hexatomic/hexatomic/compare/v1.1.7...HEAD
+[1.1.7]: https://github.com/hexatomic/hexatomic/compare/v1.1.6...v1.1.7
 [1.1.6]: https://github.com/hexatomic/hexatomic/compare/v1.1.5...v1.1.6
 [1.1.5]: https://github.com/hexatomic/hexatomic/compare/v1.1.4...v1.1.5
 [1.1.4]: https://github.com/hexatomic/hexatomic/compare/v1.1.3...v1.1.4
