### 2023-03-04 9b22f57
Use command in macOS because the system menu is not accessible to SWTBot
