+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Make preferences accessible via default means on macOS (#413)
+
 ## [1.1.7] - 2023-03-03
 
 ### Fixed
