+++ CITATION.cff
@@ -34,8 +34,8 @@ identifiers:
   type: doi
   value: 10.5281/zenodo.6900689
 doi: 10.5281/zenodo.6900689
-version: 1.1.7
-date-released: 2023-03-03
+version: 1.2.0-SNAPSHOT
+date-released: 2023-03-04
 references:
 - type: software
   title: SWT Resource Manager
