### 2023-03-04 948a5a3
Add Preferences class to contain the string constants used as preferences
