### 2023-03-04 1bdf950
Fix checkstyle and eclipse project warnings
