### 2023-03-04 9af0262
Ask user before enabling update checks.
This fixes #412
