+++ bundles/org.corpus_tools.hexatomic.core/src/main/java/org/corpus_tools/hexatomic/core/ApplicationLifecycle.java
@@ -32,8 +32,6 @@ import org.eclipse.core.runtime.FileLocator;
 import org.eclipse.core.runtime.IProgressMonitor;
 import org.eclipse.core.runtime.Path;
 import org.eclipse.core.runtime.Platform;
-import org.eclipse.core.runtime.preferences.ConfigurationScope;
-import org.eclipse.core.runtime.preferences.IEclipsePreferences;
 import org.eclipse.e4.core.services.events.IEventBroker;
 import org.eclipse.e4.ui.di.UISynchronize;
 import org.eclipse.e4.ui.workbench.UIEvents;
@@ -45,7 +43,6 @@ import org.eclipse.wb.swt.ResourceManager;
 import org.eclipse.wb.swt.SwtResourceManager;
 import org.osgi.framework.Bundle;
 import org.osgi.framework.FrameworkUtil;
-import org.osgi.service.prefs.BackingStoreException;
 import org.slf4j.LoggerFactory;
 
 
@@ -63,8 +60,6 @@ public class ApplicationLifecycle {
 
   private static final org.slf4j.Logger log =
       org.slf4j.LoggerFactory.getLogger(ApplicationLifecycle.class);
-  private static final IEclipsePreferences prefs =
-      ConfigurationScope.INSTANCE.getNode("org.corpus_tools.hexatomic.core");
 
   /**
    * Called when the model is loaded and initializes the logging.
@@ -111,21 +106,10 @@ public class ApplicationLifecycle {
       }
     });
 
-    // check if preferences are set to autoupdate and if app was recently updated
-    boolean justUpdated = prefs.getBoolean(Preferences.JUST_UPDATED, false);
-    boolean autoUpdateEnabled = prefs.getBoolean(Preferences.AUTO_UPDATE, true);
-    if (!justUpdated && autoUpdateEnabled) {
-      // add listener to perform updates as soon as workbench is created
-      eventBroker.subscribe(UIEvents.UILifeCycle.APP_STARTUP_COMPLETE,
-          new AppStartupCompleteEventHandler(updateRunner, eventBroker, null));
-    } else if (justUpdated) {
-      prefs.putBoolean(Preferences.JUST_UPDATED, false);
-      try {
-        prefs.flush();
-      } catch (BackingStoreException ex) {
-        errorService.handleException("Couldn't update preferences", ex, ApplicationLifecycle.class);
-      }
-    }
+    // Always add listener to check if update is enabled and available as soon as workbench is
+    // created. The preferences will be evaluated by the update runner itself.
+    eventBroker.subscribe(UIEvents.UILifeCycle.APP_STARTUP_COMPLETE,
+        new AppStartupCompleteEventHandler(updateRunner, eventBroker, null));
   }
 }
 
