### 2023-03-04 2338d7b
Adapt information text in preferences dialog
