### 2023-03-04 90898f5
Adapt tests to new label captions
