### 2023-03-04 a51ed4d
Refactor some smaller code smells
