### 2023-03-04 39f90f1
Make UpdateRunner more easily to mock by adding a helper function for opening the question dialog
