### 2023-03-04 4e09593
Add some unit tests for UpdateRunner#autoUpdateAllowed
