### 2023-03-04 47c3a9d
Test error handling in UpdateRunner#autoUpdateAllowed
