+++ tests/org.corpus_tools.hexatomic.core.tests/src/main/java/org/corpus_tools/hexatomic/core/update/UpdateRunnerTest.java
@@ -2,14 +2,18 @@ package org.corpus_tools.hexatomic.core.update;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoInteractions;
 import static org.mockito.Mockito.when;
 
 import org.corpus_tools.hexatomic.core.DummySync;
 import org.corpus_tools.hexatomic.core.Preferences;
+import org.corpus_tools.hexatomic.core.errors.ErrorService;
 import org.eclipse.core.runtime.preferences.ConfigurationScope;
 import org.eclipse.core.runtime.preferences.IEclipsePreferences;
 import org.eclipse.swt.widgets.Shell;
@@ -19,9 +23,6 @@ import org.junit.jupiter.api.Test;
 import org.osgi.service.prefs.BackingStoreException;
 
 class UpdateRunnerTest {
-
-  private static final IEclipsePreferences prefs =
-      ConfigurationScope.INSTANCE.getNode("org.corpus_tools.hexatomic.core");
 
 
   public interface MessageDialogMock {
@@ -33,11 +34,14 @@ class UpdateRunnerTest {
   private Shell shell;
   private MessageDialogMock dialog;
 
+  private IEclipsePreferences actualPrefs =
+      ConfigurationScope.INSTANCE.getNode("org.corpus_tools.hexatomic.core");
 
   @BeforeEach
-  void setUp() throws Exception {
-    // Always start with now preferences
-    prefs.clear();
+  void setUp() throws BackingStoreException {
+    // Always start with empty preferences
+    actualPrefs.clear();
+
 
     shell = mock(Shell.class);
     dialog = mock(MessageDialogMock.class);
@@ -49,13 +53,17 @@ class UpdateRunnerTest {
       }
     };
     fixture.sync = new DummySync();
+    // Use the actual map implementation to store state, but allow mocking e.g. for throwing
+    // exceptions
+    fixture.prefs = spy(actualPrefs);
+    fixture.errorService = mock(ErrorService.class);
+
   }
 
   @AfterEach
   void cleanup() throws BackingStoreException {
-    prefs.clear();
+    actualPrefs.clear();
   }
-
 
   @Test
   void testUpdateDeniedAtStart() {
@@ -64,39 +72,72 @@ class UpdateRunnerTest {
     assertEquals(false, fixture.autoUpdateAllowed(shell));
     
     verify(dialog).openQuestionDialog(any(), eq("Automatic update check configuration"), any());
-    assertEquals(false, prefs.getBoolean(Preferences.AUTO_UPDATE, true));
+    assertEquals(false, fixture.prefs.getBoolean(Preferences.AUTO_UPDATE, true));
   }
 
   @Test
-  void testUpdateAllowedAtStart() {
+  void testUpdateAllowedAtStart() throws BackingStoreException {
     when(dialog.openQuestionDialog(any(), any(), any())).thenReturn(true);
     
     assertEquals(true, fixture.autoUpdateAllowed(shell));
     
     verify(dialog).openQuestionDialog(any(), eq("Automatic update check configuration"), any());
-    assertEquals(true, prefs.getBoolean(Preferences.AUTO_UPDATE, false));
[…]
