### 2023-03-04 2bfb821
Add test for UpdateRunner#checkForUpdates when no update is available
