+++ tests/org.corpus_tools.hexatomic.core.tests/src/main/java/org/corpus_tools/hexatomic/core/update/UpdateRunnerTest.java
@@ -13,9 +13,16 @@ import static org.mockito.Mockito.when;
 
 import org.corpus_tools.hexatomic.core.DummySync;
 import org.corpus_tools.hexatomic.core.Preferences;
+import org.corpus_tools.hexatomic.core.Topics;
 import org.corpus_tools.hexatomic.core.errors.ErrorService;
+import org.eclipse.core.runtime.IProgressMonitor;
+import org.eclipse.core.runtime.IStatus;
+import org.eclipse.core.runtime.Status;
 import org.eclipse.core.runtime.preferences.ConfigurationScope;
 import org.eclipse.core.runtime.preferences.IEclipsePreferences;
+import org.eclipse.e4.core.services.events.IEventBroker;
+import org.eclipse.equinox.p2.core.IProvisioningAgent;
+import org.eclipse.equinox.p2.operations.UpdateOperation;
 import org.eclipse.swt.widgets.Shell;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
@@ -25,14 +32,16 @@ import org.osgi.service.prefs.BackingStoreException;
 class UpdateRunnerTest {
 
 
-  public interface MessageDialogMock {
+  public interface UpdateRunnerMock {
     boolean openQuestionDialog(Shell parent, String title, String message);
+
+    UpdateOperation createUpdateOperation();
   }
 
   private UpdateRunner fixture;
 
   private Shell shell;
-  private MessageDialogMock dialog;
+  private UpdateRunnerMock updateRunner;
 
   private IEclipsePreferences actualPrefs =
       ConfigurationScope.INSTANCE.getNode("org.corpus_tools.hexatomic.core");
@@ -44,19 +53,27 @@ class UpdateRunnerTest {
 
 
     shell = mock(Shell.class);
-    dialog = mock(MessageDialogMock.class);
+    updateRunner = mock(UpdateRunnerMock.class);
 
     fixture = new UpdateRunner() {
       @Override
       protected boolean openQuestionDialog(Shell parent, String title, String message) {
-        return dialog.openQuestionDialog(parent, title, message);
+        return updateRunner.openQuestionDialog(parent, title, message);
+      }
+
+      @Override
+      protected UpdateOperation createUpdateOperation() {
+        return updateRunner.createUpdateOperation();
       }
     };
+
     fixture.sync = new DummySync();
     // Use the actual map implementation to store state, but allow mocking e.g. for throwing
     // exceptions
     fixture.prefs = spy(actualPrefs);
     fixture.errorService = mock(ErrorService.class);
+    fixture.agent = mock(IProvisioningAgent.class);
+    fixture.events = mock(IEventBroker.class);
 
   }
 
@@ -67,28 +84,28 @@ class UpdateRunnerTest {
 
   @Test
   void testUpdateDeniedAtStart() {
-    when(dialog.openQuestionDialog(any(), any(), any())).thenReturn(false);
+    when(updateRunner.openQuestionDialog(any(), any(), any())).thenReturn(false);
     
     assertEquals(false, fixture.autoUpdateAllowed(shell));
     
-    verify(dialog).openQuestionDialog(any(), eq("Automatic update check configuration"), any());
+    verify(updateRunner).openQuestionDialog(any(), eq("Automatic update check configuration"), any());
     assertEquals(false, fixture.prefs.getBoolean(Preferences.AUTO_UPDATE, true));
   }
 
   @Test
   void testUpdateAllowedAtStart() throws BackingStoreException {
[…]
