### 2023-03-04 5d05bf2
Add more tests for UpdateRunner#checkForUpdates
