+++ bundles/org.corpus_tools.hexatomic.core/src/main/java/org/corpus_tools/hexatomic/core/update/UpdateRunner.java
@@ -62,7 +62,7 @@ public class UpdateRunner {
   IProvisioningAgent agent;
 
   @Inject
-  private IEclipseContext context;
+  IEclipseContext context;
 
   @Inject
   private IProgressMonitor monitor;
