### 2023-03-04 abb7133
Add tests for error reporting of UpdateRunner
