### 2023-03-05 f1fbd6a
Add test for the restart option when update is finished
