### 2023-03-05 6de9639
Replace string with constant
