### 2023-03-05 47b3a5e
Sketch a more detailed manual testing plan for testing the update mechanism
