### 2023-03-05 195d6f6
Switching to the tag should not change the p2.inf file.
