+++ docs/dev/src/maintenance/contributions/manual-update-tests.md
@@ -11,11 +11,12 @@ We have to test two scenarios:
 
 ## Prepare a local update site
 
-1. Adjust the file `releng/org.corpus_tools.hexatomic.product/p2.inf` and the local update site as repository by adding the following lines (but replace `<TMP>` with a temporary folder where you want to store the update site content, e.g. `/tmp/hexatomic-update-site`.
+1. Adjust the file `releng/org.corpus_tools.hexatomic.product/p2.inf` and the local update site as repository by adding the following lines. 
+   Replace `<TMP>` with a temporary folder where you want to store the update site content, e.g. `/tmp/hexatomic-update-site`.
 ```plain
   addRepository(type:0,location:file${#58}<TMP>/repository,name:Local test,enabled:true);\
   addRepository(type:1,location:file${#58}<TMP>/repository,name:Local test,enabled:true);\
-```
+``` 
 2. Change the version number of Hexatomic to a large version number
 ```bash
 mvn tycho-versions:set-version -DnewVersion=1999.0.0 
@@ -36,7 +37,7 @@ releng/org.corpus_tools.hexatomic.product/target/products/org.corpus_tools.hexat
 
 ## Updating from the latest release
 
-1. Checkout the source with the release tag `git checkout tags/v<version>`. The `p2.inf` file should stay locally altered.
+1. Check out the source with the release tag `git checkout tags/v<version>`. The `p2.inf` file should stay locally altered.
 2. Execute `mvn package` to create the product
 3. Run Hexatomic from the command line and **check that there is a message in the toolbar about an available update**.
 ```
