### 2023-03-05 fe72463
Try not to execcute the code blocks as Rust code
