### 2023-03-05 4773209
Add git command how to restore the version, but keep the p2.inf file
