### 2023-03-07 864ece0
Upgrade to cff-maven-plugin 0.8.0
