### 2023-03-07 7a9ddcc
Remove unnecessary cast to double in division.
If any of the side of the division operator is a float, this is a float division operation.
