### 2023-03-07 90246ef
Use method reference for cast
