### 2023-03-07 85a23ab
Simplify SyntaxListener#exitNewSpan
