+++ bundles/org.corpus_tools.hexatomic.graph/src/main/java/org/corpus_tools/hexatomic/console/internal/SyntaxListener.java
@@ -327,19 +327,22 @@ public class SyntaxListener extends ConsoleCommandBaseListener {
     }
   }
 
-  @Override
-  public void exitNewSpan(NewSpanContext ctx) {
-
-    boolean referencesTokensOnly = true;
+  private boolean checkReferencesOnlyTokens() {
 
     for (SStructuredNode node : referencedNodes) {
       if (!(node instanceof SToken)) {
         this.outputLines
             .add("Error: could not create the new span - " + node.getName() + " is not a token.");
-        referencesTokensOnly = false;
+        return false;
       }
     }
-    if (referencesTokensOnly) {
+    return true;
+  }
+
+  @Override
+  public void exitNewSpan(NewSpanContext ctx) {
+
+    if (checkReferencesOnlyTokens()) {
 
       // Create the span
       SSpan newSpan = this.graph.createSpan(
