### 2023-03-16 5a4abd3
Added some of the planned developments to "Future work".
To make it more clear why this was in the "Statement of need" in the first place, the possibility to add editors is now at the beginning of the "Statement of need" section.
