### 2023-03-16 hexatomic/hexatomic#446 @thomaskrause
Document how to use demo corpora from corpus-tools.org
In the review for the JOSS paper, it was mentioned that we should provide easily downloadable corpora and document how to open/annotate them.

https://github.com/openjournals/joss-reviews/issues/4825#issuecomment-1296237504

**Describe the solution you'd like**
A clear and concise description of what you want to happen.

There are demo corpora for ANNIS (<https://corpus-tools.org/annis/corpora.html>) that we can link and add a description to the HOW-TOs section. E.g. "fuerstinnenkorrespondenz" for Exmaralda and "pcc2" for Paula
