+++ docs/user/src/SUMMARY.md
@@ -21,6 +21,6 @@
   	- [Refactoring](usage/refactoring/README.md)
 - [Troubleshooting & getting help](./troubleshooting.md)
 - [HOW-TOs](./howtos/README.md)
-    - [Import an example corpus](howtos/example-import.md)
+    - [Import and annotate a corpus](howtos/example-import.md)
     - [Creating a corpus from scratch](howtos/corpus-creation.md)
     - [Import a corpus from EXMARaLDA](howtos/exmaralda-import.md)
