+++ docs/user/src/howtos/example-import.md
@@ -1,4 +1,4 @@
-# Import an example corpus
+# Import and annotate an example corpus
 
 We will import an example corpus from the [ANNIS demo corpus
 page](https://corpus-tools.org/annis/corpora.html), namely the so-called
@@ -11,9 +11,15 @@ It contains several annotation layers, like constituent trees, dependency trees 
 4. Choose the *Import* entry in the *File* menu.
 5. Click on the button with the *...* caption and navigate to the unzipped `pcc2_v6_PAULA` folder. Then click on *Next*.
 ![Select a corpus folder in the import wizard](select-pcc2-folder.png)
-1. The importer should correctly identify this corpus as “PAULA format”. Click on *Finish* to import the corpus.
+6. The importer should correctly identify this corpus as “PAULA format”. Click on *Finish* to import the corpus.
 ![Format selection wizard step](pcc2-finish.png)
-1. Unfold the corpus and in the “Corpus Structure” and right-click on the “4282” document, select “Open with Graph Editor”. You [filter](../usage/graph-editor/filter.md) the view to only show nodes with the `tiger::cat` annotation, to focus on the constituent trees.
+7. Unfold the corpus and in the “Corpus Structure” and right-click on the “4282” document, select “Open with Graph Editor”.
 ![Graph editor with the openend document](pcc2-graph-editor.png)
-1. If you like, add new annotations using the [console](../usage/graph-editor/console.md).
+8. This shows the whole document as a graph, but we are only interested in the constituent tree for now. Expand “Annotation Types” and “Node annotations” in the [Filter View](../usage/graph-editor/filter.md). Unselect “Pointing Relations”. 
+Then type `cat` into the “Search” field and click on the `tiger::cat` filter badge. 
+This still shows the whole document, but now we can select the segments we are interested in.
+Click on the first three segments while holding the <kbd>Ctrl</kbd> key.
+![Graph editor with selected segments](pcc2-select-segments.png)
+1. Add new annotations using the [console](../usage/graph-editor/console.md). To add a root node connecting the trailing token “!” with the sentence constituent node, enter `n tiger:cat:ROOT  #tok_7 #const_2` to the console and press <kbd>Enter</kbd>. Note that the segmentation changes because we connected the previously separate segments.   
+![Adding a root node with the graph editor](pcc2-add-root-node.png)
 2. Save the project via by clicking on the *File* menu and then *Save Salt Project As...* to persist the changes as a [project](../usage/projects.md).
