### 2023-03-16 cd3e0d6
Store last opened project location as preference.
This means the information survives restarts of Hexatomic.
