### 2023-03-16 1d7ad6a
Add the selected node reference on double click in the graph view to the console
