### 2023-03-16 3272711
Extend existing GraphEditor integration test to emulate double clicking on a selected node
