+++ tests/org.corpus_tools.hexatomic.it.tests/src/main/java/org/corpus_tools/hexatomic/graph/TestGraphEditor.java
@@ -1,8 +1,9 @@
-package org.corpus_tools.hexatomic.it.tests;
+package org.corpus_tools.hexatomic.graph;
 
 import static org.eclipse.swtbot.swt.finder.matchers.WidgetMatcherFactory.widgetOfType;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertInstanceOf;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
@@ -12,6 +13,7 @@ import java.lang.reflect.InvocationTargetException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.text.MessageFormat;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
@@ -25,7 +27,10 @@ import org.corpus_tools.hexatomic.core.CommandParams;
 import org.corpus_tools.hexatomic.core.ProjectManager;
 import org.corpus_tools.hexatomic.core.UiStatusReport;
 import org.corpus_tools.hexatomic.core.errors.ErrorService;
-import org.corpus_tools.hexatomic.graph.GraphEditor;
+import org.corpus_tools.hexatomic.it.tests.PartMaximizedCondition;
+import org.corpus_tools.hexatomic.it.tests.TableCellEditorInactiveCondition;
+import org.corpus_tools.hexatomic.it.tests.TestCorpusStructure;
+import org.corpus_tools.hexatomic.it.tests.TestHelper;
 import org.corpus_tools.hexatomic.it.tests.utils.SwtBotChips;
 import org.corpus_tools.salt.common.SDocument;
 import org.corpus_tools.salt.common.SDocumentGraph;
@@ -45,6 +50,7 @@ import org.eclipse.e4.core.contexts.IEclipseContext;
 import org.eclipse.e4.ui.workbench.modeling.EPartService;
 import org.eclipse.emf.common.util.URI;
 import org.eclipse.jface.bindings.keys.KeyStroke;
+import org.eclipse.jface.viewers.StructuredSelection;
 import org.eclipse.nebula.widgets.chips.Chips;
 import org.eclipse.swt.SWT;
 import org.eclipse.swtbot.e4.finder.widgets.SWTBotView;
@@ -81,6 +87,8 @@ import org.junit.jupiter.api.condition.OS;
 @TestMethodOrder(OrderAnnotation.class)
 class TestGraphEditor {
 
+  private static final String STRUCTURE5_ID = "salt:/rootCorpus/subCorpus1/doc1#structure5";
+  private static final String STRUCTURE3_ID = "salt:/rootCorpus/subCorpus1/doc1#structure3";
   private static final String DOC1_SALT_ID = "salt:/rootCorpus/subCorpus1/doc1";
   private static final String DOC1_TITLE = "doc1 (Graph Editor)";
   private static final String CONST = "const";
@@ -614,18 +622,51 @@ class TestGraphEditor {
       SDocumentGraph graph = doc.getDocumentGraph();
 
       // Before state: no edges between the two structures
-      assertEquals(0, graph.getRelations("salt:/rootCorpus/subCorpus1/doc1#structure3",
-          "salt:/rootCorpus/subCorpus1/doc1#structure5").size());
+      assertEquals(0, graph.getRelations(STRUCTURE3_ID, STRUCTURE5_ID).size());
 
-      // Add a pointing relation between the two structures
-      enterCommand(ADD_POINTING_COMMMAND);
[…]
