### 2023-03-16 872d580
Wait until part object is present
