### 2023-03-16 8622722
Show view before trying to access attached part object
