+++ tests/org.corpus_tools.hexatomic.it.tests/src/main/java/org/corpus_tools/hexatomic/graph/TestGraphEditor.java
@@ -624,6 +624,7 @@ class TestGraphEditor {
       assertEquals(0, graph.getRelations(STRUCTURE3_ID, STRUCTURE5_ID).size());
 
       SWTBotView view = bot.partByTitle(DOC1_TITLE);
+      view.show();
       bot.waitUntil(new DefaultCondition() {
 
         @Override
