+++ CITATION.cff
@@ -34,7 +34,7 @@ identifiers:
   type: doi
   value: 10.5281/zenodo.6900689
 doi: 10.5281/zenodo.6900689
-version: 1.3.0-SNAPSHOT
+version: 1.3.0
 date-released: 2023-03-16
 references:
 - type: art
