### 2023-03-20 70d192b
More clearly separate versioning from the application
