+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.4.0] - 2023-03-24
+
 ### Changed
 
 - Improve import how-to in user docs (#452)
@@ -396,7 +398,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - This is the first public preview release of the Hexatomic developer platform.
 
 
-[Unreleased]: https://github.com/hexatomic/hexatomic/compare/v1.3.0...HEAD
+[Unreleased]: https://github.com/hexatomic/hexatomic/compare/v1.4.0...HEAD
+[1.4.0]: https://github.com/hexatomic/hexatomic/compare/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/hexatomic/hexatomic/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/hexatomic/hexatomic/compare/v1.1.7...v1.2.0
 [1.1.7]: https://github.com/hexatomic/hexatomic/compare/v1.1.6...v1.1.7
