### 2023-03-24 #4825 @sdruskat
Dear @ajstewartlang, @reckart and @hsang,

Thank you @reckart for the very helpful review comments, they helped us iron out a few flaws and bugs indeed. We believe we have now addressed all comments by @reckart, but couldn't see any further comments by @hsang.

The changes are included in the [1.4.0](https://github.com/hexatomic/hexatomic/releases/tag/v1.4.0) release, changes to the overview paper are in our `JOSS-paper` branch. @ajstewartlang Can you let us know if we should update the `JOSS-paper` branch with the recent changes (to the paper, and to the software)?

## Details

> These items remain unchecked on my list:
>
> **General checks**
>
> - Reproducibility: I cannot build the repository. [Build fails with NoClassDefFound hexatomic/hexatomic#417](https://github.com/hexatomic/hexatomic/issues/417)

- This has been fixed.

> **Functionality**
>
> - Functionality: [No merging in grid view? hexatomic/hexatomic#423](https://github.com/hexatomic/hexatomic/issues/423) , no documentation on git integration claimed in paper

- The first item is documented, the paper has been updated to reflect the second item better.

> **Documentation**
>
> -  A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is? -- statement of need from paper should be transferred to documentation

- A statement of need is part of the user documentation (https://hexatomic.github.io/hexatomic/user/v1.3/index.html#what-is-hexatomic) which is adapted for the target group.

> -  Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems). -- Easily downloadable examples in different formats and for different use-cases should be provided

- How-tos for importing example data have been added to the documentation (PAULA data and using the Graph Editor, EXMARaLDA data and using the Grid Editor), including data download links.

> **Software paper**
>
> -  A statement of need: Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work? -- see [[REVIEW]: Hexatomic: An extensible, OS-independent platform for deep multi-layer linguistic annotation of corpora #4825 (comment)](https://github.com/openjournals/joss-reviews/issues/4825#issuecomment-1296233690)

- Issues have been addressed in the updated paper.

> -  State of the field: Do the authors describe how this software compares to other commonly-used packages? -- Comparison is very superficial and not in a dedicated section.
>
> -  Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)? -- Sections for "State of the field" and "Future Work" should be extracted. Otherwise ok.

- We have extracted a "Future work" section, but have refrained from extracting further sections due to the intended brevity and status of the overview papers in JOSS.
