### 2023-03-27 a2d12e6
Update version in CITATION.cff
