+++ CITATION.cff
@@ -34,8 +34,8 @@ identifiers:
   type: doi
   value: 10.5281/zenodo.6900689
 doi: 10.5281/zenodo.6900689
-version: 1.4.0
-date-released: 2023-03-24
+version: 1.4.1
+date-released: 2023-03-27
 references:
 - type: art
   title: Font Awesome
@@ -403,15 +403,15 @@ references:
   - name: The org.corpus-tools.pepper-framework 3.6.0 Team
 - type: software
   title: org.corpus-tools.salt-api
-  version: 3.4.2
+  version: 3.4.3
   license: Apache-2.0
   authors:
-  - name: The org.corpus-tools.salt-api 3.4.2 Team
+  - name: The org.corpus-tools.salt-api 3.4.3 Team
 - type: software
   title: org.corpus-tools.salt-extensions
-  version: 3.4.2
+  version: 3.4.3
   authors:
-  - name: The org.corpus-tools.salt-extensions 3.4.2 Team
+  - name: The org.corpus-tools.salt-extensions 3.4.3 Team
 - type: software
   title: org.eclipse.core.commands
   version: 3.10.300.v20221024-2137
