### 2023-03-28 #4825 @sdruskat
@ajstewartlang I have now merged the latest release into our [`JOSS-paper` branch](https://github.com/hexatomic/hexatomic/tree/JOSS-paper). The current version is now **v1.4.1**.

/cc @thomaskrause
