### 2023-03-28 0e23f44
Use the new org.corpus_tools.hexatomic.p2.inf name in documentation
