### 2023-05-31 #4825 @ajstewartlang
:wave: @sdruskat

If you could now do the following please, that would be great:

Make a [tagged release](https://docs.github.com/en/github/administering-a-repository/releasing-projects-on-github/managing-releases-in-a-repository) of the software.
Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
Please list the DOI of the archived version here.
