### 2023-06-09 #4825 @sdruskat
@ajstewartlang Just to confirm, it is you who would need to accept the submission with the above-mentioned editorialbot command, not us, right?
