### 2023-06-09 #4825 @sdruskat
> @sdruskat - one of the @openjournals/sbcs-eics needs to action this.

Thanks for the quick reply, makes sense :).
