### 2023-06-13 #4825 @sdruskat
> Congratulations @sdruskat! 🥳 Big thanks to @ajstewartlang @reckart! 😊

Yay 🎉! Thanks @ajstewartlang, @reckart and @oliviaguest. I was watching live as the acceptance went on, very exciting!
/cc @thomaskrause
